refactor: remove operation area inconsistencies. add import keys form
