+++ public/lib/index.js
@@ -1,10 +1,10 @@
-"use strict";(()=>{var gn=Object.create;var yt=Object.defineProperty;var vn=Object.getOwnPropertyDescriptor;var En=Object.getOwnPropertyNames;var Sn=Object.getPrototypeOf,wn=Object.prototype.hasOwnProperty;var ve=(e=>typeof require<"u"?require:typeof Proxy<"u"?new Proxy(e,{get:(t,n)=>(typeof require<"u"?require:t)[n]}):e)(function(e){if(typeof require<"u")return require.apply(this,arguments);throw Error('Dynamic require of "'+e+'" is not supported')});var Q=(e,t)=>()=>(t||e((t={exports:{}}).exports,t),t.exports);var Ln=(e,t,n,r)=>{if(t&&typeof t=="object"||typeof t=="function")for(let o of En(t))!wn.call(e,o)&&o!==n&&yt(e,o,{get:()=>t[o],enumerable:!(r=vn(t,o))||r.enumerable});return e};var De=(e,t,n)=>(n=e!=null?gn(Sn(e)):{},Ln(t||!e||!e.__esModule?yt(n,"default",{value:e,enumerable:!0}):n,e));var ze=Q((Et,ee)=>{(function(e,t){typeof define=="function"&&define.amd?define([],t):typeof ve=="function"&&typeof ee=="object"&&ee&&ee.exports?ee.exports=t():(e.dcodeIO=e.dcodeIO||{}).bcrypt=t()})(Et,function(){"use strict";var e={},t=null;function n(f){if(typeof ee<"u"&&ee&&ee.exports)try{return ve("crypto").randomBytes(f)}catch{}try{var d;return(self.crypto||self.msCrypto).getRandomValues(d=new Uint32Array(f)),Array.prototype.slice.call(d)}catch{}if(!t)throw Error("Neither WebCryptoAPI nor a crypto module is available. Use bcrypt.setRandomFallback to set an alternative");return t(f)}var r=!1;try{n(1),r=!0}catch{}t=null,e.setRandomFallback=function(f){t=f},e.genSaltSync=function(f,d){if(f=f||h,typeof f!="number")throw Error("Illegal arguments: "+typeof f+", "+typeof d);f<4?f=4:f>31&&(f=31);var l=[];return l.push("$2a$"),f<10&&l.push("0"),l.push(f.toString()),l.push("$"),l.push(u(n(S),S)),l.join("")},e.genSalt=function(f,d,l){if(typeof d=="function"&&(l=d,d=void 0),typeof f=="function"&&(l=f,f=void 0),typeof f>"u")f=h;else if(typeof f!="number")throw Error("illegal arguments: "+typeof f);function s(c){a(function(){try{c(null,e.genSaltSync(f))}catch(x){c(x)}})}if(l){if(typeof l!="function")throw Error("Illegal callback: "+typeof l);s(l)}else return new Promise(function(c,x){s(function(y,L){if(y){x(y);return}c(L)})})},e.hashSync=function(f,d){if(typeof d>"u"&&(d=h),typeof d=="number"&&(d=e.genSaltSync(d)),typeof f!="string"||typeof d!="string")throw Error("Illegal arguments: "+typeof f+", "+typeof d);return Ue(f,d)},e.hash=function(f,d,l,s){function c(x){typeof f=="string"&&typeof d=="number"?e.genSalt(d,function(y,L){Ue(f,L,x,s)}):typeof f=="string"&&typeof d=="string"?Ue(f,d,x,s):a(x.bind(this,Error("Illegal arguments: "+typeof f+", "+typeof d)))}if(l){if(typeof l!="function")throw Error("Illegal callback: "+typeof l);c(l)}else return new Promise(function(x,y){c(function(L,E){if(L){y(L);return}x(E)})})};function o(f,d){for(var l=0,s=0,c=0,x=f.length;c<x;++c)f.charCodeAt(c)===d.charCodeAt(c)?++l:++s;return l<0?!1:s===0}e.compareSync=function(f,d){if(typeof f!="string"||typeof d!="string")throw Error("Illegal arguments: "+typeof f+", "+typeof d);return d.length!==60?!1:o(e.hashSync(f,d.substr(0,d.length-31)),d)},e.compare=function(f,d,l,s){function c(x){if(typeof f!="string"||typeof d!="string"){a(x.bind(this,Error("Illegal arguments: "+typeof f+", "+typeof d)));return}if(d.length!==60){a(x.bind(this,null,!1));return}e.hash(f,d.substr(0,29),function(y,L){y?x(y):x(null,o(L,d))},s)}if(l){if(typeof l!="function")throw Error("Illegal callback: "+typeof l);c(l)}else return new Promise(function(x,y){c(function(L,E){if(L){y(L);return}x(E)})})},e.getRounds=function(f){if(typeof f!="string")throw Error("Illegal arguments: "+typeof f);return parseInt(f.split("$")[2],10)},e.getSalt=function(f){if(typeof f!="string")throw Error("Illegal arguments: "+typeof f);if(f.length!==60)throw Error("Illegal hash length: "+f.length+" != 60");return f.substring(0,29)};var a=typeof process<"u"&&process&&typeof process.nextTick=="function"?typeof setImmediate=="function"?setImmediate:process.nextTick:setTimeout;function i(f){var d=[],l=0;return g.encodeUTF16toUTF8(function(){return l>=f.length?null:f.charCodeAt(l++)},function(s){d.push(s)}),d}var p="./ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789".split(""),m=[-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,1,54,55,56,57,58,59,60,61,62,63,-1,-1,-1,-1,-1,-1,-1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,-1,-1,-1,-1,-1,-1,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,-1,-1,-1,-1,-1],b=String.fromCharCode;function u(f,d){var l=0,s=[],c,x;if(d<=0||d>f.length)throw Error("Illegal len: "+d);for(;l<d;){if(c=f[l++]&255,s.push(p[c>>2&63]),c=(c&3)<<4,l>=d){s.push(p[c&63]);break}if(x=f[l++]&255,c|=x>>4&15,s.push(p[c&63]),c=(x&15)<<2,l>=d){s.push(p[c&63]);break}x=f[l++]&255,c|=x>>6&3,s.push(p[c&63]),s.push(p[x&63])}return s.join("")}function v(f,d){var l=0,s=f.length,c=0,x=[],y,L,E,T,R,K;if(d<=0)throw Error("Illegal len: "+d);for(;l<s-1&&c<d&&(K=f.charCodeAt(l++),y=K<m.length?m[K]:-1,K=f.charCodeAt(l++),L=K<m.length?m[K]:-1,!(y==-1||L==-1||(R=y<<2>>>0,R|=(L&48)>>4,x.push(b(R)),++c>=d||l>=s)||(K=f.charCodeAt(l++),E=K<m.length?m[K]:-1,E==-1)||(R=(L&15)<<4>>>0,R|=(E&60)>>2,x.push(b(R)),++c>=d||l>=s)));)K=f.charCodeAt(l++),T=K<m.length?m[K]:-1,R=(E&3)<<6>>>0,R|=T,x.push(b(R)),++c;var X=[];for(l=0;l<c;l++)X.push(x[l].charCodeAt(0));return X}var g=function(){"use strict";var f={};return f.MAX_CODEPOINT=1114111,f.encodeUTF8=function(d,l){var s=null;for(typeof d=="number"&&(s=d,d=function(){return null});s!==null||(s=d())!==null;)s<128?l(s&127):s<2048?(l(s>>6&31|192),l(s&63|128)):s<65536?(l(s>>12&15|224),l(s>>6&63|128),l(s&63|128)):(l(s>>18&7|240),l(s>>12&63|128),l(s>>6&63|128),l(s&63|128)),s=null},f.decodeUTF8=function(d,l){for(var s,c,x,y,L=function(E){E=E.slice(0,E.indexOf(null));var T=Error(E.toString());throw T.name="TruncatedError",T.bytes=E,T};(s=d())!==null;)if(!(s&128))l(s);else if((s&224)===192)(c=d())===null&&L([s,c]),l((s&31)<<6|c&63);else if((s&240)===224)((c=d())===null||(x=d())===null)&&L([s,c,x]),l((s&15)<<12|(c&63)<<6|x&63);else if((s&248)===240)((c=d())===null||(x=d())===null||(y=d())===null)&&L([s,c,x,y]),l((s&7)<<18|(c&63)<<12|(x&63)<<6|y&63);else throw RangeError("Illegal starting byte: "+s)},f.UTF16toUTF8=function(d,l){for(var s,c=null;(s=c!==null?c:d())!==null;){if(s>=55296&&s<=57343&&(c=d())!==null&&c>=56320&&c<=57343){l((s-55296)*1024+c-56320+65536),c=null;continue}l(s)}c!==null&&l(c)},f.UTF8toUTF16=function(d,l){var s=null;for(typeof d=="number"&&(s=d,d=function(){return null});s!==null||(s=d())!==null;)s<=65535?l(s):(s-=65536,l((s>>10)+55296),l(s%1024+56320)),s=null},f.encodeUTF16toUTF8=function(d,l){f.UTF16toUTF8(d,function(s){f.encodeUTF8(s,l)})},f.decodeUTF8toUTF16=function(d,l){f.decodeUTF8(d,function(s){f.UTF8toUTF16(s,l)})},f.calculateCodePoint=function(d){return d<128?1:d<2048?2:d<65536?3:4},f.calculateUTF8=function(d){for(var l,s=0;(l=d())!==null;)s+=f.calculateCodePoint(l);return s},f.calculateUTF16asUTF8=function(d){var l=0,s=0;return f.UTF16toUTF8(d,function(c){++l,s+=f.calculateCodePoint(c)}),[l,s]},f}();Date.now=Date.now||function(){return+new Date};var S=16,h=10,w=16,$=100,P=[608135816,2242054355,320440878,57701188,2752067618,698298832,137296536,3964562569,1160258022,953160567,3193202383,887688300,3232508343,3380367581,1065670069,3041331479,2450970073,2306472731],A=[3509652390,2564797868,805139163,3491422135,3101798381,1780907670,3128725573,4046225305,614570311,3012652279,134345442,2240740374,1667834072,1901547113,2757295779,4103290238,227898511,1921955416,1904987480,2182433518,2069144605,3260701109,2620446009,720527379,3318853667,677414384,3393288472,3101374703,2390351024,1614419982,1822297739,2954791486,3608508353,3174124327,2024746970,1432378464,3864339955,2857741204,1464375394,1676153920,1439316330,715854006,3033291828,289532110,2706671279,2087905683,3018724369,1668267050,732546397,1947742710,3462151702,2609353502,2950085171,1814351708,2050118529,680887927,999245976,1800124847,3300911131,1713906067,1641548236,4213287313,1216130144,1575780402,4018429277,3917837745,3693486850,3949271944,596196993,3549867205,258830323,2213823033,772490370,2760122372,1774776394,2652871518,566650946,4142492826,1728879713,2882767088,1783734482,3629395816,2517608232,2874225571,1861159788,326777828,3124490320,2130389656,2716951837,967770486,1724537150,2185432712,2364442137,1164943284,2105845187,998989502,3765401048,2244026483,1075463327,1455516326,1322494562,910128902,469688178,1117454909,936433444,3490320968,3675253459,1240580251,122909385,2157517691,634681816,4142456567,3825094682,3061402683,2540495037,79693498,3249098678,1084186820,1583128258,426386531,1761308591,1047286709,322548459,995290223,1845252383,2603652396,3431023940,2942221577,3202600964,3727903485,1712269319,422464435,3234572375,1170764815,3523960633,3117677531,1434042557,442511882,3600875718,1076654713,1738483198,4213154764,2393238008,3677496056,1014306527,4251020053,793779912,2902807211,842905082,4246964064,1395751752,1040244610,2656851899,3396308128,445077038,3742853595,3577915638,679411651,2892444358,2354009459,1767581616,3150600392,3791627101,3102740896,284835224,4246832056,1258075500,768725851,2589189241,3069724005,3532540348,1274779536,3789419226,2764799539,1660621633,3471099624,4011903706,913787905,3497959166,737222580,2514213453,2928710040,3937242737,1804850592,3499020752,2949064160,2386320175,2390070455,2415321851,4061277028,2290661394,2416832540,1336762016,1754252060,3520065937,3014181293,791618072,3188594551,3933548030,2332172193,3852520463,3043980520,413987798,3465142937,3030929376,4245938359,2093235073,3534596313,375366246,2157278981,2479649556,555357303,3870105701,2008414854,3344188149,4221384143,3956125452,2067696032,3594591187,2921233993,2428461,544322398,577241275,1471733935,610547355,4027169054,1432588573,1507829418,2025931657,3646575487,545086370,48609733,2200306550,1653985193,298326376,1316178497,3007786442,2064951626,458293330,2589141269,3591329599,3164325604,727753846,2179363840,146436021,1461446943,4069977195,705550613,3059967265,3887724982,4281599278,3313849956,1404054877,2845806497,146425753,1854211946,1266315497,3048417604,3681880366,3289982499,290971e4,1235738493,2632868024,2414719590,3970600049,1771706367,1449415276,3266420449,422970021,1963543593,2690192192,3826793022,1062508698,1531092325,1804592342,2583117782,2714934279,4024971509,1294809318,4028980673,1289560198,2221992742,1669523910,35572830,157838143,1052438473,1016535060,1802137761,1753167236,1386275462,3080475397,2857371447,1040679964,2145300060,2390574316,1461121720,2956646967,4031777805,4028374788,33600511,2920084762,1018524850,629373528,3691585981,3515945977,2091462646,2486323059,586499841,988145025,935516892,3367335476,2599673255,2839830854,265290510,3972581182,2759138881,3795373465,1005194799,847297441,406762289,1314163512,1332590856,1866599683,4127851711,750260880,613907577,1450815602,3165620655,3734664991,3650291728,3012275730,3704569646,1427272223,778793252,1343938022,2676280711,2052605720,1946737175,3164576444,3914038668,3967478842,3682934266,1661551462,3294938066,4011595847,840292616,3712170807,616741398,312560963,711312465,1351876610,322626781,1910503582,271666773,2175563734,1594956187,70604529,3617834859,1007753275,1495573769,4069517037,2549218298,2663038764,504708206,2263041392,3941167025,2249088522,1514023603,1998579484,1312622330,694541497,2582060303,2151582166,1382467621,776784248,2618340202,3323268794,2497899128,2784771155,503983604,4076293799,907881277,423175695,432175456,1378068232,4145222326,3954048622,3938656102,3820766613,2793130115,2977904593,26017576,3274890735,3194772133,1700274565,1756076034,4006520079,3677328699,720338349,1533947780,354530856,688349552,3973924725,1637815568,332179504,3949051286,53804574,2852348879,3044236432,1282449977,3583942155,3416972820,4006381244,1617046695,2628476075,3002303598,1686838959,431878346,2686675385,1700445008,1080580658,1009431731,832498133,3223435511,2605976345,2271191193,2516031870,1648197032,4164389018,2548247927,300782431,375919233,238389289,3353747414,2531188641,2019080857,1475708069,455242339,2609103871,448939670,3451063019,1395535956,2413381860,1841049896,1491858159,885456874,4264095073,4001119347,1565136089,3898914787,1108368660,540939232,1173283510,2745871338,3681308437,4207628240,3343053890,4016749493,1699691293,1103962373,3625875870,2256883143,3830138730,1031889488,3479347698,1535977030,4236805024,3251091107,2132092099,1774941330,1199868427,1452454533,157007616,2904115357,342012276,595725824,1480756522,206960106,497939518,591360097,863170706,2375253569,3596610801,1814182875,2094937945,3421402208,1082520231,3463918190,2785509508,435703966,3908032597,1641649973,2842273706,3305899714,1510255612,2148256476,2655287854,3276092548,4258621189,236887753,3681803219,274041037,1734335097,3815195456,3317970021,1899903192,1026095262,4050517792,356393447,2410691914,3873677099,3682840055,3913112168,2491498743,4132185628,2489919796,1091903735,1979897079,3170134830,3567386728,3557303409,857797738,1136121015,1342202287,507115054,2535736646,337727348,3213592640,1301675037,2528481711,1895095763,1721773893,3216771564,62756741,2142006736,835421444,2531993523,1442658625,3659876326,2882144922,676362277,1392781812,170690266,3921047035,1759253602,3611846912,1745797284,664899054,1329594018,3901205900,3045908486,2062866102,2865634940,3543621612,3464012697,1080764994,553557557,3656615353,3996768171,991055499,499776247,1265440854,648242737,3940784050,980351604,3713745714,1749149687,3396870395,4211799374,3640570775,1161844396,3125318951,1431517754,545492359,4268468663,3499529547,1437099964,2702547544,3433638243,2581715763,2787789398,1060185593,1593081372,2418618748,4260947970,69676912,2159744348,86519011,2512459080,3838209314,1220612927,3339683548,133810670,1090789135,1078426020,1569222167,845107691,3583754449,4072456591,1091646820,628848692,1613405280,3757631651,526609435,236106946,48312990,2942717905,3402727701,1797494240,859738849,992217954,4005476642,2243076622,3870952857,3732016268,765654824,3490871365,2511836413,1685915746,3888969200,1414112111,2273134842,3281911079,4080962846,172450625,2569994100,980381355,4109958455,2819808352,2716589560,2568741196,3681446669,3329971472,1835478071,660984891,3704678404,4045999559,3422617507,3040415634,1762651403,1719377915,3470491036,2693910283,3642056355,3138596744,1364962596,2073328063,1983633131,926494387,3423689081,2150032023,4096667949,1749200295,3328846651,309677260,2016342300,1779581495,3079819751,111262694,1274766160,443224088,298511866,1025883608,3806446537,1145181785,168956806,3641502830,3584813610,1689216846,3666258015,3200248200,1692713982,2646376535,4042768518,1618508792,1610833997,3523052358,4130873264,2001055236,3610705100,2202168115,4028541809,2961195399,1006657119,2006996926,3186142756,1430667929,3210227297,1314452623,4074634658,4101304120,2273951170,1399257539,3367210612,3027628629,1190975929,2062231137,2333990788,2221543033,2438960610,1181637006,548689776,2362791313,3372408396,3104550113,3145860560,296247880,1970579870,3078560182,3769228297,1714227617,3291629107,3898220290,166772364,1251581989,493813264,448347421,195405023,2709975567,677966185,3703036547,1463355134,2715995803,1338867538,1343315457,2802222074,2684532164,233230375,2599980071,2000651841,3277868038,1638401717,4028070440,3237316320,6314154,819756386,300326615,590932579,1405279636,3267499572,3150704214,2428286686,3959192993,3461946742,1862657033,1266418056,963775037,2089974820,2263052895,1917689273,448879540,3550394620,3981727096,150775221,3627908307,1303187396,508620638,2975983352,2726630617,1817252668,1876281319,1457606340,908771278,3720792119,3617206836,2455994898,1729034894,1080033504,976866871,3556439503,2881648439,1522871579,1555064734,1336096578,3548522304,2579274686,3574697629,3205460757,3593280638,3338716283,3079412587,564236357,2993598910,1781952180,1464380207,3163844217,3332601554,1699332808,1393555694,1183702653,3581086237,1288719814,691649499,2847557200,2895455976,3193889540,2717570544,1781354906,1676643554,2592534050,3230253752,1126444790,2770207658,2633158820,2210423226,2615765581,2414155088,3127139286,673620729,2805611233,1269405062,4015350505,3341807571,4149409754,1057255273,2012875353,2162469141,2276492801,2601117357,993977747,3918593370,2654263191,753973209,36408145,2530585658,25011837,3520020182,2088578344,530523599,2918365339,1524020338,1518925132,3760827505,3759777254,1202760957,3985898139,3906192525,674977740,4174734889,2031300136,2019492241,3983892565,4153806404,3822280332,352677332,2297720250,60907813,90501309,3286998549,1016092578,2535922412,2839152426,457141659,509813237,4120667899,652014361,1966332200,2975202805,55981186,2327461051,676427537,3255491064,2882294119,3433927263,1307055953,942726286,933058658,2468411793,3933900994,4215176142,1361170020,2001714738,2830558078,3274259782,1222529897,1679025792,2729314320,3714953764,1770335741,151462246,3013232138,1682292957,1483529935,471910574,1539241949,458788160,3436315007,1807016891,3718408830,978976581,1043663428,3165965781,1927990952,4200891579,2372276910,3208408903,3533431907,1412390302,2931980059,4132332400,1947078029,3881505623,4168226417,2941484381,1077988104,1320477388,886195818,18198404,3786409e3,2509781533,112762804,3463356488,1866414978,891333506,18488651,661792760,1628790961,3885187036,3141171499,876946877,2693282273,1372485963,791857591,2686433993,3759982718,3167212022,3472953795,2716379847,445679433,3561995674,3504004811,3574258232,54117162,3331405415,2381918588,3769707343,4154350007,1140177722,4074052095,668550556,3214352940,367459370,261225585,2610173221,4209349473,3468074219,3265815641,314222801,3066103646,3808782860,282218597,3406013506,3773591054,379116347,1285071038,846784868,2669647154,3771962079,3550491691,2305946142,453669953,1268987020,3317592352,3279303384,3744833421,2610507566,3859509063,266596637,3847019092,517658769,3462560207,3443424879,370717030,4247526661,2224018117,4143653529,4112773975,2788324899,2477274417,1456262402,2901442914,1517677493,1846949527,2295493580,3734397586,2176403920,1280348187,1908823572,3871786941,846861322,1172426758,3287448474,3383383037,1655181056,3139813346,901632758,1897031941,2986607138,3066810236,3447102507,1393639104,373351379,950779232,625454576,3124240540,4148612726,2007998917,544563296,2244738638,2330496472,2058025392,1291430526,424198748,50039436,29584100,3605783033,2429876329,2791104160,1057563949,3255363231,3075367218,3463963227,1469046755,985887462],q=[1332899944,1700884034,1701343084,1684370003,1668446532,1869963892];function B(f,d,l,s){var c,x=f[d],y=f[d+1];return x^=l[0],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[1],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[2],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[3],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[4],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[5],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[6],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[7],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[8],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[9],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[10],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[11],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[12],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[13],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[14],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[15],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[16],f[d]=y^l[w+1],f[d+1]=x,f}function O(f,d){for(var l=0,s=0;l<4;++l)s=s<<8|f[d]&255,d=(d+1)%f.length;return{key:s,offp:d}}function z(f,d,l){for(var s=0,c=[0,0],x=d.length,y=l.length,L,E=0;E<x;E++)L=O(f,s),s=L.offp,d[E]=d[E]^L.key;for(E=0;E<x;E+=2)c=B(c,0,d,l),d[E]=c[0],d[E+1]=c[1];for(E=0;E<y;E+=2)c=B(c,0,d,l),l[E]=c[0],l[E+1]=c[1]}function Z(f,d,l,s){for(var c=0,x=[0,0],y=l.length,L=s.length,E,T=0;T<y;T++)E=O(d,c),c=E.offp,l[T]=l[T]^E.key;for(c=0,T=0;T<y;T+=2)E=O(f,c),c=E.offp,x[0]^=E.key,E=O(f,c),c=E.offp,x[1]^=E.key,x=B(x,0,l,s),l[T]=x[0],l[T+1]=x[1];for(T=0;T<L;T+=2)E=O(f,c),c=E.offp,x[0]^=E.key,E=O(f,c),c=E.offp,x[1]^=E.key,x=B(x,0,l,s),s[T]=x[0],s[T+1]=x[1]}function ht(f,d,l,s,c){var x=q.slice(),y=x.length,L;if(l<4||l>31)if(L=Error("Illegal number of rounds (4-31): "+l),s){a(s.bind(this,L));return}else throw L;if(d.length!==S)if(L=Error("Illegal salt length: "+d.length+" != "+S),s){a(s.bind(this,L));return}else throw L;l=1<<l>>>0;var E,T,R=0,K;Int32Array?(E=new Int32Array(P),T=new Int32Array(A)):(E=P.slice(),T=A.slice()),Z(d,f,E,T);function X(){if(c&&c(R/l),R<l)for(var de=Date.now();R<l&&(R=R+1,z(f,E,T),z(d,E,T),!(Date.now()-de>$)););else{for(R=0;R<64;R++)for(K=0;K<y>>1;K++)B(x,K<<1,E,T);var V=[];for(R=0;R<y;R++)V.push((x[R]>>24&255)>>>0),V.push((x[R]>>16&255)>>>0),V.push((x[R]>>8&255)>>>0),V.push((x[R]&255)>>>0);if(s){s(null,V);return}else return V}s&&a(X)}if(typeof s<"u")X();else for(var ge;;)if(typeof(ge=X())<"u")return ge||[]}function Ue(f,d,l,s){var c;if(typeof f!="string"||typeof d!="string")if(c=Error("Invalid string / salt: Not a string"),l){a(l.bind(this,c));return}else throw c;var x,y;if(d.charAt(0)!=="$"||d.charAt(1)!=="2")if(c=Error("Invalid salt version: "+d.substring(0,2)),l){a(l.bind(this,c));return}else throw c;if(d.charAt(2)==="$")x="\0",y=3;else{if(x=d.charAt(2),x!=="a"&&x!=="b"&&x!=="y"||d.charAt(3)!=="$")if(c=Error("Invalid salt revision: "+d.substring(2,4)),l){a(l.bind(this,c));return}else throw c;y=4}if(d.charAt(y+2)>"$")if(c=Error("Missing salt rounds"),l){a(l.bind(this,c));return}else throw c;var L=parseInt(d.substring(y,y+1),10)*10,E=parseInt(d.substring(y+1,y+2),10),T=L+E,R=d.substring(y+3,y+25);f+=x>="a"?"\0":"";var K=i(f),X=v(R,S);function ge(de){var V=[];return V.push("$2"),x>="a"&&V.push(x),V.push("$"),T<10&&V.push("0"),V.push(T.toString()),V.push("$"),V.push(u(X,X.length)),V.push(u(de,q.length*4-1)),V.join("")}if(typeof l>"u")return ge(ht(K,X,T));ht(K,X,T,function(de,V){de?l(de,null):l(null,ge(V))},s)}return e.encodeBase64=u,e.decodeBase64=v,e})});var tt=Q(Le=>{"use strict";Object.defineProperty(Le,"__esModule",{value:!0});Le.crossEnvCrypto=void 0;try{let e=typeof window<"u"&&window.crypto||ve("crypto").webcrypto;if(e){if(!e.subtle)throw new Error("Crypto.Subtle is undefined. Make sure you are using HTTPS in a compatible browser.");let t=n=>r=>e.subtle.digest(n,r);Le.crossEnvCrypto={randomBytes:e.getRandomValues.bind(e),hashFunctions:{SHA1:t("SHA-1"),SHA256:t("SHA-256"),SHA384:t("SHA-384"),SHA512:t("SHA-512")}}}else{let t=ve("crypto"),n=r=>o=>t.createHash(r).update(o).digest().buffer;Le.crossEnvCrypto={randomBytes:t.randomFillSync,hashFunctions:{SHA1:n("sha1"),SHA256:n("sha256"),SHA384:n("sha384"),SHA512:n("sha512")}}}}catch(e){throw console.error(e),new Error("No suitable crypto library was found. You may need a polyfill.")}});var sn=Q(Ne=>{"use strict";Object.defineProperty(Ne,"__esModule",{value:!0});Ne.SRPParameters=void 0;var Oe=tt(),Ae=class e{constructor(t=e.PrimeGroup[2048],n=e.H.SHA512){if(this.primeGroup=t,this.H=n,this.NBits=this.primeGroup.N.toString(2).length,!n)throw new Error("Hash function required")}};Ne.SRPParameters=Ae;Ae.PrimeGroup={256:{N:BigInt("125617018995153554710546479714086468244499594888726646874671447258204721048803"),g:BigInt(2)},512:{N:BigInt("11144252439149533417835749556168991736939157778924947037200268358613863350040339017097790259154750906072491181606044774215413467851989724116331597513345603"),g:BigInt(2)},768:{N:BigInt("1087179135105457859072065649059069760280540086975817629066444682366896187793570736574549981488868217843627094867924800342887096064844227836735667168319981288765377499806385489913341488724152562880918438701129530606139552645689583147"),g:BigInt(2)},1024:{N:BigInt("167609434410335061345139523764350090260135525329813904557420930309800865859473551531551523800013916573891864789934747039010546328480848979516637673776605610374669426214776197828492691384519453218253702788022233205683635831626913357154941914129985489522629902540768368409482248290641036967659389658897350067939"),g:BigInt(2)},1536:{N:BigInt("1486998185923128292816507353619409521152457662596380074614818966810244974827752411420380336514078832314731499938313197533147998565301020797040787428051479639316928015998415709101293902971072960487527411068082311763171549170528008620813391411445907584912865222076100726050255271567749213905330659264908657221124284665444825474741087704974475795505492821585749417639344967192301749033325359286273431675492866492416941152646940908101472416714421046022696100064262587"),g:BigInt(2)},2048:{N:BigInt("21766174458617435773191008891802753781907668374255538511144643224689886235383840957210909013086056401571399717235807266581649606472148410291413364152197364477180887395655483738115072677402235101762521901569820740293149529620419333266262073471054548368736039519702486226506248861060256971802984953561121442680157668000761429988222457090413873973970171927093992114751765168063614761119615476233422096442783117971236371647333871414335895773474667308967050807005509320424799678417036867928316761272274230314067548291133582479583061439577559347101961771406173684378522703483495337037655006751328447510550299250924469288819"),g:BigInt(2)}};Ae.H={SHA1:Oe.crossEnvCrypto.hashFunctions.SHA1,SHA256:Oe.crossEnvCrypto.hashFunctions.SHA256,SHA384:Oe.crossEnvCrypto.hashFunctions.SHA384,SHA512:Oe.crossEnvCrypto.hashFunctions.SHA512}});var $e=Q(I=>{"use strict";var rt=I&&I.__awaiter||function(e,t,n,r){function o(a){return a instanceof n?a:new n(function(i){i(a)})}return new(n||(n=Promise))(function(a,i){function p(u){try{b(r.next(u))}catch(v){i(v)}}function m(u){try{b(r.throw(u))}catch(v){i(v)}}function b(u){u.done?a(u.value):o(u.value).then(p,m)}b((r=r.apply(e,t||[])).next())})};Object.defineProperty(I,"__esModule",{value:!0});I.modPow=I.hashBitCount=I.createVerifierAndSalt=I.createVerifier=I.generateRandomBigInt=I.generateRandomString=I.hashPadded=I.hash=I.padStartArrayBuffer=I.stringToArrayBuffer=I.arrayBufferToBigInt=I.bigIntToArrayBuffer=void 0;var dn=tt(),jn=BigInt(0),nt=BigInt(1),ln=BigInt(2),zn=e=>{let t=e.toString(16),n=new ArrayBuffer(Math.ceil(t.length/2)),r=new Uint8Array(n),o=0;t.length%2!==0&&(r[0]=parseInt(t[0],16),o=1);for(let a=0;a<n.byteLength;a++)r[a+o]=parseInt(t.slice(2*a+o,2*a+2+o),16);return n};I.bigIntToArrayBuffer=zn;var Yn=e=>{let t=[];return new Uint8Array(e).forEach(n=>{t.push(("0"+n.toString(16)).slice(-2))}),BigInt(`0x${t.join("")}`)};I.arrayBufferToBigInt=Yn;function Zn(e){return new TextEncoder().encode(e).buffer}I.stringToArrayBuffer=Zn;var Xn=(e,t)=>{let n=new Uint8Array(e);if(n.length<t){let r=new Uint8Array(t);return r.fill(0,0,t-n.length),r.set(n,t-n.length),r}return n};I.padStartArrayBuffer=Xn;function at(e,...t){let n=t.reduce((o,a)=>o+a.byteLength,0),r=new Uint8Array(n);for(let o=0,a=0;a<t.length;a++)r.set(new Uint8Array(t[a]),o),o+=t[a].byteLength;return e.H(r)}I.hash=at;function Jn(e,t,...n){let r=n.map(o=>I.padStartArrayBuffer(o,t));return at(e,...r)}I.hashPadded=Jn;function Wn(e=10){let t=new Uint8Array(Math.ceil(Math.ceil(e/2)));return dn.crossEnvCrypto.randomBytes(t),t.reduce((n,r)=>{let o=r.toString(16).toString();return o.length===1?n+"0"+o:n+o},"").slice(0,e)}I.generateRandomString=Wn;function Qn(e=16){return I.arrayBufferToBigInt(rr(e))}I.generateRandomBigInt=Qn;function un(e,t,n,r){return rt(this,void 0,void 0,function*(){if(!t||!t.trim())throw new Error("Identity (I) must not be null or empty.");if(!n)throw new Error("Salt (s) must not be null.");if(!r)throw new Error("Password (P) must not be null");let o=yield e.computeX(t,n,r);return e.computeVerifier(o)})}I.createVerifier=un;function er(e,t,n,r){return rt(this,void 0,void 0,function*(){let o=yield e.generateRandomSalt(r);return{s:o,v:yield un(e,t,o,n)}})}I.createVerifierAndSalt=er;var tr=e=>rt(void 0,void 0,void 0,function*(){return(yield at(e,I.bigIntToArrayBuffer(BigInt(1)))).byteLength*8});I.hashBitCount=tr;function nr(e,t,n){let r=nt;for(;t>jn;)t%ln==nt?(r=e*r%n,t-=nt):(e=e*e%n,t/=ln);return r}I.modPow=nr;var rr=e=>{let t=new Uint8Array(e);return dn.crossEnvCrypto.randomBytes(t),t.buffer}});var fn=Q(he=>{"use strict";var te=he&&he.__awaiter||function(e,t,n,r){function o(a){return a instanceof n?a:new n(function(i){i(a)})}return new(n||(n=Promise))(function(a,i){function p(u){try{b(r.next(u))}catch(v){i(v)}}function m(u){try{b(r.throw(u))}catch(v){i(v)}}function b(u){u.done?a(u.value):o(u.value).then(p,m)}b((r=r.apply(e,t||[])).next())})};Object.defineProperty(he,"__esModule",{value:!0});he.SRPRoutines=void 0;var M=$e(),ot=class{constructor(t){this.parameters=t}hash(...t){return M.hash(this.parameters,...t)}hashPadded(...t){let n=Math.trunc((this.parameters.NBits+7)/8);return M.hashPadded(this.parameters,n,...t)}computeK(){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hashPadded(M.bigIntToArrayBuffer(this.parameters.primeGroup.N),M.bigIntToArrayBuffer(this.parameters.primeGroup.g)))})}generateRandomSalt(t){return te(this,void 0,void 0,function*(){let n=yield M.hashBitCount(this.parameters),r=t||2*n/8;return M.generateRandomBigInt(r)})}computeX(t,n,r){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hash(M.bigIntToArrayBuffer(n),yield this.computeIdentityHash(t,r)))})}computeXStep2(t,n){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hash(M.bigIntToArrayBuffer(t),n))})}computeIdentityHash(t,n){return te(this,void 0,void 0,function*(){return yield this.hash(M.stringToArrayBuffer(n))})}computeVerifier(t){return M.modPow(this.parameters.primeGroup.g,t,this.parameters.primeGroup.N)}generatePrivateValue(){let t=Math.max(256,this.parameters.NBits),n;do n=M.generateRandomBigInt(t/8)%this.parameters.primeGroup.N;while(n===BigInt(0));return n}computeClientPublicValue(t){return M.modPow(this.parameters.primeGroup.g,t,this.parameters.primeGroup.N)}isValidPublicValue(t){return t%this.parameters.primeGroup.N!==BigInt(0)}computeU(t,n){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hashPadded(M.bigIntToArrayBuffer(t),M.bigIntToArrayBuffer(n)))})}computeClientEvidence(t,n,r,o,a){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hash(M.bigIntToArrayBuffer(r),M.bigIntToArrayBuffer(o),M.bigIntToArrayBuffer(a)))})}computeServerEvidence(t,n,r){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hash(M.bigIntToArrayBuffer(t),M.bigIntToArrayBuffer(n),M.bigIntToArrayBuffer(r)))})}computeClientSessionKey(t,n,r,o,a){let i=this.parameters.primeGroup.N,p=r*n+o,m=M.modPow(this.parameters.primeGroup.g,n,i)*t%i;return M.modPow(a+i-m,p,i)}};he.SRPRoutines=ot});var xn=Q(W=>{"use strict";var st=W&&W.__awaiter||function(e,t,n,r){function o(a){return a instanceof n?a:new n(function(i){i(a)})}return new(n||(n=Promise))(function(a,i){function p(u){try{b(r.next(u))}catch(v){i(v)}}function m(u){try{b(r.throw(u))}catch(v){i(v)}}function b(u){u.done?a(u.value):o(u.value).then(p,m)}b((r=r.apply(e,t||[])).next())})};Object.defineProperty(W,"__esModule",{value:!0});W.SRPClientSessionStep2=W.SRPClientSessionStep1=W.SRPClientSession=void 0;var ct=class{constructor(t){this.routines=t}step1(t,n){return st(this,void 0,void 0,function*(){if(!t||!t.trim())throw new Error("User identity must not be null nor empty");if(!n)throw new Error("User password must not be null");let r=yield this.routines.computeIdentityHash(t,n);return new ke(this.routines,t,r)})}};W.SRPClientSession=ct;var ke=class e{constructor(t,n,r){this.routines=t,this.I=n,this.IH=r}step2(t,n){return st(this,void 0,void 0,function*(){if(!t)throw new Error("Salt (s) must not be null");if(!n)throw new Error("Public server value (B) must not be null");let r=yield this.routines.computeXStep2(t,this.IH),o=this.routines.generatePrivateValue(),a=this.routines.computeClientPublicValue(o),i=yield this.routines.computeK(),p=yield this.routines.computeU(a,n),m=this.routines.computeClientSessionKey(i,r,p,o,n),b=yield this.routines.computeClientEvidence(this.I,t,a,n,m);return new _e(this.routines,a,b,m)})}toJSON(){return{I:this.I,IH:Array.from(new Uint8Array(this.IH))}}static fromState(t,n){return new e(t,n.I,new Uint8Array(n.IH).buffer)}};W.SRPClientSessionStep1=ke;var _e=class e{constructor(t,n,r,o){this.routines=t,this.A=n,this.M1=r,this.S=o}step3(t){return st(this,void 0,void 0,function*(){if(!t)throw new Error("Server evidence (M2) must not be null");if((yield this.routines.computeServerEvidence(this.A,this.M1,this.S))!==t)throw new Error("Bad server credentials")})}toJSON(){return{A:this.A.toString(16),M1:this.M1.toString(16),S:this.S.toString(16)}}static fromState(t,n){return new e(t,BigInt("0x"+n.A),BigInt("0x"+n.M1),BigInt("0x"+n.S))}};W.SRPClientSessionStep2=_e});var pn=Q(ne=>{"use strict";var it=ne&&ne.__awaiter||function(e,t,n,r){function o(a){return a instanceof n?a:new n(function(i){i(a)})}return new(n||(n=Promise))(function(a,i){function p(u){try{b(r.next(u))}catch(v){i(v)}}function m(u){try{b(r.throw(u))}catch(v){i(v)}}function b(u){u.done?a(u.value):o(u.value).then(p,m)}b((r=r.apply(e,t||[])).next())})};Object.defineProperty(ne,"__esModule",{value:!0});ne.SRPServerSessionStep1=ne.SRPServerSession=void 0;var lt=$e(),dt=class{constructor(t){this.routines=t}step1(t,n,r){return it(this,void 0,void 0,function*(){let o=this.routines.generatePrivateValue(),a=yield this.routines.computeK(),i=ar(this.routines.parameters,a,r,o);return new Fe(this.routines,t,n,r,o,i)})}};ne.SRPServerSession=dt;var Fe=class e{constructor(t,n,r,o,a,i){this.routines=t,this.identifier=n,this.salt=r,this.verifier=o,this.b=a,this.B=i}sessionKey(t){return it(this,void 0,void 0,function*(){if(t===null)throw new Error("Client public value (A) must not be null");if(!this.routines.isValidPublicValue(t))throw new Error(`Invalid Client public value (A): ${t.toString(16)}`);let n=yield this.routines.computeU(t,this.B);return or(this.routines.parameters.primeGroup.N,this.verifier,n,t,this.b)})}step2(t,n){return it(this,void 0,void 0,function*(){if(!n)throw new Error("Client evidence (M1) must not be null");let r=yield this.sessionKey(t);if((yield this.routines.computeClientEvidence(this.identifier,this.salt,t,this.B,r))!==n)throw new Error("Bad client credentials");return this.routines.computeServerEvidence(t,n,r)})}toJSON(){return{identifier:this.identifier,salt:this.salt.toString(16),verifier:this.verifier.toString(16),b:this.b.toString(16),B:this.B.toString(16)}}static fromState(t,n){return new e(t,n.identifier,BigInt("0x"+n.salt),BigInt("0x"+n.verifier),BigInt("0x"+n.b),BigInt("0x"+n.B))}};ne.SRPServerSessionStep1=Fe;var ar=(e,t,n,r)=>(lt.modPow(e.primeGroup.g,r,e.primeGroup.N)+n*t)%e.primeGroup.N,or=(e,t,n,r,o)=>lt.modPow(lt.modPow(t,n,e)*r,o,e)});var mn=Q(_=>{"use strict";Object.defineProperty(_,"__esModule",{value:!0});_.generateRandomBigInt=_.arrayBufferToBigInt=_.bigIntToArrayBuffer=_.createVerifierAndSalt=_.SRPServerSessionStep1=_.SRPServerSession=_.SRPClientSessionStep2=_.SRPClientSessionStep1=_.SRPClientSession=_.SRPRoutines=_.SRPParameters=void 0;var cr=sn();Object.defineProperty(_,"SRPParameters",{enumerable:!0,get:function(){return cr.SRPParameters}});var sr=fn();Object.defineProperty(_,"SRPRoutines",{enumerable:!0,get:function(){return sr.SRPRoutines}});var ut=xn();Object.defineProperty(_,"SRPClientSession",{enumerable:!0,get:function(){return ut.SRPClientSession}});Object.defineProperty(_,"SRPClientSessionStep1",{enumerable:!0,get:function(){return ut.SRPClientSessionStep1}});Object.defineProperty(_,"SRPClientSessionStep2",{enumerable:!0,get:function(){return ut.SRPClientSessionStep2}});var bn=pn();Object.defineProperty(_,"SRPServerSession",{enumerable:!0,get:function(){return bn.SRPServerSession}});Object.defineProperty(_,"SRPServerSessionStep1",{enumerable:!0,get:function(){return bn.SRPServerSessionStep1}});var Ke=$e();Object.defineProperty(_,"createVerifierAndSalt",{enumerable:!0,get:function(){return Ke.createVerifierAndSalt}});Object.defineProperty(_,"bigIntToArrayBuffer",{enumerable:!0,get:function(){return Ke.bigIntToArrayBuffer}});Object.defineProperty(_,"arrayBufferToBigInt",{enumerable:!0,get:function(){return Ke.arrayBufferToBigInt}});Object.defineProperty(_,"generateRandomBigInt",{enumerable:!0,get:function(){return Ke.generateRandomBigInt}})});var gt=document.querySelector("#warning"),ue=[],An=(()=>{try{if(window.BigInt instanceof Function)return!0}catch{}return!1})();An||ue.push("BigInt is not supported in this browser. Many operations, including SRP, will fail.");var Tn=(()=>{try{if(window.indexedDB instanceof IDBFactory)return!0}catch{}return!1})();Tn||ue.push("IndexedDB is not supported in this browser. Cryptographic keys cannot be stored.");var Cn=(()=>{try{if(window.crypto.getRandomValues instanceof Function)return!0}catch{}return!1})();Cn||ue.push("This browser cannot securely source random data. Many operations will fail. The outputs of any operations that do not fail should not be trusted.");var In=(()=>{try{if(window.crypto.subtle.constructor instanceof Function)return!0}catch{}return!1})();In||ue.push("The WebCrypto API subtle crypto interface is not available. Is this page being delivered over HTTPS? Many operations will fail. The outputs of any operations that do not fail should not be trusted.");if(ue.length){gt.classList.add("active");let e=gt.querySelector(".content");ue.forEach(t=>{let n=document.createElement("p");n.textContent=t,e.appendChild(n)})}var Ce=document.querySelector("#loader"),Bn=e=>{Ce&&e>=0&&e<=100&&(Ce.style.width=`${e}%`),e===0?Ce?.classList.remove("animate"):Ce?.classList.add("animate")},N=Bn;var Ge=document.querySelector("#error"),vt=Ge.querySelector(".content"),C=e=>{let t;e instanceof Error?t=e:t=new Error(String(e)),Ge.classList.add("active"),vt.textContent=t.message,N(100)},je=()=>{vt.textContent="",Ge.classList.remove("active")};var Ie=De(ze()),J=(h=>(h[h.BOOLEAN=-1]="BOOLEAN",h[h.BIGINT=-2]="BIGINT",h[h.INTEGER=-3]="INTEGER",h[h.UUID=-4]="UUID",h[h.JSON=-5]="JSON",h[h.PEM=-6]="PEM",h[h.UNKNOWN=0]="UNKNOWN",h[h.BINARY=2]="BINARY",h[h.OCTAL=8]="OCTAL",h[h["UTF-8"]=100]="UTF-8",h[h.HEXADECIMAL=16]="HEXADECIMAL",h[h.BASE64=64]="BASE64",h[h.BASE64_CRYPT=640]="BASE64_CRYPT",h[h.PHC_STRING=-100]="PHC_STRING",h[h.MODULAR_CRYPT_FORMAT=-200]="MODULAR_CRYPT_FORMAT",h))(J||{}),j=(e,t)=>{let n=new Uint8Array(e);if(t===64)return btoa(n.reduce((o,a)=>o+String.fromCharCode(a),""));if(t===640)return Ie.encodeBase64(n,n.length);if(t===100)return new TextDecoder().decode(e);let r=0;if(t===2&&(r=8),t===8&&(r=3),t===16&&(r=2),!r)throw new Error(`Radix ${t} is not supported`);return n.reduce((o,a)=>o+a.toString(t).padStart(r,"0"),"")},U=(e,t)=>{let n;if(t===-1)return n=new Uint8Array(1),n[0]=Number(JSON.parse(e)),n.buffer;if(t===64){let a=atob(e);n=new Uint8Array(a.length);for(let i=0;i<a.length;i+=1)n[i]=a.charCodeAt(i);return n.buffer}if(t===640){let a=Ie.decodeBase64(e,1/0);return Uint8Array.from(a).buffer}if(t===-2){let a=BigInt(e),i=[];for(;a>0;)i.push(Number(a&0xffn)),a>>=8n;return Uint8Array.from(i.reverse()).buffer}if(t===-4){let a=e.replace(/-/g,"");return n=new Uint8Array(a.length/2),a.match(/.{1,2}/g)?.forEach((i,p)=>{n[p]=parseInt(i,16)}),n.buffer}if(t===100)return new TextEncoder().encode(e);let r=0;if(t===2&&(r=8),t===8&&(r=3),t===16&&(r=2),!r)throw new Error(`Radix ${t} is not supported`);let o=e.length/r;return n=new Uint8Array(o),e.match(new RegExp(`.{1,${r}}`,"g"))?.forEach((a,i)=>{n[i]=parseInt(a,t)}),n.buffer},fe=e=>{if(e.toLowerCase()==="true"||e.toLowerCase()==="false")return-1;if(/^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i.test(e))return-4;if(/^([01]{8})+$/.test(e))return 2;if(/^([0-7]{3})+$/.test(e))return 8;if(/^([0-9a-f]{2})+$|^([0-9A-F]{2})+$/.test(e))return 16;if(/^([0-9a-zA-Z+/]{4})*[0-9a-zA-Z+/]{2}[0-9a-zA-Z+/=]{2}$/.test(e))return 64;if(/^-{5}BEGIN .+-{5}(\r\n?|\n)([0-9a-zA-Z+/]{4})*[0-9a-zA-Z+/]{2}[0-9a-zA-Z+/=]{2}(\r\n?|\n)-{5}END .+-{5}(\r\n?|\n)?$/.test(e))return-6;if(/^[0-9a-zA-Z./]+$/.test(e))return 640;if(/{/.test(e))try{return JSON.parse(e),-5}catch{}return/^\$[a-z0-9-]{4,32}(\$v=[0-9]+)?(\$[a-z0-9-]{1,32}=[a-zA-Z0-9/+.-]+(,[a-z0-9-]{1,32}=[a-zA-Z0-9/+.-]+)*)?(\$[a-zA-Z0-9/+.-]+(\$[a-zA-Z0-9/+]{2,})?)?$/.test(e)?-100:/^\$[a-zA-Z0-9]+(\$[a-zA-Z0-9./]+)+$/.test(e)?-200:0};var St='<svg viewBox="0 0 24 24"><path fill="currentColor" d="M7.41 8.58L12 13.17l4.59-4.59L18 10l-6 6l-6-6z"></path></svg>',Ye='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M240 128c-64 0 0 88-64 88H80c-64 0 0-88-64-88c64 0 0-88 64-88h96c64 0 0 88 64 88" opacity=".2"/><path d="M43.18 128a29.78 29.78 0 0 1 8 10.26c4.8 9.9 4.8 22 4.8 33.74c0 24.31 1 36 24 36a8 8 0 0 1 0 16c-17.48 0-29.32-6.14-35.2-18.26c-4.8-9.9-4.8-22-4.8-33.74c0-24.31-1-36-24-36a8 8 0 0 1 0-16c23 0 24-11.69 24-36c0-11.72 0-23.84 4.8-33.74C50.68 38.14 62.52 32 80 32a8 8 0 0 1 0 16c-23 0-24 11.69-24 36c0 11.72 0 23.84-4.8 33.74A29.78 29.78 0 0 1 43.18 128M240 120c-23 0-24-11.69-24-36c0-11.72 0-23.84-4.8-33.74C205.32 38.14 193.48 32 176 32a8 8 0 0 0 0 16c23 0 24 11.69 24 36c0 11.72 0 23.84 4.8 33.74a29.78 29.78 0 0 0 8 10.26a29.78 29.78 0 0 0-8 10.26c-4.8 9.9-4.8 22-4.8 33.74c0 24.31-1 36-24 36a8 8 0 0 0 0 16c17.48 0 29.32-6.14 35.2-18.26c4.8-9.9 4.8-22 4.8-33.74c0-24.31 1-36 24-36a8 8 0 0 0 0-16"/></g></svg>',wt='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M184 72v144H40V72Z" opacity=".2"/><path d="M184 64H40a8 8 0 0 0-8 8v144a8 8 0 0 0 8 8h144a8 8 0 0 0 8-8V72a8 8 0 0 0-8-8m-8 144H48V80h128Zm48-168v144a8 8 0 0 1-16 0V48H72a8 8 0 0 1 0-16h144a8 8 0 0 1 8 8"/></g></svg>',Lt='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M216 91.31V208a8 8 0 0 1-8 8h-32v-64a8 8 0 0 0-8-8H88a8 8 0 0 0-8 8v64H48a8 8 0 0 1-8-8V48a8 8 0 0 1 8-8h116.69a8 8 0 0 1 5.65 2.34l43.32 43.31a8 8 0 0 1 2.34 5.66" opacity=".2"/><path d="M219.31 80L176 36.69A15.86 15.86 0 0 0 164.69 32H48a16 16 0 0 0-16 16v160a16 16 0 0 0 16 16h160a16 16 0 0 0 16-16V91.31A15.86 15.86 0 0 0 219.31 80M168 208H88v-56h80Zm40 0h-24v-56a16 16 0 0 0-16-16H88a16 16 0 0 0-16 16v56H48V48h116.69L208 91.31ZM160 72a8 8 0 0 1-8 8H96a8 8 0 0 1 0-16h56a8 8 0 0 1 8 8"/></g></svg>',At='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M232 56v144a16 16 0 0 1-16 16H40a16 16 0 0 1-16-16V56a16 16 0 0 1 16-16h176a16 16 0 0 1 16 16" opacity=".2"/><path d="m205.66 85.66l-96 96a8 8 0 0 1-11.32 0l-40-40a8 8 0 0 1 11.32-11.32L104 164.69l90.34-90.35a8 8 0 0 1 11.32 11.32"/></g></svg>',Tt='<svg viewBox="0 0 24 24"><path fill="currentColor" d="m17.16 7.59l-5.66 5.66l-5.66-5.66l.71-.7l4.95 4.95l4.95-4.95zm0 4l-5.66 5.66l-5.66-5.66l.71-.7l4.95 4.95l4.95-4.95z"/></svg>',Ct='<svg viewBox="0 0 16 16"><path fill="currentColor" fill-rule="evenodd" d="m10.57 1.14l3.28 3.3l.15.36v9.7l-.5.5h-11l-.5-.5v-13l.5-.5h7.72zM3 2v12h10V5l-3-3zm1.46 4.052c0 1.287.458 1.93 1.374 1.93c.457 0 .807-.173 1.05-.52c.246-.348.368-.847.368-1.499C7.252 4.654 6.805 4 5.91 4c-.471 0-.831.175-1.08.526c-.247.35-.37.858-.37 1.526m.862-.022c0-.922.183-1.383.55-1.383c.344 0 .516.448.516 1.343s-.176 1.343-.527 1.343c-.36 0-.54-.434-.54-1.303zm3.187 1.886h2.435v-.672h-.792V4l-1.665.336v.687l.82-.177v2.398h-.798zm-1.337 5H4.736v-.672h.798V9.846l-.82.177v-.687L6.38 9v3.244h.792v.671zm1.035-1.931c0 1.287.458 1.93 1.375 1.93c.457 0 .807-.173 1.05-.52c.245-.348.368-.847.368-1.499c0-1.309-.448-1.963-1.343-1.963c-.47 0-.83.175-1.08.526c-.246.35-.37.858-.37 1.526m.862-.022c0-.922.184-1.383.55-1.383c.344 0 .516.448.516 1.343s-.175 1.343-.526 1.343c-.36 0-.54-.434-.54-1.303" clip-rule="evenodd"/></svg>',It='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M232 136v64a8 8 0 0 1-8 8H32a8 8 0 0 1-8-8v-64a8 8 0 0 1 8-8h192a8 8 0 0 1 8 8" opacity=".2"/><path d="M240 136v64a16 16 0 0 1-16 16H32a16 16 0 0 1-16-16v-64a16 16 0 0 1 16-16h40a8 8 0 0 1 0 16H32v64h192v-64h-40a8 8 0 0 1 0-16h40a16 16 0 0 1 16 16m-117.66-2.34a8 8 0 0 0 11.32 0l48-48a8 8 0 0 0-11.32-11.32L136 108.69V24a8 8 0 0 0-16 0v84.69L85.66 74.34a8 8 0 0 0-11.32 11.32ZM200 168a12 12 0 1 0-12 12a12 12 0 0 0 12-12"/></g></svg>',Be='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M232 96a72 72 0 0 1-98.83 66.83L120 176H96v24H72v24H32v-40l61.17-61.17A72 72 0 1 1 232 96" opacity=".2"/><path d="M160 16a80.07 80.07 0 0 0-76.09 104.78l-57.57 57.56A8 8 0 0 0 24 184v40a8 8 0 0 0 8 8h40a8 8 0 0 0 8-8v-16h16a8 8 0 0 0 8-8v-16h16a8 8 0 0 0 5.66-2.34l9.56-9.57A80 80 0 1 0 160 16m0 144a63.7 63.7 0 0 1-23.65-4.51a8 8 0 0 0-8.84 1.68L116.69 168H96a8 8 0 0 0-8 8v16H72a8 8 0 0 0-8 8v16H40v-28.69l58.83-58.82a8 8 0 0 0 1.68-8.84A64 64 0 1 1 160 160m32-84a12 12 0 1 1-12-12a12 12 0 0 1 12 12"/></g></svg>',Bt='<svg viewBox="0 0 512 512"><path fill="currentColor" d="M261.1 24.8c-6.3 0-12.7.43-19.2 1.18c-34.6 4.01-64.8 17.59-86.1 37.06c-21.4 19.48-34.2 45.56-31 73.16c2.8 24.6 17.8 45.2 39.1 59.4c2.6-6.2 5.9-11.9 9.2-16.5c-17.6-11.6-28.4-27.3-30.4-45c-2.3-19.7 6.7-39.58 24.8-56.14c18.2-16.57 45.3-29.06 76.6-32.68c31.3-3.63 60.6 2.33 82.1 14.3c21.4 11.98 34.7 29.31 37 48.92c2.2 19.3-6.2 38.8-23.4 55a69.91 69.91 0 0 0-35.4-10.6h-2.2c-5.1.1-10.1.7-15.3 1.8c-37.5 8.7-60.8 45.5-52.2 82.7c5.3 23 21.6 40.6 42.2 48.5l39.7 172.2l47 29.1l29.5-46.7l-23.5-14.5l14.8-23.4l-23.5-14.6l14.7-23.3l-23.5-14.6l14.8-23.4l-13.5-58.4c15.1-16.1 22-39.1 16.7-62.2c-2.7-11.7-8.2-22-15.8-30.4c18.9-19 29.8-43.5 26.8-69.2c-3.2-27.55-21.6-50.04-46.9-64.11c-20.5-11.45-45.8-17.77-73.1-17.59m-20.2 135.5c-25.9 1.1-49.9 16.8-60.4 42.2c-9.1 21.9-6 45.7 6.2 64.2l-67.8 163l21.3 51l51.2-20.9l-10.7-25.5l25.6-10.4l-10.6-25.5l25.6-10.4l-10.7-25.5l25.6-10.5l22.8-54.8c-20.5-11.5-36.2-31.2-41.9-55.8c-6.9-30.3 3.1-60.6 23.8-81.1m58 7.2c8.9-.1 17.3 3.5 23.4 9.4c-5.5 3.5-11.6 6.6-18 9.4c-1.6-.6-3.3-.8-5.1-.8c-.6 0-1.1 0-1.6.1c-7 .8-12.2 6.1-13.1 12.7c-.2 1-.2 2-.2 2.9c.1.3.1.7.1 1c1 8.4 8.3 14.2 16.7 13.2c6.8-.8 12-5.9 13-12.3c6.2-2.8 12-5.9 17.5-9.4c.2 1 .4 2 .5 3c2.1 18-11 34.5-29 36.6c-17.9 2.1-34.5-11-36.5-29c-2.1-18 11-34.5 29-36.6c1.1-.1 2.2-.2 3.3-.2"/></svg>',Ht='<svg viewBox="0 0 48 48"><defs><mask id="ipTPeopleDownload0"><g fill="none" stroke="#fff" stroke-linejoin="round" stroke-width="4"><path fill="#555" d="M19 20a7 7 0 1 0 0-14a7 7 0 0 0 0 14Z"/><path stroke-linecap="round" d="M37 40V30m-4 6l4 4l4-4m-14-8h-8.2c-4.48 0-6.72 0-8.432.872a8 8 0 0 0-3.496 3.496C6 34.08 6 36.32 6 40.8V42h21"/></g></mask></defs><path fill="currentColor" d="M0 0h48v48H0z" mask="url(#ipTPeopleDownload0)"/></svg>',Mt='<svg viewBox="0 0 24 24"><g fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"><path d="M21 12a9 9 0 1 0-9 9M3.6 9h16.8M3.6 15H12"/><path d="M11.578 3a17 17 0 0 0 0 18M12.5 3c1.719 2.755 2.5 5.876 2.5 9m3 2v7m-3-3l3 3l3-3"/></g></svg>',Rt='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M229.66 90.34L90.34 229.66a8 8 0 0 1-11.31 0L26.34 177a8 8 0 0 1 0-11.31L165.66 26.34a8 8 0 0 1 11.31 0L229.66 79a8 8 0 0 1 0 11.34" opacity=".2"/><path d="m235.32 73.37l-52.69-52.68a16 16 0 0 0-22.63 0L20.68 160a16 16 0 0 0 0 22.63l52.69 52.68a16 16 0 0 0 22.63 0L235.32 96a16 16 0 0 0 0-22.63M84.68 224L32 171.31l32-32l26.34 26.35a8 8 0 0 0 11.32-11.32L75.31 128L96 107.31l26.34 26.35a8 8 0 0 0 11.32-11.32L107.31 96L128 75.31l26.34 26.35a8 8 0 0 0 11.32-11.32L139.31 64l32-32L224 84.69Z"/></g></svg>',Pt='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M200 56v136a8 8 0 0 1-8 8H64a8 8 0 0 1-8-8V56Z" opacity=".2"/><path d="M208 56v32a8 8 0 0 1-16 0V64h-56v128h24a8 8 0 0 1 0 16H96a8 8 0 0 1 0-16h24V64H64v24a8 8 0 0 1-16 0V56a8 8 0 0 1 8-8h144a8 8 0 0 1 8 8"/></g></svg>',Ze='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M200 56v152a8 8 0 0 1-8 8H64a8 8 0 0 1-8-8V56Z" opacity=".2"/><path d="M216 48h-40v-8a24 24 0 0 0-24-24h-48a24 24 0 0 0-24 24v8H40a8 8 0 0 0 0 16h8v144a16 16 0 0 0 16 16h128a16 16 0 0 0 16-16V64h8a8 8 0 0 0 0-16M96 40a8 8 0 0 1 8-8h48a8 8 0 0 1 8 8v8H96Zm96 168H64V64h128Zm-80-104v64a8 8 0 0 1-16 0v-64a8 8 0 0 1 16 0m48 0v64a8 8 0 0 1-16 0v-64a8 8 0 0 1 16 0"/></g></svg>';var xe=document.querySelector("#results"),qt=(e,t,n)=>async()=>{let r,o=n===null?"":n||"bin";if(typeof e=="string"){if(r=new TextEncoder().encode(e),!n)switch(fe(e)){case-5:o="json";break;case-6:o="pem";break;default:o="txt"}}else r=e;let a=new Blob([r],{type:"application/octet-stream"}),i=window.URL.createObjectURL(a),p=document.createElement("a");p.href=i;let m=o?`.${o}`:"";p.download=t?`${t}${m}`:`cryptotools-result${m}`,document.body.appendChild(p),p.click(),document.body.removeChild(p),window.URL.revokeObjectURL(i)},Ot=({label:e,value:t,encoding:n,bitLength:r,rawData:o,filename:a,extension:i})=>{let p=document.createElement("div");p.classList.add("result");let m=document.createElement("h2");m.textContent=e,p.appendChild(m);let b=document.createElement("div");b.classList.add("content"),b.textContent=`${t.substring(0,1e3)}${t.length>1e3?"\u2026":""}`,p.appendChild(b);let u=document.createElement("div");u.classList.add("stats");let v=[{icon:Pt,tooltip:"Character length",statValue:`${t.length.toLocaleString()} characters`},{icon:Rt,tooltip:"Output length",statValue:`${r.toLocaleString()} bits`}],g=n>0&&typeof o=="object",S=[{tooltip:"Download Text",tooltipAfter:"Downloaded!",icon:It,callback:qt(t,a,g?`${J[n].toLocaleLowerCase()}.txt`:i)},{tooltip:"Copy",tooltipAfter:"Copied!",icon:wt,callback:()=>navigator.clipboard.writeText(b.textContent)}];if(g){S.unshift({tooltip:"Download Raw Data",tooltipAfter:"Downloaded!",icon:Ct,callback:qt(o,a,i)});let w=document.createElement("label"),$=document.createElement("span");w.classList.add("control"),$.classList.add("label"),$.textContent="Encoding",w.innerHTML=Ye,w.insertBefore($,w.firstChild);let P=document.createElement("select");P.addEventListener("change",()=>{let q=Number(P.selectedOptions[0].value);try{let B=j(o,q),O=Ot({label:e,value:B,encoding:q,bitLength:r,rawData:o,filename:a,extension:i});p.replaceWith(O)}catch(B){C(B)}}),Object.entries(J).forEach(([q,B])=>{if(typeof B=="number"&&B>0){let O=document.createElement("option");O.textContent=q.toLowerCase(),O.value=String(B),B===n&&(O.selected=!0),P.appendChild(O)}}),w.appendChild(P);let A=document.createElement("span");A.innerHTML=St,w.appendChild(A.firstChild),u.appendChild(w)}else v.unshift({icon:Ye,tooltip:"Encoding",statValue:J[n].toLowerCase()});v.forEach(({icon:w,tooltip:$,statValue:P})=>{let A=document.createElement("div");A.classList.add("datum"),$&&(A.dataset.tooltip=$),A.innerHTML=w;let q=document.createElement("span");q.textContent=P,A.appendChild(q),u.appendChild(A)}),p.appendChild(u);let h=document.createElement("div");return h.classList.add("links"),u.appendChild(h),S.forEach(({tooltip:w,tooltipAfter:$,icon:P,callback:A})=>{let q=document.createElement("a");q.dataset.tooltip=w,q.href="#",q.addEventListener("click",async B=>{B.preventDefault();try{await A(),q.innerHTML=At,q.dataset.tooltip=$,setTimeout(()=>{q.innerHTML=P,q.dataset.tooltip=w},5e3)}catch(O){console.error(O)}}),q.innerHTML=P,h.appendChild(q)}),p},F=e=>{xe.innerHTML=Tt,je(),xe.scrollIntoView({block:"start",behavior:"smooth"}),e.forEach(({label:t,value:n,defaultEncoding:r,filename:o,extension:a})=>{let i,p,m;if(n instanceof ArrayBuffer)i=r||16,p=j(n,i),m=n.byteLength;else{p=n,i=r||fe(n);try{m=U(n,i).byteLength}catch{console.warn(`Cannot calculate raw data byte length for "${t}" because no decoder has been implemented for \`${J[i].toLowerCase()}\`. Assuming utf-8 encoded text instead.`),m=new TextEncoder().encode(n).byteLength}}try{let b=Ot({label:t,value:p,encoding:i,bitLength:m*8,rawData:n,filename:o,extension:a});xe.appendChild(b)}catch(b){C(b)}}),xe.style.opacity="100%",N(100)},Y=()=>{xe.style.opacity="0",xe.textContent="",je()};var Nt=document.querySelectorAll("#main-menu li"),Hn=document.querySelectorAll("main > section");Nt.forEach(e=>e.addEventListener("click",()=>{Nt.forEach(t=>t.classList.remove("active")),Hn.forEach(t=>t.classList.remove("active")),e.classList.add("active"),e.dataset.target&&document.querySelector(e.dataset.target)?.classList.add("active"),Y()}));var oe=document.querySelector("#digest"),He=oe.querySelector(".digest-select"),kt=oe.querySelector("textarea"),Ee=oe.querySelector("details"),Xe=oe.querySelector("button");Ee?.addEventListener("toggle",()=>{Ee.open?Xe.textContent="Verify Digest":Xe.textContent="Generate Digest"});async function Mn(e,t){let n=Number(oe.querySelector(".encoding select")?.selectedOptions[0].value),r=U(e,n);return await crypto.subtle.digest(t,r)}Xe.addEventListener("click",async()=>{N(0);let e=He.selectedOptions[0].dataset.alg,t=kt.value;try{let n=await Mn(t,e);if(Ee?.open){let r=Number(Ee.querySelector("select.input-encoding")?.selectedOptions[0].value),o=Ee.querySelector("input.hash")?.value||"",a=U(o,r);F([{label:`${e} Digest Verified?`,value:String(j(n,16)===j(a,16))}])}else F([{label:`${e} Digest`,value:n,defaultEncoding:16,filename:"digest",extension:e.replace(/-/g,"").toLowerCase()}])}catch(n){C(n)}});He.addEventListener("change",()=>{let e=oe.querySelector("menu"),t=e.querySelector(".block-size span"),n=e.querySelector(".method span"),r=e.querySelector(".specification span"),o=He.selectedOptions[0].dataset;t.textContent=o.bs||"",n.textContent=o.method||"",r.textContent=o.spec||"",Y()});var Rn=async(e,t,n)=>{let r=[];try{let o={};t&&t.split(/\r?\n/).forEach(a=>{let[i,p]=a.split(/\s[\s*]/);o[p]=i}),await Promise.all(e.map(async([a,i])=>{let p=o[a.name],m=await crypto.subtle.digest(n,i),b,u,v="digest",g;p?(b=`${n} Digest of ${a.name} Verifies Checksum? \u2022 ${a.size.toLocaleString()} bytes \u2022 ${a.type||"Unknown type"}`,u=String(p===j(m,16))):(b=`${n} Digest of ${a.name} \u2022 ${a.size.toLocaleString()} bytes \u2022 ${a.type||"Unknown type"}`,u=m,v=a.name,g=n.replace(/-/g,"").toLowerCase()),r.push({label:b,value:u,filename:v,extension:g})}))}catch(o){C(o)}F(r)},_t=e=>{if(!e||!e.length)return;N(0);let t=Array.from(e),n=t.reduce((p,m)=>p+m.size,0),r=[],o=[],a=He.selectedOptions[0].dataset.alg,i;t.forEach((p,m)=>{let b=new FileReader;b.onprogress=u=>{r[m]=u.loaded,N(r.reduce((v,g)=>v+g,0)/n*100)},b.onload=async u=>{if(!u.target?.result||!(u.target.result instanceof ArrayBuffer)){C(new Error(`File [${p.name}] of type [${p.type}] failed to load.`));return}try{let g=new Error("More than one checksum file exists in the file list.");switch(!0){case/^.*\.sha1$|^sha1sums?(\.txt)?$/i.test(p.name):if(a="SHA-1",i)throw g;i=new TextDecoder().decode(u.target.result);break;case/^.*\.sha256$|^sha256sums?(\.txt)?$/i.test(p.name):if(a="SHA-256",i)throw g;i=new TextDecoder().decode(u.target.result);break;case/^.*\.sha384$|^sha384sums?(\.txt)?$/i.test(p.name):if(a="SHA-384",i)throw g;i=new TextDecoder().decode(u.target.result);break;case/^.*\.sha512$|^sha512sums?(\.txt)?$/i.test(p.name):if(a="SHA-512",i)throw g;i=new TextDecoder().decode(u.target.result);break;default:o.push([p,u.target.result])}}catch(g){C(g)}o.length===(i?e.length-1:e.length)&&Rn(o,i,a)},b.readAsArrayBuffer(p)})},$t=oe.querySelector(".upload input");$t?.addEventListener("change",()=>_t($t.files));kt.addEventListener("drop",e=>{_t(e.dataTransfer?.files)});var pe=document.querySelector("#random"),Ft=pe.querySelectorAll("menu li"),Pn=pe.querySelectorAll(".settings");Ft.forEach(e=>e.addEventListener("click",()=>{Ft.forEach(t=>t.classList.remove("active")),Pn.forEach(t=>t.classList.remove("active")),e.classList.add("active"),e.dataset.target&&document.querySelector(`.${e.dataset.target}`)?.classList.add("active"),Y()}));var qn=pe.querySelector("button");qn.addEventListener("click",()=>{N(0);try{let e=pe.querySelector(".byte-length"),t=pe.querySelector(".output");if(pe.querySelector("menu li.active").dataset.target==="uuid"){F([{label:"UUID",value:window.crypto.randomUUID(),filename:"prng-uuid"}]);return}let r=Number(e.value.length?e.value:64),o=new Uint8Array(r);if(window.crypto.getRandomValues(o),t.selectedOptions[0].value==="display"){F([{label:"Random Values",value:o.buffer,filename:`prng-${r}-bytes`}]);return}let i=new Blob([o],{type:"application/octet-stream"}),p=window.URL.createObjectURL(i),m=document.createElement("a");m.href=p,m.download=`prng-${r}-bytes.bin`,document.body.appendChild(m),m.click(),document.body.removeChild(m),window.URL.revokeObjectURL(p)}catch(e){C(e)}N(100)});var Kt=e=>t=>{t.preventDefault(),e.remove()},On=(e,t,n,r,o)=>{let a=document.createElement("div");a.classList.add("modal-background");let i=document.createElement("aside");i.classList.add("modal"),a.appendChild(i);let p=document.createElement("div");p.classList.add("container"),i.appendChild(p);let m=document.createElement("span");m.innerHTML=e,p.appendChild(m);let b=document.createElement("div");b.classList.add("content"),p.appendChild(b);let u=document.createElement("span");u.textContent=t,b.appendChild(u);let v=document.createElement("p");v.textContent=n,b.appendChild(v);let g=document.createElement("div");g.classList.add("actions"),i.appendChild(g);let S=document.createElement("button");S.classList.add("cancel"),S.textContent="Cancel",S.addEventListener("click",Kt(a)),g.appendChild(S);let h=document.createElement("button");h.classList.add("confirm"),h.textContent=r,h.addEventListener("click",Kt(a)),h.addEventListener("click",()=>o()),g.appendChild(h),document.body.appendChild(a)},Vt=On;var Nn="cryptotools",se="keys";var ce,Ut=()=>new Promise((e,t)=>{let n=window.indexedDB.open(Nn,1);n.onerror=()=>{t(new Error("Cannot open database."))},n.onupgradeneeded=()=>{ce=n.result,ce.createObjectStore(se,{keyPath:"name"})},n.onsuccess=()=>{ce=n.result,ce.onerror=r=>console.error(new Error(`A database error ocurred: ${JSON.stringify(r.target)}`)),e(!0)}}),Dt=(e,t)=>new Promise((n,r)=>{let i=ce.transaction([se],"readwrite").objectStore(se).add({name:e,key:t});i.onerror=p=>r(new Error(JSON.stringify(p))),i.onsuccess=()=>n(!0)}),Gt=e=>new Promise((t,n)=>{let a=ce.transaction([se],"readwrite").objectStore(se).delete(e);a.onerror=i=>n(new Error(JSON.stringify(i))),a.onsuccess=()=>t(!0)}),jt=()=>new Promise((e,t)=>{let o=ce.transaction([se],"readonly").objectStore(se).getAll();o.onerror=a=>t(new Error(JSON.stringify(a))),o.onsuccess=a=>e(a.target.result)});var zt=document.querySelector("#encryption"),D=zt.querySelector(".operation-area"),Yt=D?.querySelectorAll("menu li");zt.querySelectorAll(".keys menu li").forEach(e=>e.addEventListener("click",()=>D?.classList.remove("active")));var $n=(e,t)=>{let n=D?.querySelector("textarea"),r=D?.querySelector(".signature"),o=D?.querySelector("button"),a="";switch(e){case"encrypt":a="Paste plain text or drop a file...";break;case"decrypt":a="Paste ciphertext or drop a file...";break;case"sign":a="Paste plain text or drop a file...";break;case"verify":a="Paste plain text or drop a file...";break;default:}n&&(n.placeholder=a),D?.querySelectorAll(".input > .settings").forEach(i=>i.classList.remove("active")),D?.querySelector(`.settings.${t}.${e}`)?.classList.add("active"),e==="verify"?r?.classList.add("active"):r?.classList.remove("active"),o&&(o.textContent=`${e.charAt(0).toUpperCase()}${e.slice(1)}`)},Zt=(e,t,n)=>()=>{D?.classList.add("active"),D?.scrollIntoView({block:"center",behavior:"smooth"});let r=t instanceof CryptoKey,o=n,a=e,i,p;if(r)i=t,p=i.usages,o||(p.includes("encrypt")?o="encrypt":p.includes("sign")?o="sign":[o]=p);else{p=t.privateKey.usages.concat(t.publicKey.usages),o||(p.includes("encrypt")?o="encrypt":p.includes("sign")?o="sign":[o]=p);let u="";switch(o){case"verify":case"encrypt":u="Public Key",i=t.publicKey;break;case"sign":case"decrypt":u="Private Key",i=t.privateKey;break;default:throw new Error(`Operation ${o} is not implemented for key "${e}".`)}a=`${e} (${u})`}let m=i.algorithm.name.toLowerCase();D&&(D.dataset.key=e,D.dataset.alg=m),Yt?.forEach(u=>u.classList.remove("active","show")),D?.querySelector(`menu li[data-op="${o}"]`)?.classList.add("active"),p.forEach(u=>D?.querySelector(`menu li[data-op="${u}"]`)?.classList.add("show"));let b=D?.querySelector(".footer .key");if(b){b.innerHTML=Be;let u=document.createElement("span");u.textContent=a,b.appendChild(u)}$n(o,m)};Yt?.forEach(e=>{let t=e.dataset.op;e.addEventListener("click",()=>{let n=D.dataset.key||"",{key:r}=Me(n);Zt(n,r,t)()})});var Xt=Zt;var ie={},Je=()=>{document.querySelector("#encryption .operation-area")?.classList.remove("active"),Y();let t=document.querySelector("#encryption .list-keys ul");t.textContent="";let n=Object.entries(ie);if(!n.length){let r=document.createElement("li");r.textContent="No Keys.  Generate a key to get started.",t.appendChild(r)}n.forEach(([r,o])=>{let{key:a}=o,i=document.createElement("li");i.addEventListener("click",Xt(r,a));let p=document.createElement("div");p.classList.add("container"),i.appendChild(p);let m=document.createElement("span");p.appendChild(m);let b=document.createElement("div");b.classList.add("meta"),p.appendChild(b);let u=document.createElement("span");if(u.textContent=r,b.appendChild(u),o.saved){let A=document.createElement("span");A.classList.add("saved"),A.dataset.tooltip="Persists after page refresh",A.innerHTML=Lt,u.appendChild(A)}let v=a instanceof CryptoKey,g=document.createElement("span"),S,h,w;v?(m.innerHTML=Be,S="Symmetric Key",h=a.algorithm.name,w=a.usages):(m.innerHTML=Bt,S="Asymmetric Key Pair",h=a.privateKey.algorithm.name,w=a.privateKey.usages.concat(a.publicKey.usages)),g.textContent=`${S} \u2022 ${h} \u2022 ${w.join(", ")}`,b.appendChild(g);let $=document.createElement("div");if($.classList.add("actions"),i.appendChild($),!v){let A=document.createElement("a");A.href="#",A.innerHTML=Mt,A.dataset.tooltip="Export Public Key",A.addEventListener("click",async q=>{q.preventDefault(),q.stopPropagation();let B=a.publicKey,O=`${r.replace(/\s/g,"-")}.public`,z=[];try{switch(B.algorithm.name){case"RSASSA-PKCS1-v1_5":case"RSA-PSS":case"RSA-OAEP":case"ECDSA":{let Z=await be(B,"spki")();z.push({label:`Public Key "${r}" in Subject Public Key Info (SPKI) Format, DER-Encoded`,value:Z,defaultEncoding:64,filename:O,extension:"der"},{label:`Public Key "${r}" in Subject Public Key Info (SPKI) Format, PEM-Encoded`,value:`-----BEGIN PUBLIC KEY-----
+"use strict";(()=>{var gn=Object.create;var yt=Object.defineProperty;var vn=Object.getOwnPropertyDescriptor;var En=Object.getOwnPropertyNames;var Sn=Object.getPrototypeOf,wn=Object.prototype.hasOwnProperty;var ve=(e=>typeof require<"u"?require:typeof Proxy<"u"?new Proxy(e,{get:(t,n)=>(typeof require<"u"?require:t)[n]}):e)(function(e){if(typeof require<"u")return require.apply(this,arguments);throw Error('Dynamic require of "'+e+'" is not supported')});var Q=(e,t)=>()=>(t||e((t={exports:{}}).exports,t),t.exports);var Ln=(e,t,n,r)=>{if(t&&typeof t=="object"||typeof t=="function")for(let o of En(t))!wn.call(e,o)&&o!==n&&yt(e,o,{get:()=>t[o],enumerable:!(r=vn(t,o))||r.enumerable});return e};var De=(e,t,n)=>(n=e!=null?gn(Sn(e)):{},Ln(t||!e||!e.__esModule?yt(n,"default",{value:e,enumerable:!0}):n,e));var ze=Q((Et,ee)=>{(function(e,t){typeof define=="function"&&define.amd?define([],t):typeof ve=="function"&&typeof ee=="object"&&ee&&ee.exports?ee.exports=t():(e.dcodeIO=e.dcodeIO||{}).bcrypt=t()})(Et,function(){"use strict";var e={},t=null;function n(f){if(typeof ee<"u"&&ee&&ee.exports)try{return ve("crypto").randomBytes(f)}catch{}try{var d;return(self.crypto||self.msCrypto).getRandomValues(d=new Uint32Array(f)),Array.prototype.slice.call(d)}catch{}if(!t)throw Error("Neither WebCryptoAPI nor a crypto module is available. Use bcrypt.setRandomFallback to set an alternative");return t(f)}var r=!1;try{n(1),r=!0}catch{}t=null,e.setRandomFallback=function(f){t=f},e.genSaltSync=function(f,d){if(f=f||h,typeof f!="number")throw Error("Illegal arguments: "+typeof f+", "+typeof d);f<4?f=4:f>31&&(f=31);var l=[];return l.push("$2a$"),f<10&&l.push("0"),l.push(f.toString()),l.push("$"),l.push(u(n(S),S)),l.join("")},e.genSalt=function(f,d,l){if(typeof d=="function"&&(l=d,d=void 0),typeof f=="function"&&(l=f,f=void 0),typeof f>"u")f=h;else if(typeof f!="number")throw Error("illegal arguments: "+typeof f);function s(c){a(function(){try{c(null,e.genSaltSync(f))}catch(x){c(x)}})}if(l){if(typeof l!="function")throw Error("Illegal callback: "+typeof l);s(l)}else return new Promise(function(c,x){s(function(y,L){if(y){x(y);return}c(L)})})},e.hashSync=function(f,d){if(typeof d>"u"&&(d=h),typeof d=="number"&&(d=e.genSaltSync(d)),typeof f!="string"||typeof d!="string")throw Error("Illegal arguments: "+typeof f+", "+typeof d);return Ue(f,d)},e.hash=function(f,d,l,s){function c(x){typeof f=="string"&&typeof d=="number"?e.genSalt(d,function(y,L){Ue(f,L,x,s)}):typeof f=="string"&&typeof d=="string"?Ue(f,d,x,s):a(x.bind(this,Error("Illegal arguments: "+typeof f+", "+typeof d)))}if(l){if(typeof l!="function")throw Error("Illegal callback: "+typeof l);c(l)}else return new Promise(function(x,y){c(function(L,E){if(L){y(L);return}x(E)})})};function o(f,d){for(var l=0,s=0,c=0,x=f.length;c<x;++c)f.charCodeAt(c)===d.charCodeAt(c)?++l:++s;return l<0?!1:s===0}e.compareSync=function(f,d){if(typeof f!="string"||typeof d!="string")throw Error("Illegal arguments: "+typeof f+", "+typeof d);return d.length!==60?!1:o(e.hashSync(f,d.substr(0,d.length-31)),d)},e.compare=function(f,d,l,s){function c(x){if(typeof f!="string"||typeof d!="string"){a(x.bind(this,Error("Illegal arguments: "+typeof f+", "+typeof d)));return}if(d.length!==60){a(x.bind(this,null,!1));return}e.hash(f,d.substr(0,29),function(y,L){y?x(y):x(null,o(L,d))},s)}if(l){if(typeof l!="function")throw Error("Illegal callback: "+typeof l);c(l)}else return new Promise(function(x,y){c(function(L,E){if(L){y(L);return}x(E)})})},e.getRounds=function(f){if(typeof f!="string")throw Error("Illegal arguments: "+typeof f);return parseInt(f.split("$")[2],10)},e.getSalt=function(f){if(typeof f!="string")throw Error("Illegal arguments: "+typeof f);if(f.length!==60)throw Error("Illegal hash length: "+f.length+" != 60");return f.substring(0,29)};var a=typeof process<"u"&&process&&typeof process.nextTick=="function"?typeof setImmediate=="function"?setImmediate:process.nextTick:setTimeout;function i(f){var d=[],l=0;return g.encodeUTF16toUTF8(function(){return l>=f.length?null:f.charCodeAt(l++)},function(s){d.push(s)}),d}var p="./ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789".split(""),m=[-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,0,1,54,55,56,57,58,59,60,61,62,63,-1,-1,-1,-1,-1,-1,-1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,-1,-1,-1,-1,-1,-1,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,-1,-1,-1,-1,-1],b=String.fromCharCode;function u(f,d){var l=0,s=[],c,x;if(d<=0||d>f.length)throw Error("Illegal len: "+d);for(;l<d;){if(c=f[l++]&255,s.push(p[c>>2&63]),c=(c&3)<<4,l>=d){s.push(p[c&63]);break}if(x=f[l++]&255,c|=x>>4&15,s.push(p[c&63]),c=(x&15)<<2,l>=d){s.push(p[c&63]);break}x=f[l++]&255,c|=x>>6&3,s.push(p[c&63]),s.push(p[x&63])}return s.join("")}function v(f,d){var l=0,s=f.length,c=0,x=[],y,L,E,T,R,K;if(d<=0)throw Error("Illegal len: "+d);for(;l<s-1&&c<d&&(K=f.charCodeAt(l++),y=K<m.length?m[K]:-1,K=f.charCodeAt(l++),L=K<m.length?m[K]:-1,!(y==-1||L==-1||(R=y<<2>>>0,R|=(L&48)>>4,x.push(b(R)),++c>=d||l>=s)||(K=f.charCodeAt(l++),E=K<m.length?m[K]:-1,E==-1)||(R=(L&15)<<4>>>0,R|=(E&60)>>2,x.push(b(R)),++c>=d||l>=s)));)K=f.charCodeAt(l++),T=K<m.length?m[K]:-1,R=(E&3)<<6>>>0,R|=T,x.push(b(R)),++c;var X=[];for(l=0;l<c;l++)X.push(x[l].charCodeAt(0));return X}var g=function(){"use strict";var f={};return f.MAX_CODEPOINT=1114111,f.encodeUTF8=function(d,l){var s=null;for(typeof d=="number"&&(s=d,d=function(){return null});s!==null||(s=d())!==null;)s<128?l(s&127):s<2048?(l(s>>6&31|192),l(s&63|128)):s<65536?(l(s>>12&15|224),l(s>>6&63|128),l(s&63|128)):(l(s>>18&7|240),l(s>>12&63|128),l(s>>6&63|128),l(s&63|128)),s=null},f.decodeUTF8=function(d,l){for(var s,c,x,y,L=function(E){E=E.slice(0,E.indexOf(null));var T=Error(E.toString());throw T.name="TruncatedError",T.bytes=E,T};(s=d())!==null;)if(!(s&128))l(s);else if((s&224)===192)(c=d())===null&&L([s,c]),l((s&31)<<6|c&63);else if((s&240)===224)((c=d())===null||(x=d())===null)&&L([s,c,x]),l((s&15)<<12|(c&63)<<6|x&63);else if((s&248)===240)((c=d())===null||(x=d())===null||(y=d())===null)&&L([s,c,x,y]),l((s&7)<<18|(c&63)<<12|(x&63)<<6|y&63);else throw RangeError("Illegal starting byte: "+s)},f.UTF16toUTF8=function(d,l){for(var s,c=null;(s=c!==null?c:d())!==null;){if(s>=55296&&s<=57343&&(c=d())!==null&&c>=56320&&c<=57343){l((s-55296)*1024+c-56320+65536),c=null;continue}l(s)}c!==null&&l(c)},f.UTF8toUTF16=function(d,l){var s=null;for(typeof d=="number"&&(s=d,d=function(){return null});s!==null||(s=d())!==null;)s<=65535?l(s):(s-=65536,l((s>>10)+55296),l(s%1024+56320)),s=null},f.encodeUTF16toUTF8=function(d,l){f.UTF16toUTF8(d,function(s){f.encodeUTF8(s,l)})},f.decodeUTF8toUTF16=function(d,l){f.decodeUTF8(d,function(s){f.UTF8toUTF16(s,l)})},f.calculateCodePoint=function(d){return d<128?1:d<2048?2:d<65536?3:4},f.calculateUTF8=function(d){for(var l,s=0;(l=d())!==null;)s+=f.calculateCodePoint(l);return s},f.calculateUTF16asUTF8=function(d){var l=0,s=0;return f.UTF16toUTF8(d,function(c){++l,s+=f.calculateCodePoint(c)}),[l,s]},f}();Date.now=Date.now||function(){return+new Date};var S=16,h=10,w=16,$=100,P=[608135816,2242054355,320440878,57701188,2752067618,698298832,137296536,3964562569,1160258022,953160567,3193202383,887688300,3232508343,3380367581,1065670069,3041331479,2450970073,2306472731],A=[3509652390,2564797868,805139163,3491422135,3101798381,1780907670,3128725573,4046225305,614570311,3012652279,134345442,2240740374,1667834072,1901547113,2757295779,4103290238,227898511,1921955416,1904987480,2182433518,2069144605,3260701109,2620446009,720527379,3318853667,677414384,3393288472,3101374703,2390351024,1614419982,1822297739,2954791486,3608508353,3174124327,2024746970,1432378464,3864339955,2857741204,1464375394,1676153920,1439316330,715854006,3033291828,289532110,2706671279,2087905683,3018724369,1668267050,732546397,1947742710,3462151702,2609353502,2950085171,1814351708,2050118529,680887927,999245976,1800124847,3300911131,1713906067,1641548236,4213287313,1216130144,1575780402,4018429277,3917837745,3693486850,3949271944,596196993,3549867205,258830323,2213823033,772490370,2760122372,1774776394,2652871518,566650946,4142492826,1728879713,2882767088,1783734482,3629395816,2517608232,2874225571,1861159788,326777828,3124490320,2130389656,2716951837,967770486,1724537150,2185432712,2364442137,1164943284,2105845187,998989502,3765401048,2244026483,1075463327,1455516326,1322494562,910128902,469688178,1117454909,936433444,3490320968,3675253459,1240580251,122909385,2157517691,634681816,4142456567,3825094682,3061402683,2540495037,79693498,3249098678,1084186820,1583128258,426386531,1761308591,1047286709,322548459,995290223,1845252383,2603652396,3431023940,2942221577,3202600964,3727903485,1712269319,422464435,3234572375,1170764815,3523960633,3117677531,1434042557,442511882,3600875718,1076654713,1738483198,4213154764,2393238008,3677496056,1014306527,4251020053,793779912,2902807211,842905082,4246964064,1395751752,1040244610,2656851899,3396308128,445077038,3742853595,3577915638,679411651,2892444358,2354009459,1767581616,3150600392,3791627101,3102740896,284835224,4246832056,1258075500,768725851,2589189241,3069724005,3532540348,1274779536,3789419226,2764799539,1660621633,3471099624,4011903706,913787905,3497959166,737222580,2514213453,2928710040,3937242737,1804850592,3499020752,2949064160,2386320175,2390070455,2415321851,4061277028,2290661394,2416832540,1336762016,1754252060,3520065937,3014181293,791618072,3188594551,3933548030,2332172193,3852520463,3043980520,413987798,3465142937,3030929376,4245938359,2093235073,3534596313,375366246,2157278981,2479649556,555357303,3870105701,2008414854,3344188149,4221384143,3956125452,2067696032,3594591187,2921233993,2428461,544322398,577241275,1471733935,610547355,4027169054,1432588573,1507829418,2025931657,3646575487,545086370,48609733,2200306550,1653985193,298326376,1316178497,3007786442,2064951626,458293330,2589141269,3591329599,3164325604,727753846,2179363840,146436021,1461446943,4069977195,705550613,3059967265,3887724982,4281599278,3313849956,1404054877,2845806497,146425753,1854211946,1266315497,3048417604,3681880366,3289982499,290971e4,1235738493,2632868024,2414719590,3970600049,1771706367,1449415276,3266420449,422970021,1963543593,2690192192,3826793022,1062508698,1531092325,1804592342,2583117782,2714934279,4024971509,1294809318,4028980673,1289560198,2221992742,1669523910,35572830,157838143,1052438473,1016535060,1802137761,1753167236,1386275462,3080475397,2857371447,1040679964,2145300060,2390574316,1461121720,2956646967,4031777805,4028374788,33600511,2920084762,1018524850,629373528,3691585981,3515945977,2091462646,2486323059,586499841,988145025,935516892,3367335476,2599673255,2839830854,265290510,3972581182,2759138881,3795373465,1005194799,847297441,406762289,1314163512,1332590856,1866599683,4127851711,750260880,613907577,1450815602,3165620655,3734664991,3650291728,3012275730,3704569646,1427272223,778793252,1343938022,2676280711,2052605720,1946737175,3164576444,3914038668,3967478842,3682934266,1661551462,3294938066,4011595847,840292616,3712170807,616741398,312560963,711312465,1351876610,322626781,1910503582,271666773,2175563734,1594956187,70604529,3617834859,1007753275,1495573769,4069517037,2549218298,2663038764,504708206,2263041392,3941167025,2249088522,1514023603,1998579484,1312622330,694541497,2582060303,2151582166,1382467621,776784248,2618340202,3323268794,2497899128,2784771155,503983604,4076293799,907881277,423175695,432175456,1378068232,4145222326,3954048622,3938656102,3820766613,2793130115,2977904593,26017576,3274890735,3194772133,1700274565,1756076034,4006520079,3677328699,720338349,1533947780,354530856,688349552,3973924725,1637815568,332179504,3949051286,53804574,2852348879,3044236432,1282449977,3583942155,3416972820,4006381244,1617046695,2628476075,3002303598,1686838959,431878346,2686675385,1700445008,1080580658,1009431731,832498133,3223435511,2605976345,2271191193,2516031870,1648197032,4164389018,2548247927,300782431,375919233,238389289,3353747414,2531188641,2019080857,1475708069,455242339,2609103871,448939670,3451063019,1395535956,2413381860,1841049896,1491858159,885456874,4264095073,4001119347,1565136089,3898914787,1108368660,540939232,1173283510,2745871338,3681308437,4207628240,3343053890,4016749493,1699691293,1103962373,3625875870,2256883143,3830138730,1031889488,3479347698,1535977030,4236805024,3251091107,2132092099,1774941330,1199868427,1452454533,157007616,2904115357,342012276,595725824,1480756522,206960106,497939518,591360097,863170706,2375253569,3596610801,1814182875,2094937945,3421402208,1082520231,3463918190,2785509508,435703966,3908032597,1641649973,2842273706,3305899714,1510255612,2148256476,2655287854,3276092548,4258621189,236887753,3681803219,274041037,1734335097,3815195456,3317970021,1899903192,1026095262,4050517792,356393447,2410691914,3873677099,3682840055,3913112168,2491498743,4132185628,2489919796,1091903735,1979897079,3170134830,3567386728,3557303409,857797738,1136121015,1342202287,507115054,2535736646,337727348,3213592640,1301675037,2528481711,1895095763,1721773893,3216771564,62756741,2142006736,835421444,2531993523,1442658625,3659876326,2882144922,676362277,1392781812,170690266,3921047035,1759253602,3611846912,1745797284,664899054,1329594018,3901205900,3045908486,2062866102,2865634940,3543621612,3464012697,1080764994,553557557,3656615353,3996768171,991055499,499776247,1265440854,648242737,3940784050,980351604,3713745714,1749149687,3396870395,4211799374,3640570775,1161844396,3125318951,1431517754,545492359,4268468663,3499529547,1437099964,2702547544,3433638243,2581715763,2787789398,1060185593,1593081372,2418618748,4260947970,69676912,2159744348,86519011,2512459080,3838209314,1220612927,3339683548,133810670,1090789135,1078426020,1569222167,845107691,3583754449,4072456591,1091646820,628848692,1613405280,3757631651,526609435,236106946,48312990,2942717905,3402727701,1797494240,859738849,992217954,4005476642,2243076622,3870952857,3732016268,765654824,3490871365,2511836413,1685915746,3888969200,1414112111,2273134842,3281911079,4080962846,172450625,2569994100,980381355,4109958455,2819808352,2716589560,2568741196,3681446669,3329971472,1835478071,660984891,3704678404,4045999559,3422617507,3040415634,1762651403,1719377915,3470491036,2693910283,3642056355,3138596744,1364962596,2073328063,1983633131,926494387,3423689081,2150032023,4096667949,1749200295,3328846651,309677260,2016342300,1779581495,3079819751,111262694,1274766160,443224088,298511866,1025883608,3806446537,1145181785,168956806,3641502830,3584813610,1689216846,3666258015,3200248200,1692713982,2646376535,4042768518,1618508792,1610833997,3523052358,4130873264,2001055236,3610705100,2202168115,4028541809,2961195399,1006657119,2006996926,3186142756,1430667929,3210227297,1314452623,4074634658,4101304120,2273951170,1399257539,3367210612,3027628629,1190975929,2062231137,2333990788,2221543033,2438960610,1181637006,548689776,2362791313,3372408396,3104550113,3145860560,296247880,1970579870,3078560182,3769228297,1714227617,3291629107,3898220290,166772364,1251581989,493813264,448347421,195405023,2709975567,677966185,3703036547,1463355134,2715995803,1338867538,1343315457,2802222074,2684532164,233230375,2599980071,2000651841,3277868038,1638401717,4028070440,3237316320,6314154,819756386,300326615,590932579,1405279636,3267499572,3150704214,2428286686,3959192993,3461946742,1862657033,1266418056,963775037,2089974820,2263052895,1917689273,448879540,3550394620,3981727096,150775221,3627908307,1303187396,508620638,2975983352,2726630617,1817252668,1876281319,1457606340,908771278,3720792119,3617206836,2455994898,1729034894,1080033504,976866871,3556439503,2881648439,1522871579,1555064734,1336096578,3548522304,2579274686,3574697629,3205460757,3593280638,3338716283,3079412587,564236357,2993598910,1781952180,1464380207,3163844217,3332601554,1699332808,1393555694,1183702653,3581086237,1288719814,691649499,2847557200,2895455976,3193889540,2717570544,1781354906,1676643554,2592534050,3230253752,1126444790,2770207658,2633158820,2210423226,2615765581,2414155088,3127139286,673620729,2805611233,1269405062,4015350505,3341807571,4149409754,1057255273,2012875353,2162469141,2276492801,2601117357,993977747,3918593370,2654263191,753973209,36408145,2530585658,25011837,3520020182,2088578344,530523599,2918365339,1524020338,1518925132,3760827505,3759777254,1202760957,3985898139,3906192525,674977740,4174734889,2031300136,2019492241,3983892565,4153806404,3822280332,352677332,2297720250,60907813,90501309,3286998549,1016092578,2535922412,2839152426,457141659,509813237,4120667899,652014361,1966332200,2975202805,55981186,2327461051,676427537,3255491064,2882294119,3433927263,1307055953,942726286,933058658,2468411793,3933900994,4215176142,1361170020,2001714738,2830558078,3274259782,1222529897,1679025792,2729314320,3714953764,1770335741,151462246,3013232138,1682292957,1483529935,471910574,1539241949,458788160,3436315007,1807016891,3718408830,978976581,1043663428,3165965781,1927990952,4200891579,2372276910,3208408903,3533431907,1412390302,2931980059,4132332400,1947078029,3881505623,4168226417,2941484381,1077988104,1320477388,886195818,18198404,3786409e3,2509781533,112762804,3463356488,1866414978,891333506,18488651,661792760,1628790961,3885187036,3141171499,876946877,2693282273,1372485963,791857591,2686433993,3759982718,3167212022,3472953795,2716379847,445679433,3561995674,3504004811,3574258232,54117162,3331405415,2381918588,3769707343,4154350007,1140177722,4074052095,668550556,3214352940,367459370,261225585,2610173221,4209349473,3468074219,3265815641,314222801,3066103646,3808782860,282218597,3406013506,3773591054,379116347,1285071038,846784868,2669647154,3771962079,3550491691,2305946142,453669953,1268987020,3317592352,3279303384,3744833421,2610507566,3859509063,266596637,3847019092,517658769,3462560207,3443424879,370717030,4247526661,2224018117,4143653529,4112773975,2788324899,2477274417,1456262402,2901442914,1517677493,1846949527,2295493580,3734397586,2176403920,1280348187,1908823572,3871786941,846861322,1172426758,3287448474,3383383037,1655181056,3139813346,901632758,1897031941,2986607138,3066810236,3447102507,1393639104,373351379,950779232,625454576,3124240540,4148612726,2007998917,544563296,2244738638,2330496472,2058025392,1291430526,424198748,50039436,29584100,3605783033,2429876329,2791104160,1057563949,3255363231,3075367218,3463963227,1469046755,985887462],q=[1332899944,1700884034,1701343084,1684370003,1668446532,1869963892];function B(f,d,l,s){var c,x=f[d],y=f[d+1];return x^=l[0],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[1],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[2],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[3],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[4],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[5],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[6],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[7],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[8],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[9],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[10],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[11],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[12],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[13],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[14],c=s[x>>>24],c+=s[256|x>>16&255],c^=s[512|x>>8&255],c+=s[768|x&255],y^=c^l[15],c=s[y>>>24],c+=s[256|y>>16&255],c^=s[512|y>>8&255],c+=s[768|y&255],x^=c^l[16],f[d]=y^l[w+1],f[d+1]=x,f}function O(f,d){for(var l=0,s=0;l<4;++l)s=s<<8|f[d]&255,d=(d+1)%f.length;return{key:s,offp:d}}function z(f,d,l){for(var s=0,c=[0,0],x=d.length,y=l.length,L,E=0;E<x;E++)L=O(f,s),s=L.offp,d[E]=d[E]^L.key;for(E=0;E<x;E+=2)c=B(c,0,d,l),d[E]=c[0],d[E+1]=c[1];for(E=0;E<y;E+=2)c=B(c,0,d,l),l[E]=c[0],l[E+1]=c[1]}function Z(f,d,l,s){for(var c=0,x=[0,0],y=l.length,L=s.length,E,T=0;T<y;T++)E=O(d,c),c=E.offp,l[T]=l[T]^E.key;for(c=0,T=0;T<y;T+=2)E=O(f,c),c=E.offp,x[0]^=E.key,E=O(f,c),c=E.offp,x[1]^=E.key,x=B(x,0,l,s),l[T]=x[0],l[T+1]=x[1];for(T=0;T<L;T+=2)E=O(f,c),c=E.offp,x[0]^=E.key,E=O(f,c),c=E.offp,x[1]^=E.key,x=B(x,0,l,s),s[T]=x[0],s[T+1]=x[1]}function ht(f,d,l,s,c){var x=q.slice(),y=x.length,L;if(l<4||l>31)if(L=Error("Illegal number of rounds (4-31): "+l),s){a(s.bind(this,L));return}else throw L;if(d.length!==S)if(L=Error("Illegal salt length: "+d.length+" != "+S),s){a(s.bind(this,L));return}else throw L;l=1<<l>>>0;var E,T,R=0,K;Int32Array?(E=new Int32Array(P),T=new Int32Array(A)):(E=P.slice(),T=A.slice()),Z(d,f,E,T);function X(){if(c&&c(R/l),R<l)for(var ue=Date.now();R<l&&(R=R+1,z(f,E,T),z(d,E,T),!(Date.now()-ue>$)););else{for(R=0;R<64;R++)for(K=0;K<y>>1;K++)B(x,K<<1,E,T);var V=[];for(R=0;R<y;R++)V.push((x[R]>>24&255)>>>0),V.push((x[R]>>16&255)>>>0),V.push((x[R]>>8&255)>>>0),V.push((x[R]&255)>>>0);if(s){s(null,V);return}else return V}s&&a(X)}if(typeof s<"u")X();else for(var ge;;)if(typeof(ge=X())<"u")return ge||[]}function Ue(f,d,l,s){var c;if(typeof f!="string"||typeof d!="string")if(c=Error("Invalid string / salt: Not a string"),l){a(l.bind(this,c));return}else throw c;var x,y;if(d.charAt(0)!=="$"||d.charAt(1)!=="2")if(c=Error("Invalid salt version: "+d.substring(0,2)),l){a(l.bind(this,c));return}else throw c;if(d.charAt(2)==="$")x="\0",y=3;else{if(x=d.charAt(2),x!=="a"&&x!=="b"&&x!=="y"||d.charAt(3)!=="$")if(c=Error("Invalid salt revision: "+d.substring(2,4)),l){a(l.bind(this,c));return}else throw c;y=4}if(d.charAt(y+2)>"$")if(c=Error("Missing salt rounds"),l){a(l.bind(this,c));return}else throw c;var L=parseInt(d.substring(y,y+1),10)*10,E=parseInt(d.substring(y+1,y+2),10),T=L+E,R=d.substring(y+3,y+25);f+=x>="a"?"\0":"";var K=i(f),X=v(R,S);function ge(ue){var V=[];return V.push("$2"),x>="a"&&V.push(x),V.push("$"),T<10&&V.push("0"),V.push(T.toString()),V.push("$"),V.push(u(X,X.length)),V.push(u(ue,q.length*4-1)),V.join("")}if(typeof l>"u")return ge(ht(K,X,T));ht(K,X,T,function(ue,V){ue?l(ue,null):l(null,ge(V))},s)}return e.encodeBase64=u,e.decodeBase64=v,e})});var tt=Q(Le=>{"use strict";Object.defineProperty(Le,"__esModule",{value:!0});Le.crossEnvCrypto=void 0;try{let e=typeof window<"u"&&window.crypto||ve("crypto").webcrypto;if(e){if(!e.subtle)throw new Error("Crypto.Subtle is undefined. Make sure you are using HTTPS in a compatible browser.");let t=n=>r=>e.subtle.digest(n,r);Le.crossEnvCrypto={randomBytes:e.getRandomValues.bind(e),hashFunctions:{SHA1:t("SHA-1"),SHA256:t("SHA-256"),SHA384:t("SHA-384"),SHA512:t("SHA-512")}}}else{let t=ve("crypto"),n=r=>o=>t.createHash(r).update(o).digest().buffer;Le.crossEnvCrypto={randomBytes:t.randomFillSync,hashFunctions:{SHA1:n("sha1"),SHA256:n("sha256"),SHA384:n("sha384"),SHA512:n("sha512")}}}}catch(e){throw console.error(e),new Error("No suitable crypto library was found. You may need a polyfill.")}});var sn=Q(Ne=>{"use strict";Object.defineProperty(Ne,"__esModule",{value:!0});Ne.SRPParameters=void 0;var Oe=tt(),Ae=class e{constructor(t=e.PrimeGroup[2048],n=e.H.SHA512){if(this.primeGroup=t,this.H=n,this.NBits=this.primeGroup.N.toString(2).length,!n)throw new Error("Hash function required")}};Ne.SRPParameters=Ae;Ae.PrimeGroup={256:{N:BigInt("125617018995153554710546479714086468244499594888726646874671447258204721048803"),g:BigInt(2)},512:{N:BigInt("11144252439149533417835749556168991736939157778924947037200268358613863350040339017097790259154750906072491181606044774215413467851989724116331597513345603"),g:BigInt(2)},768:{N:BigInt("1087179135105457859072065649059069760280540086975817629066444682366896187793570736574549981488868217843627094867924800342887096064844227836735667168319981288765377499806385489913341488724152562880918438701129530606139552645689583147"),g:BigInt(2)},1024:{N:BigInt("167609434410335061345139523764350090260135525329813904557420930309800865859473551531551523800013916573891864789934747039010546328480848979516637673776605610374669426214776197828492691384519453218253702788022233205683635831626913357154941914129985489522629902540768368409482248290641036967659389658897350067939"),g:BigInt(2)},1536:{N:BigInt("1486998185923128292816507353619409521152457662596380074614818966810244974827752411420380336514078832314731499938313197533147998565301020797040787428051479639316928015998415709101293902971072960487527411068082311763171549170528008620813391411445907584912865222076100726050255271567749213905330659264908657221124284665444825474741087704974475795505492821585749417639344967192301749033325359286273431675492866492416941152646940908101472416714421046022696100064262587"),g:BigInt(2)},2048:{N:BigInt("21766174458617435773191008891802753781907668374255538511144643224689886235383840957210909013086056401571399717235807266581649606472148410291413364152197364477180887395655483738115072677402235101762521901569820740293149529620419333266262073471054548368736039519702486226506248861060256971802984953561121442680157668000761429988222457090413873973970171927093992114751765168063614761119615476233422096442783117971236371647333871414335895773474667308967050807005509320424799678417036867928316761272274230314067548291133582479583061439577559347101961771406173684378522703483495337037655006751328447510550299250924469288819"),g:BigInt(2)}};Ae.H={SHA1:Oe.crossEnvCrypto.hashFunctions.SHA1,SHA256:Oe.crossEnvCrypto.hashFunctions.SHA256,SHA384:Oe.crossEnvCrypto.hashFunctions.SHA384,SHA512:Oe.crossEnvCrypto.hashFunctions.SHA512}});var $e=Q(I=>{"use strict";var rt=I&&I.__awaiter||function(e,t,n,r){function o(a){return a instanceof n?a:new n(function(i){i(a)})}return new(n||(n=Promise))(function(a,i){function p(u){try{b(r.next(u))}catch(v){i(v)}}function m(u){try{b(r.throw(u))}catch(v){i(v)}}function b(u){u.done?a(u.value):o(u.value).then(p,m)}b((r=r.apply(e,t||[])).next())})};Object.defineProperty(I,"__esModule",{value:!0});I.modPow=I.hashBitCount=I.createVerifierAndSalt=I.createVerifier=I.generateRandomBigInt=I.generateRandomString=I.hashPadded=I.hash=I.padStartArrayBuffer=I.stringToArrayBuffer=I.arrayBufferToBigInt=I.bigIntToArrayBuffer=void 0;var dn=tt(),jn=BigInt(0),nt=BigInt(1),ln=BigInt(2),zn=e=>{let t=e.toString(16),n=new ArrayBuffer(Math.ceil(t.length/2)),r=new Uint8Array(n),o=0;t.length%2!==0&&(r[0]=parseInt(t[0],16),o=1);for(let a=0;a<n.byteLength;a++)r[a+o]=parseInt(t.slice(2*a+o,2*a+2+o),16);return n};I.bigIntToArrayBuffer=zn;var Yn=e=>{let t=[];return new Uint8Array(e).forEach(n=>{t.push(("0"+n.toString(16)).slice(-2))}),BigInt(`0x${t.join("")}`)};I.arrayBufferToBigInt=Yn;function Zn(e){return new TextEncoder().encode(e).buffer}I.stringToArrayBuffer=Zn;var Xn=(e,t)=>{let n=new Uint8Array(e);if(n.length<t){let r=new Uint8Array(t);return r.fill(0,0,t-n.length),r.set(n,t-n.length),r}return n};I.padStartArrayBuffer=Xn;function at(e,...t){let n=t.reduce((o,a)=>o+a.byteLength,0),r=new Uint8Array(n);for(let o=0,a=0;a<t.length;a++)r.set(new Uint8Array(t[a]),o),o+=t[a].byteLength;return e.H(r)}I.hash=at;function Jn(e,t,...n){let r=n.map(o=>I.padStartArrayBuffer(o,t));return at(e,...r)}I.hashPadded=Jn;function Wn(e=10){let t=new Uint8Array(Math.ceil(Math.ceil(e/2)));return dn.crossEnvCrypto.randomBytes(t),t.reduce((n,r)=>{let o=r.toString(16).toString();return o.length===1?n+"0"+o:n+o},"").slice(0,e)}I.generateRandomString=Wn;function Qn(e=16){return I.arrayBufferToBigInt(rr(e))}I.generateRandomBigInt=Qn;function un(e,t,n,r){return rt(this,void 0,void 0,function*(){if(!t||!t.trim())throw new Error("Identity (I) must not be null or empty.");if(!n)throw new Error("Salt (s) must not be null.");if(!r)throw new Error("Password (P) must not be null");let o=yield e.computeX(t,n,r);return e.computeVerifier(o)})}I.createVerifier=un;function er(e,t,n,r){return rt(this,void 0,void 0,function*(){let o=yield e.generateRandomSalt(r);return{s:o,v:yield un(e,t,o,n)}})}I.createVerifierAndSalt=er;var tr=e=>rt(void 0,void 0,void 0,function*(){return(yield at(e,I.bigIntToArrayBuffer(BigInt(1)))).byteLength*8});I.hashBitCount=tr;function nr(e,t,n){let r=nt;for(;t>jn;)t%ln==nt?(r=e*r%n,t-=nt):(e=e*e%n,t/=ln);return r}I.modPow=nr;var rr=e=>{let t=new Uint8Array(e);return dn.crossEnvCrypto.randomBytes(t),t.buffer}});var fn=Q(he=>{"use strict";var te=he&&he.__awaiter||function(e,t,n,r){function o(a){return a instanceof n?a:new n(function(i){i(a)})}return new(n||(n=Promise))(function(a,i){function p(u){try{b(r.next(u))}catch(v){i(v)}}function m(u){try{b(r.throw(u))}catch(v){i(v)}}function b(u){u.done?a(u.value):o(u.value).then(p,m)}b((r=r.apply(e,t||[])).next())})};Object.defineProperty(he,"__esModule",{value:!0});he.SRPRoutines=void 0;var M=$e(),ot=class{constructor(t){this.parameters=t}hash(...t){return M.hash(this.parameters,...t)}hashPadded(...t){let n=Math.trunc((this.parameters.NBits+7)/8);return M.hashPadded(this.parameters,n,...t)}computeK(){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hashPadded(M.bigIntToArrayBuffer(this.parameters.primeGroup.N),M.bigIntToArrayBuffer(this.parameters.primeGroup.g)))})}generateRandomSalt(t){return te(this,void 0,void 0,function*(){let n=yield M.hashBitCount(this.parameters),r=t||2*n/8;return M.generateRandomBigInt(r)})}computeX(t,n,r){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hash(M.bigIntToArrayBuffer(n),yield this.computeIdentityHash(t,r)))})}computeXStep2(t,n){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hash(M.bigIntToArrayBuffer(t),n))})}computeIdentityHash(t,n){return te(this,void 0,void 0,function*(){return yield this.hash(M.stringToArrayBuffer(n))})}computeVerifier(t){return M.modPow(this.parameters.primeGroup.g,t,this.parameters.primeGroup.N)}generatePrivateValue(){let t=Math.max(256,this.parameters.NBits),n;do n=M.generateRandomBigInt(t/8)%this.parameters.primeGroup.N;while(n===BigInt(0));return n}computeClientPublicValue(t){return M.modPow(this.parameters.primeGroup.g,t,this.parameters.primeGroup.N)}isValidPublicValue(t){return t%this.parameters.primeGroup.N!==BigInt(0)}computeU(t,n){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hashPadded(M.bigIntToArrayBuffer(t),M.bigIntToArrayBuffer(n)))})}computeClientEvidence(t,n,r,o,a){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hash(M.bigIntToArrayBuffer(r),M.bigIntToArrayBuffer(o),M.bigIntToArrayBuffer(a)))})}computeServerEvidence(t,n,r){return te(this,void 0,void 0,function*(){return M.arrayBufferToBigInt(yield this.hash(M.bigIntToArrayBuffer(t),M.bigIntToArrayBuffer(n),M.bigIntToArrayBuffer(r)))})}computeClientSessionKey(t,n,r,o,a){let i=this.parameters.primeGroup.N,p=r*n+o,m=M.modPow(this.parameters.primeGroup.g,n,i)*t%i;return M.modPow(a+i-m,p,i)}};he.SRPRoutines=ot});var xn=Q(W=>{"use strict";var st=W&&W.__awaiter||function(e,t,n,r){function o(a){return a instanceof n?a:new n(function(i){i(a)})}return new(n||(n=Promise))(function(a,i){function p(u){try{b(r.next(u))}catch(v){i(v)}}function m(u){try{b(r.throw(u))}catch(v){i(v)}}function b(u){u.done?a(u.value):o(u.value).then(p,m)}b((r=r.apply(e,t||[])).next())})};Object.defineProperty(W,"__esModule",{value:!0});W.SRPClientSessionStep2=W.SRPClientSessionStep1=W.SRPClientSession=void 0;var ct=class{constructor(t){this.routines=t}step1(t,n){return st(this,void 0,void 0,function*(){if(!t||!t.trim())throw new Error("User identity must not be null nor empty");if(!n)throw new Error("User password must not be null");let r=yield this.routines.computeIdentityHash(t,n);return new ke(this.routines,t,r)})}};W.SRPClientSession=ct;var ke=class e{constructor(t,n,r){this.routines=t,this.I=n,this.IH=r}step2(t,n){return st(this,void 0,void 0,function*(){if(!t)throw new Error("Salt (s) must not be null");if(!n)throw new Error("Public server value (B) must not be null");let r=yield this.routines.computeXStep2(t,this.IH),o=this.routines.generatePrivateValue(),a=this.routines.computeClientPublicValue(o),i=yield this.routines.computeK(),p=yield this.routines.computeU(a,n),m=this.routines.computeClientSessionKey(i,r,p,o,n),b=yield this.routines.computeClientEvidence(this.I,t,a,n,m);return new _e(this.routines,a,b,m)})}toJSON(){return{I:this.I,IH:Array.from(new Uint8Array(this.IH))}}static fromState(t,n){return new e(t,n.I,new Uint8Array(n.IH).buffer)}};W.SRPClientSessionStep1=ke;var _e=class e{constructor(t,n,r,o){this.routines=t,this.A=n,this.M1=r,this.S=o}step3(t){return st(this,void 0,void 0,function*(){if(!t)throw new Error("Server evidence (M2) must not be null");if((yield this.routines.computeServerEvidence(this.A,this.M1,this.S))!==t)throw new Error("Bad server credentials")})}toJSON(){return{A:this.A.toString(16),M1:this.M1.toString(16),S:this.S.toString(16)}}static fromState(t,n){return new e(t,BigInt("0x"+n.A),BigInt("0x"+n.M1),BigInt("0x"+n.S))}};W.SRPClientSessionStep2=_e});var pn=Q(ne=>{"use strict";var it=ne&&ne.__awaiter||function(e,t,n,r){function o(a){return a instanceof n?a:new n(function(i){i(a)})}return new(n||(n=Promise))(function(a,i){function p(u){try{b(r.next(u))}catch(v){i(v)}}function m(u){try{b(r.throw(u))}catch(v){i(v)}}function b(u){u.done?a(u.value):o(u.value).then(p,m)}b((r=r.apply(e,t||[])).next())})};Object.defineProperty(ne,"__esModule",{value:!0});ne.SRPServerSessionStep1=ne.SRPServerSession=void 0;var lt=$e(),dt=class{constructor(t){this.routines=t}step1(t,n,r){return it(this,void 0,void 0,function*(){let o=this.routines.generatePrivateValue(),a=yield this.routines.computeK(),i=ar(this.routines.parameters,a,r,o);return new Fe(this.routines,t,n,r,o,i)})}};ne.SRPServerSession=dt;var Fe=class e{constructor(t,n,r,o,a,i){this.routines=t,this.identifier=n,this.salt=r,this.verifier=o,this.b=a,this.B=i}sessionKey(t){return it(this,void 0,void 0,function*(){if(t===null)throw new Error("Client public value (A) must not be null");if(!this.routines.isValidPublicValue(t))throw new Error(`Invalid Client public value (A): ${t.toString(16)}`);let n=yield this.routines.computeU(t,this.B);return or(this.routines.parameters.primeGroup.N,this.verifier,n,t,this.b)})}step2(t,n){return it(this,void 0,void 0,function*(){if(!n)throw new Error("Client evidence (M1) must not be null");let r=yield this.sessionKey(t);if((yield this.routines.computeClientEvidence(this.identifier,this.salt,t,this.B,r))!==n)throw new Error("Bad client credentials");return this.routines.computeServerEvidence(t,n,r)})}toJSON(){return{identifier:this.identifier,salt:this.salt.toString(16),verifier:this.verifier.toString(16),b:this.b.toString(16),B:this.B.toString(16)}}static fromState(t,n){return new e(t,n.identifier,BigInt("0x"+n.salt),BigInt("0x"+n.verifier),BigInt("0x"+n.b),BigInt("0x"+n.B))}};ne.SRPServerSessionStep1=Fe;var ar=(e,t,n,r)=>(lt.modPow(e.primeGroup.g,r,e.primeGroup.N)+n*t)%e.primeGroup.N,or=(e,t,n,r,o)=>lt.modPow(lt.modPow(t,n,e)*r,o,e)});var mn=Q(_=>{"use strict";Object.defineProperty(_,"__esModule",{value:!0});_.generateRandomBigInt=_.arrayBufferToBigInt=_.bigIntToArrayBuffer=_.createVerifierAndSalt=_.SRPServerSessionStep1=_.SRPServerSession=_.SRPClientSessionStep2=_.SRPClientSessionStep1=_.SRPClientSession=_.SRPRoutines=_.SRPParameters=void 0;var cr=sn();Object.defineProperty(_,"SRPParameters",{enumerable:!0,get:function(){return cr.SRPParameters}});var sr=fn();Object.defineProperty(_,"SRPRoutines",{enumerable:!0,get:function(){return sr.SRPRoutines}});var ut=xn();Object.defineProperty(_,"SRPClientSession",{enumerable:!0,get:function(){return ut.SRPClientSession}});Object.defineProperty(_,"SRPClientSessionStep1",{enumerable:!0,get:function(){return ut.SRPClientSessionStep1}});Object.defineProperty(_,"SRPClientSessionStep2",{enumerable:!0,get:function(){return ut.SRPClientSessionStep2}});var bn=pn();Object.defineProperty(_,"SRPServerSession",{enumerable:!0,get:function(){return bn.SRPServerSession}});Object.defineProperty(_,"SRPServerSessionStep1",{enumerable:!0,get:function(){return bn.SRPServerSessionStep1}});var Ke=$e();Object.defineProperty(_,"createVerifierAndSalt",{enumerable:!0,get:function(){return Ke.createVerifierAndSalt}});Object.defineProperty(_,"bigIntToArrayBuffer",{enumerable:!0,get:function(){return Ke.bigIntToArrayBuffer}});Object.defineProperty(_,"arrayBufferToBigInt",{enumerable:!0,get:function(){return Ke.arrayBufferToBigInt}});Object.defineProperty(_,"generateRandomBigInt",{enumerable:!0,get:function(){return Ke.generateRandomBigInt}})});var gt=document.querySelector("#warning"),fe=[],An=(()=>{try{if(window.BigInt instanceof Function)return!0}catch{}return!1})();An||fe.push("BigInt is not supported in this browser. Many operations, including SRP, will fail.");var Tn=(()=>{try{if(window.indexedDB instanceof IDBFactory)return!0}catch{}return!1})();Tn||fe.push("IndexedDB is not supported in this browser. Cryptographic keys cannot be stored.");var Cn=(()=>{try{if(window.crypto.getRandomValues instanceof Function)return!0}catch{}return!1})();Cn||fe.push("This browser cannot securely source random data. Many operations will fail. The outputs of any operations that do not fail should not be trusted.");var In=(()=>{try{if(window.crypto.subtle.constructor instanceof Function)return!0}catch{}return!1})();In||fe.push("The WebCrypto API subtle crypto interface is not available. Is this page being delivered over HTTPS? Many operations will fail. The outputs of any operations that do not fail should not be trusted.");if(fe.length){gt.classList.add("active");let e=gt.querySelector(".content");fe.forEach(t=>{let n=document.createElement("p");n.textContent=t,e.appendChild(n)})}var Ce=document.querySelector("#loader"),Bn=e=>{Ce&&e>=0&&e<=100&&(Ce.style.width=`${e}%`),e===0?Ce?.classList.remove("animate"):Ce?.classList.add("animate")},N=Bn;var Ge=document.querySelector("#error"),vt=Ge.querySelector(".content"),C=e=>{let t;e instanceof Error?t=e:t=new Error(String(e)),Ge.classList.add("active"),vt.textContent=t.message,N(100)},je=()=>{vt.textContent="",Ge.classList.remove("active")};var Ie=De(ze()),J=(h=>(h[h.BOOLEAN=-1]="BOOLEAN",h[h.BIGINT=-2]="BIGINT",h[h.INTEGER=-3]="INTEGER",h[h.UUID=-4]="UUID",h[h.JSON=-5]="JSON",h[h.PEM=-6]="PEM",h[h.UNKNOWN=0]="UNKNOWN",h[h.BINARY=2]="BINARY",h[h.OCTAL=8]="OCTAL",h[h["UTF-8"]=100]="UTF-8",h[h.HEXADECIMAL=16]="HEXADECIMAL",h[h.BASE64=64]="BASE64",h[h.BASE64_CRYPT=640]="BASE64_CRYPT",h[h.PHC_STRING=-100]="PHC_STRING",h[h.MODULAR_CRYPT_FORMAT=-200]="MODULAR_CRYPT_FORMAT",h))(J||{}),j=(e,t)=>{let n=new Uint8Array(e);if(t===64)return btoa(n.reduce((o,a)=>o+String.fromCharCode(a),""));if(t===640)return Ie.encodeBase64(n,n.length);if(t===100)return new TextDecoder().decode(e);let r=0;if(t===2&&(r=8),t===8&&(r=3),t===16&&(r=2),!r)throw new Error(`Radix ${t} is not supported`);return n.reduce((o,a)=>o+a.toString(t).padStart(r,"0"),"")},U=(e,t)=>{let n;if(t===-1)return n=new Uint8Array(1),n[0]=Number(JSON.parse(e)),n.buffer;if(t===64){let a=atob(e);n=new Uint8Array(a.length);for(let i=0;i<a.length;i+=1)n[i]=a.charCodeAt(i);return n.buffer}if(t===640){let a=Ie.decodeBase64(e,1/0);return Uint8Array.from(a).buffer}if(t===-2){let a=BigInt(e),i=[];for(;a>0;)i.push(Number(a&0xffn)),a>>=8n;return Uint8Array.from(i.reverse()).buffer}if(t===-4){let a=e.replace(/-/g,"");return n=new Uint8Array(a.length/2),a.match(/.{1,2}/g)?.forEach((i,p)=>{n[p]=parseInt(i,16)}),n.buffer}if(t===100)return new TextEncoder().encode(e);let r=0;if(t===2&&(r=8),t===8&&(r=3),t===16&&(r=2),!r)throw new Error(`Radix ${t} is not supported`);let o=e.length/r;return n=new Uint8Array(o),e.match(new RegExp(`.{1,${r}}`,"g"))?.forEach((a,i)=>{n[i]=parseInt(a,t)}),n.buffer},oe=e=>{if(e.toLowerCase()==="true"||e.toLowerCase()==="false")return-1;if(/^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i.test(e))return-4;if(/^([01]{8})+$/.test(e))return 2;if(/^([0-7]{3})+$/.test(e))return 8;if(/^([0-9a-f]{2})+$|^([0-9A-F]{2})+$/.test(e))return 16;if(/^([0-9a-zA-Z+/]{4})*[0-9a-zA-Z+/]{2}[0-9a-zA-Z+/=]{2}$/.test(e))return 64;if(/^-{5}BEGIN .+-{5}(\r\n?|\n)([0-9a-zA-Z+/]{4})*[0-9a-zA-Z+/]{2}[0-9a-zA-Z+/=]{2}(\r\n?|\n)-{5}END .+-{5}(\r\n?|\n)?$/.test(e))return-6;if(/^[0-9a-zA-Z./]+$/.test(e))return 640;if(/{/.test(e))try{return JSON.parse(e),-5}catch{}return/^\$[a-z0-9-]{4,32}(\$v=[0-9]+)?(\$[a-z0-9-]{1,32}=[a-zA-Z0-9/+.-]+(,[a-z0-9-]{1,32}=[a-zA-Z0-9/+.-]+)*)?(\$[a-zA-Z0-9/+.-]+(\$[a-zA-Z0-9/+]{2,})?)?$/.test(e)?-100:/^\$[a-zA-Z0-9]+(\$[a-zA-Z0-9./]+)+$/.test(e)?-200:0};var St='<svg viewBox="0 0 24 24"><path fill="currentColor" d="M7.41 8.58L12 13.17l4.59-4.59L18 10l-6 6l-6-6z"></path></svg>',Ye='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M240 128c-64 0 0 88-64 88H80c-64 0 0-88-64-88c64 0 0-88 64-88h96c64 0 0 88 64 88" opacity=".2"/><path d="M43.18 128a29.78 29.78 0 0 1 8 10.26c4.8 9.9 4.8 22 4.8 33.74c0 24.31 1 36 24 36a8 8 0 0 1 0 16c-17.48 0-29.32-6.14-35.2-18.26c-4.8-9.9-4.8-22-4.8-33.74c0-24.31-1-36-24-36a8 8 0 0 1 0-16c23 0 24-11.69 24-36c0-11.72 0-23.84 4.8-33.74C50.68 38.14 62.52 32 80 32a8 8 0 0 1 0 16c-23 0-24 11.69-24 36c0 11.72 0 23.84-4.8 33.74A29.78 29.78 0 0 1 43.18 128M240 120c-23 0-24-11.69-24-36c0-11.72 0-23.84-4.8-33.74C205.32 38.14 193.48 32 176 32a8 8 0 0 0 0 16c23 0 24 11.69 24 36c0 11.72 0 23.84 4.8 33.74a29.78 29.78 0 0 0 8 10.26a29.78 29.78 0 0 0-8 10.26c-4.8 9.9-4.8 22-4.8 33.74c0 24.31-1 36-24 36a8 8 0 0 0 0 16c17.48 0 29.32-6.14 35.2-18.26c4.8-9.9 4.8-22 4.8-33.74c0-24.31 1-36 24-36a8 8 0 0 0 0-16"/></g></svg>',wt='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M184 72v144H40V72Z" opacity=".2"/><path d="M184 64H40a8 8 0 0 0-8 8v144a8 8 0 0 0 8 8h144a8 8 0 0 0 8-8V72a8 8 0 0 0-8-8m-8 144H48V80h128Zm48-168v144a8 8 0 0 1-16 0V48H72a8 8 0 0 1 0-16h144a8 8 0 0 1 8 8"/></g></svg>',Lt='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M216 91.31V208a8 8 0 0 1-8 8h-32v-64a8 8 0 0 0-8-8H88a8 8 0 0 0-8 8v64H48a8 8 0 0 1-8-8V48a8 8 0 0 1 8-8h116.69a8 8 0 0 1 5.65 2.34l43.32 43.31a8 8 0 0 1 2.34 5.66" opacity=".2"/><path d="M219.31 80L176 36.69A15.86 15.86 0 0 0 164.69 32H48a16 16 0 0 0-16 16v160a16 16 0 0 0 16 16h160a16 16 0 0 0 16-16V91.31A15.86 15.86 0 0 0 219.31 80M168 208H88v-56h80Zm40 0h-24v-56a16 16 0 0 0-16-16H88a16 16 0 0 0-16 16v56H48V48h116.69L208 91.31ZM160 72a8 8 0 0 1-8 8H96a8 8 0 0 1 0-16h56a8 8 0 0 1 8 8"/></g></svg>',At='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M232 56v144a16 16 0 0 1-16 16H40a16 16 0 0 1-16-16V56a16 16 0 0 1 16-16h176a16 16 0 0 1 16 16" opacity=".2"/><path d="m205.66 85.66l-96 96a8 8 0 0 1-11.32 0l-40-40a8 8 0 0 1 11.32-11.32L104 164.69l90.34-90.35a8 8 0 0 1 11.32 11.32"/></g></svg>',Tt='<svg viewBox="0 0 24 24"><path fill="currentColor" d="m17.16 7.59l-5.66 5.66l-5.66-5.66l.71-.7l4.95 4.95l4.95-4.95zm0 4l-5.66 5.66l-5.66-5.66l.71-.7l4.95 4.95l4.95-4.95z"/></svg>',Ct='<svg viewBox="0 0 16 16"><path fill="currentColor" fill-rule="evenodd" d="m10.57 1.14l3.28 3.3l.15.36v9.7l-.5.5h-11l-.5-.5v-13l.5-.5h7.72zM3 2v12h10V5l-3-3zm1.46 4.052c0 1.287.458 1.93 1.374 1.93c.457 0 .807-.173 1.05-.52c.246-.348.368-.847.368-1.499C7.252 4.654 6.805 4 5.91 4c-.471 0-.831.175-1.08.526c-.247.35-.37.858-.37 1.526m.862-.022c0-.922.183-1.383.55-1.383c.344 0 .516.448.516 1.343s-.176 1.343-.527 1.343c-.36 0-.54-.434-.54-1.303zm3.187 1.886h2.435v-.672h-.792V4l-1.665.336v.687l.82-.177v2.398h-.798zm-1.337 5H4.736v-.672h.798V9.846l-.82.177v-.687L6.38 9v3.244h.792v.671zm1.035-1.931c0 1.287.458 1.93 1.375 1.93c.457 0 .807-.173 1.05-.52c.245-.348.368-.847.368-1.499c0-1.309-.448-1.963-1.343-1.963c-.47 0-.83.175-1.08.526c-.246.35-.37.858-.37 1.526m.862-.022c0-.922.184-1.383.55-1.383c.344 0 .516.448.516 1.343s-.175 1.343-.526 1.343c-.36 0-.54-.434-.54-1.303" clip-rule="evenodd"/></svg>',It='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M232 136v64a8 8 0 0 1-8 8H32a8 8 0 0 1-8-8v-64a8 8 0 0 1 8-8h192a8 8 0 0 1 8 8" opacity=".2"/><path d="M240 136v64a16 16 0 0 1-16 16H32a16 16 0 0 1-16-16v-64a16 16 0 0 1 16-16h40a8 8 0 0 1 0 16H32v64h192v-64h-40a8 8 0 0 1 0-16h40a16 16 0 0 1 16 16m-117.66-2.34a8 8 0 0 0 11.32 0l48-48a8 8 0 0 0-11.32-11.32L136 108.69V24a8 8 0 0 0-16 0v84.69L85.66 74.34a8 8 0 0 0-11.32 11.32ZM200 168a12 12 0 1 0-12 12a12 12 0 0 0 12-12"/></g></svg>',Be='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M232 96a72 72 0 0 1-98.83 66.83L120 176H96v24H72v24H32v-40l61.17-61.17A72 72 0 1 1 232 96" opacity=".2"/><path d="M160 16a80.07 80.07 0 0 0-76.09 104.78l-57.57 57.56A8 8 0 0 0 24 184v40a8 8 0 0 0 8 8h40a8 8 0 0 0 8-8v-16h16a8 8 0 0 0 8-8v-16h16a8 8 0 0 0 5.66-2.34l9.56-9.57A80 80 0 1 0 160 16m0 144a63.7 63.7 0 0 1-23.65-4.51a8 8 0 0 0-8.84 1.68L116.69 168H96a8 8 0 0 0-8 8v16H72a8 8 0 0 0-8 8v16H40v-28.69l58.83-58.82a8 8 0 0 0 1.68-8.84A64 64 0 1 1 160 160m32-84a12 12 0 1 1-12-12a12 12 0 0 1 12 12"/></g></svg>',Bt='<svg viewBox="0 0 512 512"><path fill="currentColor" d="M261.1 24.8c-6.3 0-12.7.43-19.2 1.18c-34.6 4.01-64.8 17.59-86.1 37.06c-21.4 19.48-34.2 45.56-31 73.16c2.8 24.6 17.8 45.2 39.1 59.4c2.6-6.2 5.9-11.9 9.2-16.5c-17.6-11.6-28.4-27.3-30.4-45c-2.3-19.7 6.7-39.58 24.8-56.14c18.2-16.57 45.3-29.06 76.6-32.68c31.3-3.63 60.6 2.33 82.1 14.3c21.4 11.98 34.7 29.31 37 48.92c2.2 19.3-6.2 38.8-23.4 55a69.91 69.91 0 0 0-35.4-10.6h-2.2c-5.1.1-10.1.7-15.3 1.8c-37.5 8.7-60.8 45.5-52.2 82.7c5.3 23 21.6 40.6 42.2 48.5l39.7 172.2l47 29.1l29.5-46.7l-23.5-14.5l14.8-23.4l-23.5-14.6l14.7-23.3l-23.5-14.6l14.8-23.4l-13.5-58.4c15.1-16.1 22-39.1 16.7-62.2c-2.7-11.7-8.2-22-15.8-30.4c18.9-19 29.8-43.5 26.8-69.2c-3.2-27.55-21.6-50.04-46.9-64.11c-20.5-11.45-45.8-17.77-73.1-17.59m-20.2 135.5c-25.9 1.1-49.9 16.8-60.4 42.2c-9.1 21.9-6 45.7 6.2 64.2l-67.8 163l21.3 51l51.2-20.9l-10.7-25.5l25.6-10.4l-10.6-25.5l25.6-10.4l-10.7-25.5l25.6-10.5l22.8-54.8c-20.5-11.5-36.2-31.2-41.9-55.8c-6.9-30.3 3.1-60.6 23.8-81.1m58 7.2c8.9-.1 17.3 3.5 23.4 9.4c-5.5 3.5-11.6 6.6-18 9.4c-1.6-.6-3.3-.8-5.1-.8c-.6 0-1.1 0-1.6.1c-7 .8-12.2 6.1-13.1 12.7c-.2 1-.2 2-.2 2.9c.1.3.1.7.1 1c1 8.4 8.3 14.2 16.7 13.2c6.8-.8 12-5.9 13-12.3c6.2-2.8 12-5.9 17.5-9.4c.2 1 .4 2 .5 3c2.1 18-11 34.5-29 36.6c-17.9 2.1-34.5-11-36.5-29c-2.1-18 11-34.5 29-36.6c1.1-.1 2.2-.2 3.3-.2"/></svg>',Ht='<svg viewBox="0 0 48 48"><defs><mask id="ipTPeopleDownload0"><g fill="none" stroke="#fff" stroke-linejoin="round" stroke-width="4"><path fill="#555" d="M19 20a7 7 0 1 0 0-14a7 7 0 0 0 0 14Z"/><path stroke-linecap="round" d="M37 40V30m-4 6l4 4l4-4m-14-8h-8.2c-4.48 0-6.72 0-8.432.872a8 8 0 0 0-3.496 3.496C6 34.08 6 36.32 6 40.8V42h21"/></g></mask></defs><path fill="currentColor" d="M0 0h48v48H0z" mask="url(#ipTPeopleDownload0)"/></svg>',Mt='<svg viewBox="0 0 24 24"><g fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"><path d="M21 12a9 9 0 1 0-9 9M3.6 9h16.8M3.6 15H12"/><path d="M11.578 3a17 17 0 0 0 0 18M12.5 3c1.719 2.755 2.5 5.876 2.5 9m3 2v7m-3-3l3 3l3-3"/></g></svg>',Rt='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M229.66 90.34L90.34 229.66a8 8 0 0 1-11.31 0L26.34 177a8 8 0 0 1 0-11.31L165.66 26.34a8 8 0 0 1 11.31 0L229.66 79a8 8 0 0 1 0 11.34" opacity=".2"/><path d="m235.32 73.37l-52.69-52.68a16 16 0 0 0-22.63 0L20.68 160a16 16 0 0 0 0 22.63l52.69 52.68a16 16 0 0 0 22.63 0L235.32 96a16 16 0 0 0 0-22.63M84.68 224L32 171.31l32-32l26.34 26.35a8 8 0 0 0 11.32-11.32L75.31 128L96 107.31l26.34 26.35a8 8 0 0 0 11.32-11.32L107.31 96L128 75.31l26.34 26.35a8 8 0 0 0 11.32-11.32L139.31 64l32-32L224 84.69Z"/></g></svg>',Pt='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M200 56v136a8 8 0 0 1-8 8H64a8 8 0 0 1-8-8V56Z" opacity=".2"/><path d="M208 56v32a8 8 0 0 1-16 0V64h-56v128h24a8 8 0 0 1 0 16H96a8 8 0 0 1 0-16h24V64H64v24a8 8 0 0 1-16 0V56a8 8 0 0 1 8-8h144a8 8 0 0 1 8 8"/></g></svg>',Ze='<svg viewBox="0 0 256 256"><g fill="currentColor"><path d="M200 56v152a8 8 0 0 1-8 8H64a8 8 0 0 1-8-8V56Z" opacity=".2"/><path d="M216 48h-40v-8a24 24 0 0 0-24-24h-48a24 24 0 0 0-24 24v8H40a8 8 0 0 0 0 16h8v144a16 16 0 0 0 16 16h128a16 16 0 0 0 16-16V64h8a8 8 0 0 0 0-16M96 40a8 8 0 0 1 8-8h48a8 8 0 0 1 8 8v8H96Zm96 168H64V64h128Zm-80-104v64a8 8 0 0 1-16 0v-64a8 8 0 0 1 16 0m48 0v64a8 8 0 0 1-16 0v-64a8 8 0 0 1 16 0"/></g></svg>';var xe=document.querySelector("#results"),qt=(e,t,n)=>async()=>{let r,o=n===null?"":n||"bin";if(typeof e=="string"){if(r=new TextEncoder().encode(e),!n)switch(oe(e)){case-5:o="json";break;case-6:o="pem";break;default:o="txt"}}else r=e;let a=new Blob([r],{type:"application/octet-stream"}),i=window.URL.createObjectURL(a),p=document.createElement("a");p.href=i;let m=o?`.${o}`:"";p.download=t?`${t}${m}`:`cryptotools-result${m}`,document.body.appendChild(p),p.click(),document.body.removeChild(p),window.URL.revokeObjectURL(i)},Ot=({label:e,value:t,encoding:n,bitLength:r,rawData:o,filename:a,extension:i})=>{let p=document.createElement("div");p.classList.add("result");let m=document.createElement("h2");m.textContent=e,p.appendChild(m);let b=document.createElement("div");b.classList.add("content"),b.textContent=`${t.substring(0,1e3)}${t.length>1e3?"\u2026":""}`,p.appendChild(b);let u=document.createElement("div");u.classList.add("stats");let v=[{icon:Pt,tooltip:"Character length",statValue:`${t.length.toLocaleString()} characters`},{icon:Rt,tooltip:"Output length",statValue:`${r.toLocaleString()} bits`}],g=n>0&&typeof o=="object",S=[{tooltip:"Download Text",tooltipAfter:"Downloaded!",icon:It,callback:qt(t,a,g?`${J[n].toLocaleLowerCase()}.txt`:i)},{tooltip:"Copy",tooltipAfter:"Copied!",icon:wt,callback:()=>navigator.clipboard.writeText(b.textContent)}];if(g){S.unshift({tooltip:"Download Raw Data",tooltipAfter:"Downloaded!",icon:Ct,callback:qt(o,a,i)});let w=document.createElement("label"),$=document.createElement("span");w.classList.add("control"),$.classList.add("label"),$.textContent="Encoding",w.innerHTML=Ye,w.insertBefore($,w.firstChild);let P=document.createElement("select");P.addEventListener("change",()=>{let q=Number(P.selectedOptions[0].value);try{let B=j(o,q),O=Ot({label:e,value:B,encoding:q,bitLength:r,rawData:o,filename:a,extension:i});p.replaceWith(O)}catch(B){C(B)}}),Object.entries(J).forEach(([q,B])=>{if(typeof B=="number"&&B>0){let O=document.createElement("option");O.textContent=q.toLowerCase(),O.value=String(B),B===n&&(O.selected=!0),P.appendChild(O)}}),w.appendChild(P);let A=document.createElement("span");A.innerHTML=St,w.appendChild(A.firstChild),u.appendChild(w)}else v.unshift({icon:Ye,tooltip:"Encoding",statValue:J[n].toLowerCase()});v.forEach(({icon:w,tooltip:$,statValue:P})=>{let A=document.createElement("div");A.classList.add("datum"),$&&(A.dataset.tooltip=$),A.innerHTML=w;let q=document.createElement("span");q.textContent=P,A.appendChild(q),u.appendChild(A)}),p.appendChild(u);let h=document.createElement("div");return h.classList.add("links"),u.appendChild(h),S.forEach(({tooltip:w,tooltipAfter:$,icon:P,callback:A})=>{let q=document.createElement("a");q.dataset.tooltip=w,q.href="#",q.addEventListener("click",async B=>{B.preventDefault();try{await A(),q.innerHTML=At,q.dataset.tooltip=$,setTimeout(()=>{q.innerHTML=P,q.dataset.tooltip=w},5e3)}catch(O){console.error(O)}}),q.innerHTML=P,h.appendChild(q)}),p},F=e=>{xe.innerHTML=Tt,je(),xe.scrollIntoView({block:"start",behavior:"smooth"}),e.forEach(({label:t,value:n,defaultEncoding:r,filename:o,extension:a})=>{let i,p,m;if(n instanceof ArrayBuffer)i=r||16,p=j(n,i),m=n.byteLength;else{p=n,i=r||oe(n);try{m=U(n,i).byteLength}catch{console.warn(`Cannot calculate raw data byte length for "${t}" because no decoder has been implemented for \`${J[i].toLowerCase()}\`. Assuming utf-8 encoded text instead.`),m=new TextEncoder().encode(n).byteLength}}try{let b=Ot({label:t,value:p,encoding:i,bitLength:m*8,rawData:n,filename:o,extension:a});xe.appendChild(b)}catch(b){C(b)}}),xe.style.opacity="100%",N(100)},Y=()=>{xe.style.opacity="0",xe.textContent="",je()};var Nt=document.querySelectorAll("#main-menu li"),Hn=document.querySelectorAll("main > section");Nt.forEach(e=>e.addEventListener("click",()=>{Nt.forEach(t=>t.classList.remove("active")),Hn.forEach(t=>t.classList.remove("active")),e.classList.add("active"),e.dataset.target&&document.querySelector(e.dataset.target)?.classList.add("active"),Y()}));var ce=document.querySelector("#digest"),He=ce.querySelector(".digest-select"),kt=ce.querySelector("textarea"),Ee=ce.querySelector("details"),Xe=ce.querySelector("button");Ee?.addEventListener("toggle",()=>{Ee.open?Xe.textContent="Verify Digest":Xe.textContent="Generate Digest"});async function Mn(e,t){let n=Number(ce.querySelector(".encoding select")?.selectedOptions[0].value),r=U(e,n);return await crypto.subtle.digest(t,r)}Xe.addEventListener("click",async()=>{N(0);let e=He.selectedOptions[0].dataset.alg,t=kt.value;try{let n=await Mn(t,e);if(Ee?.open){let r=Number(Ee.querySelector("select.input-encoding")?.selectedOptions[0].value),o=Ee.querySelector("input.hash")?.value||"",a=U(o,r);F([{label:`${e} Digest Verified?`,value:String(j(n,16)===j(a,16))}])}else F([{label:`${e} Digest`,value:n,defaultEncoding:16,filename:"digest",extension:e.replace(/-/g,"").toLowerCase()}])}catch(n){C(n)}});He.addEventListener("change",()=>{let e=ce.querySelector("menu"),t=e.querySelector(".block-size span"),n=e.querySelector(".method span"),r=e.querySelector(".specification span"),o=He.selectedOptions[0].dataset;t.textContent=o.bs||"",n.textContent=o.method||"",r.textContent=o.spec||"",Y()});var Rn=async(e,t,n)=>{let r=[];try{let o={};t&&t.split(/\r?\n/).forEach(a=>{let[i,p]=a.split(/\s[\s*]/);o[p]=i}),await Promise.all(e.map(async([a,i])=>{let p=o[a.name],m=await crypto.subtle.digest(n,i),b,u,v="digest",g;p?(b=`${n} Digest of ${a.name} Verifies Checksum? \u2022 ${a.size.toLocaleString()} bytes \u2022 ${a.type||"Unknown type"}`,u=String(p===j(m,16))):(b=`${n} Digest of ${a.name} \u2022 ${a.size.toLocaleString()} bytes \u2022 ${a.type||"Unknown type"}`,u=m,v=a.name,g=n.replace(/-/g,"").toLowerCase()),r.push({label:b,value:u,filename:v,extension:g})}))}catch(o){C(o)}F(r)},_t=e=>{if(!e||!e.length)return;N(0);let t=Array.from(e),n=t.reduce((p,m)=>p+m.size,0),r=[],o=[],a=He.selectedOptions[0].dataset.alg,i;t.forEach((p,m)=>{let b=new FileReader;b.onprogress=u=>{r[m]=u.loaded,N(r.reduce((v,g)=>v+g,0)/n*100)},b.onload=async u=>{if(!u.target?.result||!(u.target.result instanceof ArrayBuffer)){C(new Error(`File [${p.name}] of type [${p.type}] failed to load.`));return}try{let g=new Error("More than one checksum file exists in the file list.");switch(!0){case/^.*\.sha1$|^sha1sums?(\.txt)?$/i.test(p.name):if(a="SHA-1",i)throw g;i=new TextDecoder().decode(u.target.result);break;case/^.*\.sha256$|^sha256sums?(\.txt)?$/i.test(p.name):if(a="SHA-256",i)throw g;i=new TextDecoder().decode(u.target.result);break;case/^.*\.sha384$|^sha384sums?(\.txt)?$/i.test(p.name):if(a="SHA-384",i)throw g;i=new TextDecoder().decode(u.target.result);break;case/^.*\.sha512$|^sha512sums?(\.txt)?$/i.test(p.name):if(a="SHA-512",i)throw g;i=new TextDecoder().decode(u.target.result);break;default:o.push([p,u.target.result])}}catch(g){C(g)}o.length===(i?e.length-1:e.length)&&Rn(o,i,a)},b.readAsArrayBuffer(p)})},$t=ce.querySelector(".upload input");$t?.addEventListener("change",()=>_t($t.files));kt.addEventListener("drop",e=>{_t(e.dataTransfer?.files)});var pe=document.querySelector("#random"),Ft=pe.querySelectorAll("menu li"),Pn=pe.querySelectorAll(".settings");Ft.forEach(e=>e.addEventListener("click",()=>{Ft.forEach(t=>t.classList.remove("active")),Pn.forEach(t=>t.classList.remove("active")),e.classList.add("active"),e.dataset.target&&document.querySelector(`.${e.dataset.target}`)?.classList.add("active"),Y()}));var qn=pe.querySelector("button");qn.addEventListener("click",()=>{N(0);try{let e=pe.querySelector(".byte-length"),t=pe.querySelector(".output");if(pe.querySelector("menu li.active").dataset.target==="uuid"){F([{label:"UUID",value:window.crypto.randomUUID(),filename:"prng-uuid"}]);return}let r=Number(e.value.length?e.value:64),o=new Uint8Array(r);if(window.crypto.getRandomValues(o),t.selectedOptions[0].value==="display"){F([{label:"Random Values",value:o.buffer,filename:`prng-${r}-bytes`}]);return}let i=new Blob([o],{type:"application/octet-stream"}),p=window.URL.createObjectURL(i),m=document.createElement("a");m.href=p,m.download=`prng-${r}-bytes.bin`,document.body.appendChild(m),m.click(),document.body.removeChild(m),window.URL.revokeObjectURL(p)}catch(e){C(e)}N(100)});var Kt=e=>t=>{t.preventDefault(),e.remove()},On=(e,t,n,r,o)=>{let a=document.createElement("div");a.classList.add("modal-background");let i=document.createElement("aside");i.classList.add("modal"),a.appendChild(i);let p=document.createElement("div");p.classList.add("container"),i.appendChild(p);let m=document.createElement("span");m.innerHTML=e,p.appendChild(m);let b=document.createElement("div");b.classList.add("content"),p.appendChild(b);let u=document.createElement("span");u.textContent=t,b.appendChild(u);let v=document.createElement("p");v.textContent=n,b.appendChild(v);let g=document.createElement("div");g.classList.add("actions"),i.appendChild(g);let S=document.createElement("button");S.classList.add("cancel"),S.textContent="Cancel",S.addEventListener("click",Kt(a)),g.appendChild(S);let h=document.createElement("button");h.classList.add("confirm"),h.textContent=r,h.addEventListener("click",Kt(a)),h.addEventListener("click",()=>o()),g.appendChild(h),document.body.appendChild(a)},Vt=On;var Nn="cryptotools",ie="keys";var se,Ut=()=>new Promise((e,t)=>{let n=window.indexedDB.open(Nn,1);n.onerror=()=>{t(new Error("Cannot open database."))},n.onupgradeneeded=()=>{se=n.result,se.createObjectStore(ie,{keyPath:"name"})},n.onsuccess=()=>{se=n.result,se.onerror=r=>console.error(new Error(`A database error ocurred: ${JSON.stringify(r.target)}`)),e(!0)}}),Dt=(e,t)=>new Promise((n,r)=>{let i=se.transaction([ie],"readwrite").objectStore(ie).add({name:e,key:t});i.onerror=p=>r(new Error(JSON.stringify(p))),i.onsuccess=()=>n(!0)}),Gt=e=>new Promise((t,n)=>{let a=se.transaction([ie],"readwrite").objectStore(ie).delete(e);a.onerror=i=>n(new Error(JSON.stringify(i))),a.onsuccess=()=>t(!0)}),jt=()=>new Promise((e,t)=>{let o=se.transaction([ie],"readonly").objectStore(ie).getAll();o.onerror=a=>t(new Error(JSON.stringify(a))),o.onsuccess=a=>e(a.target.result)});var zt=document.querySelector("#encryption"),D=zt.querySelector(".crypto-operations"),Yt=D?.querySelectorAll("menu li");zt.querySelectorAll(".keys menu li").forEach(e=>e.addEventListener("click",()=>D?.classList.remove("active")));var $n=(e,t)=>{let n=D?.querySelector("textarea"),r=D?.querySelector(".signature"),o=D?.querySelector("button"),a="";switch(e){case"encrypt":a="Paste plain text or drop a file...";break;case"decrypt":a="Paste ciphertext or drop a file...";break;case"sign":a="Paste plain text or drop a file...";break;case"verify":a="Paste plain text or drop a file...";break;default:}n&&(n.placeholder=a),D?.querySelectorAll(".input > .settings").forEach(i=>i.classList.remove("active")),D?.querySelector(`.settings.${t}.${e}`)?.classList.add("active"),e==="verify"?r?.classList.add("active"):r?.classList.remove("active"),o&&(o.textContent=`${e.charAt(0).toUpperCase()}${e.slice(1)}`)},Zt=(e,t,n)=>()=>{D?.classList.add("active"),D?.scrollIntoView({block:"center",behavior:"smooth"});let r=t instanceof CryptoKey,o=n,a=e,i,p;if(r)i=t,p=i.usages,o||(p.includes("encrypt")?o="encrypt":p.includes("sign")?o="sign":[o]=p);else{p=t.privateKey.usages.concat(t.publicKey.usages),o||(p.includes("encrypt")?o="encrypt":p.includes("sign")?o="sign":[o]=p);let u="";switch(o){case"verify":case"encrypt":u="Public Key",i=t.publicKey;break;case"sign":case"decrypt":u="Private Key",i=t.privateKey;break;default:throw new Error(`Operation ${o} is not implemented for key "${e}".`)}a=`${e} (${u})`}let m=i.algorithm.name.toLowerCase();D&&(D.dataset.key=e,D.dataset.alg=m),Yt?.forEach(u=>u.classList.remove("active","show")),D?.querySelector(`menu li[data-op="${o}"]`)?.classList.add("active"),p.forEach(u=>D?.querySelector(`menu li[data-op="${u}"]`)?.classList.add("show"));let b=D?.querySelector(".footer .key");if(b){b.innerHTML=Be;let u=document.createElement("span");u.textContent=a,b.appendChild(u)}$n(o,m)};Yt?.forEach(e=>{let t=e.dataset.op;e.addEventListener("click",()=>{let n=D.dataset.key||"",{key:r}=Me(n);Zt(n,r,t)()})});var Xt=Zt;var le={},Je=()=>{document.querySelector("#encryption .operation-area")?.classList.remove("active"),Y();let t=document.querySelector('#encryption [data-tab="list-keys"] ul');t.textContent="";let n=Object.entries(le);if(!n.length){let r=document.createElement("li");r.textContent="No Keys.  Generate a key to get started.",t.appendChild(r)}n.forEach(([r,o])=>{let{key:a}=o,i=document.createElement("li");i.addEventListener("click",Xt(r,a));let p=document.createElement("div");p.classList.add("container"),i.appendChild(p);let m=document.createElement("span");p.appendChild(m);let b=document.createElement("div");b.classList.add("meta"),p.appendChild(b);let u=document.createElement("span");if(u.textContent=r,b.appendChild(u),o.saved){let A=document.createElement("span");A.classList.add("saved"),A.dataset.tooltip="Persists after page refresh",A.innerHTML=Lt,u.appendChild(A)}let v=a instanceof CryptoKey,g=document.createElement("span"),S,h,w;v?(m.innerHTML=Be,S="Symmetric Key",h=a.algorithm.name,w=a.usages):(m.innerHTML=Bt,S="Asymmetric Key Pair",h=a.privateKey.algorithm.name,w=a.privateKey.usages.concat(a.publicKey.usages)),g.textContent=`${S} \u2022 ${h} \u2022 ${w.join(", ")}`,b.appendChild(g);let $=document.createElement("div");if($.classList.add("actions"),i.appendChild($),!v){let A=document.createElement("a");A.href="#",A.innerHTML=Mt,A.dataset.tooltip="Export Public Key",A.addEventListener("click",async q=>{q.preventDefault(),q.stopPropagation();let B=a.publicKey,O=`${r.replace(/\s/g,"-")}.public`,z=[];try{switch(B.algorithm.name){case"RSASSA-PKCS1-v1_5":case"RSA-PSS":case"RSA-OAEP":case"ECDSA":{let Z=await be(B,"spki")();z.push({label:`Public Key "${r}" in Subject Public Key Info (SPKI) Format, DER-Encoded`,value:Z,defaultEncoding:64,filename:O,extension:"der"},{label:`Public Key "${r}" in Subject Public Key Info (SPKI) Format, PEM-Encoded`,value:`-----BEGIN PUBLIC KEY-----
 ${j(Z,64)}
 -----END PUBLIC KEY-----
 `,filename:O,extension:"pem"});break}default:throw new Error(`Unsupported key algorithm ${B.algorithm.name}`)}B.algorithm.name==="ECDSA"&&z.push({label:`Public Key "${r}" in Raw Format`,value:await be(B,"raw")(),defaultEncoding:64,filename:O}),z.push({label:`Public Key "${r}" in JSON Web Key (JWK) Format`,value:JSON.stringify(await be(B,"jwk")(),null,2),filename:O}),F(z)}catch(Z){C(Z)}}),$.appendChild(A)}if(!v&&a.privateKey.extractable||v&&a.extractable){let A=document.createElement("a");A.href="#",A.innerHTML=Ht,A.dataset.tooltip=v?"Export Secret Key":"Export Private Key",A.addEventListener("click",async q=>{q.preventDefault(),q.stopPropagation();let B=v?a:a.privateKey,O=`${r.replace(/\s/g,"-")}.${v?"secret":"private"}`,z=[];try{switch(B.algorithm.name){case"AES-CTR":case"AES-CBC":case"AES-GCM":case"HMAC":z.push({label:`Secret Key "${r}" in Raw Format`,value:await be(B,"raw")(),filename:O,extension:"key"});break;case"RSASSA-PKCS1-v1_5":case"RSA-PSS":case"RSA-OAEP":case"ECDSA":{let Z=await be(B,"pkcs8")();z.push({label:`Private Key "${r}" in PKCS#8 Format, DER-Encoded`,value:Z,defaultEncoding:64,filename:O,extension:"der"},{label:`Private Key "${r}" in PKCS#8 Format, PEM-Encoded`,value:`-----BEGIN PRIVATE KEY-----
 ${j(Z,64)}
 -----END PRIVATE KEY-----
-`,filename:O,extension:"key"});break}default:throw new Error(`Unsupported key algorithm ${B.algorithm.name}`)}z.push({label:`${v?"Secret":"Private"} Key "${r}" in JSON Web Key (JWK) Format`,value:JSON.stringify(await be(B,"jwk")(),null,2),filename:O}),F(z)}catch(Z){C(Z)}}),$.appendChild(A)}let P=document.createElement("a");P.href="#",P.innerHTML=Ze,P.dataset.tooltip="Delete key",P.addEventListener("click",A=>{A.preventDefault(),A.stopPropagation(),Vt(Ze,"Are you sure?","This is an unrecoverable operation.","Delete",_n(r))}),$.appendChild(P),t.appendChild(i)})},kn=()=>document.querySelector('#encryption li[data-target="list-keys"]')?.click(),Jt=(e,t,n=!1)=>{if(ie[e])throw new Error(`A key by the name of "${e}" already exists.`);ie[e]={saved:n,key:t},n&&Dt(e,t),Je(),kn(),N(100)},Me=e=>{if(!ie[e])throw new Error(`A key by the name of "${e}" does not exist.`);return ie[e]},_n=e=>async()=>{try{await Gt(e),delete ie[e],Je()}catch(t){C(t)}},be=(e,t)=>async()=>window.crypto.subtle.exportKey(t,e);(async()=>{try{await Ut(),(await jt()).forEach(({name:t,key:n})=>{ie[t]={saved:!0,key:n}}),Je()}catch(e){C(e)}})();var k=document.querySelector("#encryption .generate-key"),Pe=k.querySelector(".control.purpose select"),Wt=()=>{let e=k.querySelector(".settings.algorithm.active select"),t=e.selectedOptions[0],n=k.querySelector(".stats .type span"),r=k.querySelector(".stats .method span"),o=k.querySelector(".stats .specification span"),a=k.querySelector(".stats .confidentiality"),i=k.querySelector(".stats .integrity");k.querySelectorAll("div.settings").forEach(m=>{m.classList.contains("algorithm")||m.classList.remove("active")}),k.querySelector(`.settings.${t.dataset.target}`)?.classList.add("active"),e.parentElement.querySelector(".label").textContent=`Algorithm \u2022 ${t.dataset.mode}`,n.textContent=t.dataset.type||"",r.textContent=t.dataset.method||"",o.textContent=t.dataset.spec||"",t.dataset.confidentiality==="true"?(a.classList.remove("no"),a.dataset.tooltip="Provides Confidentiality"):(a.classList.add("no"),a.dataset.tooltip="Provides NO Confidentiality"),t.dataset.integrity==="true"?(i.classList.remove("no"),i.dataset.tooltip="Provides Integrity Protection"):(i.classList.add("no"),i.dataset.tooltip="Provides NO Integrity Protection")};Pe.addEventListener("change",()=>{let e=Pe.selectedOptions[0].value;k.querySelectorAll(".settings.algorithm").forEach(n=>n.classList.remove("active")),k.querySelector(`.settings.${e}`)?.classList.add("active"),Wt()});k.querySelectorAll(".settings.algorithm").forEach(e=>e.addEventListener("change",()=>Wt()));var Fn=e=>{let t=e,n=[];for(;t>0;)n.unshift(t&255),t>>=8;return new Uint8Array(n)},Re=k.querySelector("button");Re?.addEventListener("click",async()=>{Re.disabled=!0,N(0);let t=k.querySelector(".name input").value,r=k.querySelector(".settings.algorithm.active select").selectedOptions[0];try{if(!t)throw new Error("A key name is required.");let a,i=Pe.value==="encryption"?["encrypt","decrypt"]:["sign","verify"];switch(r.dataset.target){case"rsa":{let u=Number(k.querySelector(".settings.rsa input.modulus-length")?.value)||2048,v=Number(k.querySelector(".settings.rsa input.public-exponent")?.value)||65537,g=k.querySelector(".settings.rsa select.hash")?.value||"SHA-256";a={name:r.dataset.alg||"",modulusLength:u,publicExponent:Fn(v),hash:g};break}case"ec":{let u=k.querySelector(".settings.ec select.curve")?.value||"P-256";a={name:r.dataset.alg||"",namedCurve:u};break}case"hmac":{let u=k.querySelector(".settings.hmac select.hash")?.value||"SHA-256";a={name:r.dataset.alg||"",hash:u};break}case"aes":{let u=Number(k.querySelector(".settings.aes select.key-length")?.value)||256;a={name:r.dataset.alg||"",length:u};break}default:throw new Error("No algorithm selected.")}let p=k.querySelector(".extractable input")?.checked||!1,m=await window.crypto.subtle.generateKey(a,p,i),b=k.querySelector(".save input")?.checked;Jt(t,m,b)}catch(a){C(a),Re.disabled=!1;return}k.querySelector("form")?.reset();let o=new Event("change");Pe.dispatchEvent(o),Re.disabled=!1});var Kn=document.querySelector("#encryption"),H=Kn.querySelector(".encrypt.operation-area"),en=H.querySelector(".input textarea"),tn=async(e,t,n,r="")=>{let o=[],a=r?` \u2022 ${r}`:"",i=Me(H.dataset.key||"").key,p=i instanceof CryptoKey,m=n?`${n}-${e}`:e;switch(e){case"verify":{let b=H.querySelector(".signature textarea");if(!b.value)throw new Error("A signature is required to verify the plaintext.");let u=Number(H.querySelector(".signature .encoding select")?.selectedOptions[0].value),v=U(b.value,u),g=p?i:i.publicKey,S={name:g.algorithm.name};if(g.algorithm.name==="RSA-PSS"){let w=H.querySelector(".rsa-pss.verify input.salt-length")?.value;S.saltLength=w?Number(w):32}if(g.algorithm.name==="ECDSA"){let w=H.querySelector(".ecdsa.verify select")?.selectedOptions[0].value;S.hash=w}let h=await window.crypto.subtle.verify(S,g,v,t);o.unshift({label:"Signature verifies data?",value:String(h)});break}case"encrypt":{let b=p?i:i.publicKey,u={name:b.algorithm.name};if(b.algorithm.name==="AES-CTR"){let g=H.querySelector(".aes-ctr.encrypt input.counter")?.value,S=H.querySelector(".aes-ctr.encrypt select")?.selectedOptions[0].value,h;if(g){let w=Number(S);if(h=U(g,w),h.byteLength!==16)throw new Error("Counter length must be exactly 16 bytes.")}else h=crypto.getRandomValues(new Uint8Array(16)).buffer;u.counter=h,u.length=64,o.push({label:"Counter \u2022 Needed to decrypt",value:h,defaultEncoding:64,filename:`${m}-counter`},{label:"Counter Length \u2022 Needed to decrypt",value:u.length,defaultEncoding:-3,filename:`${m}-counter-length`})}if(b.algorithm.name==="AES-CBC"){let g=H.querySelector(".aes-cbc.encrypt .iv input")?.value,S=H.querySelector(".aes-cbc.encrypt .iv select")?.selectedOptions[0].value,h;if(g){let w=Number(S);if(h=U(g,w),h.byteLength!==16)throw new Error("IV length must be exactly 16 bytes.")}else h=crypto.getRandomValues(new Uint8Array(16)).buffer;u.iv=h,o.push({label:"Initialization Vector (IV) \u2022 Needed to decrypt \u2022 Not Secret",value:h,defaultEncoding:64,filename:`${m}-iv`})}if(b.algorithm.name==="AES-GCM"){let g=H.querySelector(".aes-gcm.encrypt .iv input")?.value,S=H.querySelector(".aes-gcm.encrypt .iv select")?.selectedOptions[0].value,h=H.querySelector(".aes-gcm.encrypt textarea"),w=H.querySelector(".aes-gcm.encrypt .tag-length select")?.selectedOptions[0].value;u.tagLength=Number(w)||128;let $;if(g){let P=Number(S);$=U(g,P)}else $=crypto.getRandomValues(new Uint8Array(12)).buffer;if(u.iv=$,o.unshift({label:"Initialization Vector (IV) \u2022 Needed to decrypt \u2022 Not Secret",value:$,defaultEncoding:64,filename:`${m}-iv`}),h?.value.length){let P=new TextEncoder().encode(h?.value);u.additionalData=P,o.push({label:"Authenticated but Unencrypted Data \u2022 Needed to decrypt",value:h.value,defaultEncoding:100,filename:`${m}-authenticated-data`})}}let v=await window.crypto.subtle.encrypt(u,b,t);o.unshift({label:`Ciphertext${a}`,value:v,defaultEncoding:64,filename:n?`${n}.enc`:"ciphertext"});break}case"sign":{let b=p?i:i.privateKey,u={name:b.algorithm.name};if(b.algorithm.name==="RSA-PSS"){let g=H.querySelector(".rsa-pss.sign input.salt-length")?.value,S;switch(b.algorithm.hash.name){case"SHA-1":S=160/8;break;case"SHA-256":S=256/8;break;case"SHA-384":S=384/8;break;case"SHA-512":S=512/8;break;default:S=0}u.saltLength=g===""?S:Number(g),o.push({label:"RSA-PSS Salt Length \u2022 Needed for signature verification",value:String(u.saltLength),defaultEncoding:-3,filename:`${m}-salt-length`});let h=b.algorithm.modulusLength;console.debug("Limit",Math.ceil((h-1)/8)-S-2)}if(b.algorithm.name==="ECDSA"){let g=H.querySelector(".ecdsa.sign select")?.selectedOptions[0].value;u.hash=g,o.push({label:"Hash Function \u2022 Needed for signature verification",value:String(g),defaultEncoding:100,filename:`${m}-hash`})}let v=await window.crypto.subtle.sign(u,b,t);o.unshift({label:`Signature${a}`,value:v,defaultEncoding:64,filename:n?`${n}.sig`:"signature"});break}case"decrypt":{let b=p?i:i.privateKey,u={name:b.algorithm.name};if(b.algorithm.name==="AES-CTR"){let g=H.querySelector(".aes-ctr.decrypt input.counter")?.value,S=H.querySelector(".aes-ctr.decrypt select")?.selectedOptions[0].value,h;if(g){let w=Number(S);if(h=U(g,w),h.byteLength!==16)throw new Error("Counter length must be exactly 16 bytes.")}else throw new Error("AES-CTR requires the initial counter value in order to decrypt the ciphertext.");u.counter=h,u.length=64}if(b.algorithm.name==="AES-CBC"){let g=H.querySelector(".aes-cbc.decrypt .iv input")?.value,S=H.querySelector(".aes-cbc.decrypt .iv select")?.selectedOptions[0].value,h;if(g){let w=Number(S);if(h=U(g,w),h.byteLength!==16)throw new Error("IV length must be exactly 16 bytes.")}else throw new Error("AES-CBC requires the Initialization Vector (IV) in order to decrypt the ciphertext.");u.iv=h}if(b.algorithm.name==="AES-GCM"){let g=H.querySelector(".aes-gcm.decrypt .iv input")?.value,S=H.querySelector(".aes-gcm.decrypt .iv select")?.selectedOptions[0].value,h=H.querySelector(".aes-gcm.decrypt textarea"),w=H.querySelector(".aes-gcm.decrypt .tag-length select")?.selectedOptions[0].value;u.tagLength=Number(w)||128;let $;if(g){let P=Number(S);$=U(g,P)}else throw new Error("AES-GCM requires the Initialization Vector (IV) in order to decrypt the ciphertext.");if(u.iv=$,h?.value.length){let P=new TextEncoder().encode(h?.value);u.additionalData=P}}let v=await window.crypto.subtle.decrypt(u,b,t);o.unshift({label:`Plaintext${a}`,value:v,defaultEncoding:100,filename:n?n.replace(/(\.?enc)|(\.?bin)/ig,""):"plaintext",extension:null});break}default:throw new Error(`Operation ${e} is not implemented.`)}return o},Vn=H.querySelector("button");Vn?.addEventListener("click",async()=>{N(0);try{let e=H.querySelector("menu li.active")?.dataset.op;if(!e)throw new Error("An operation must be defined.");let t=Number(H.querySelector(".input .encoding select")?.selectedOptions[0].value),n=U(en.value,t),r=await tn(e,n);F(r)}catch(e){C(e)}});var nn=e=>{if(!(!e||!e.length)){N(0);try{let t=H.querySelector("menu li.active")?.dataset.op;if(!t)throw new Error("An operation must be defined.");let n=Array.from(e),r=n.reduce((i,p)=>i+p.size,0),o=[],a=[];n.forEach((i,p)=>{let m=new FileReader;m.onprogress=b=>{o[p]=b.loaded,N(o.reduce((u,v)=>u+v,0)/r*100)},m.onload=async b=>{if(!b.target?.result||!(b.target.result instanceof ArrayBuffer)){C(new Error(`File [${i.name}] of type [${i.type}] failed to load.`));return}let u=`${i.name} \u2022 ${i.size.toLocaleString()} bytes \u2022 ${i.type||"Unknown type"}`;try{a.push(await tn(t,b.target.result,i.name,u))}catch(v){C(v)}a.length===e.length&&F(a.flat())},m.readAsArrayBuffer(i)})}catch(t){C(t)}}},Qt=H.querySelector(".upload input");Qt?.addEventListener("change",()=>nn(Qt.files));en.addEventListener("drop",e=>{nn(e.dataTransfer?.files)});var me=De(ze());var Un=document.querySelector("#hash-bcrypt .control.cost span:first-child"),We=document.querySelector("#hash-bcrypt input.cost");We.addEventListener("change",()=>{let e=parseInt(We.value,10);Un.textContent=`Cost \u2022 ${(2**e).toLocaleString()} iterations`});var rn=(e,t)=>(n,r)=>{if(n&&C(n),typeof r=="boolean")F([{label:"Bcrypt Verification Result",value:String(r)}]);else{let a=me.getSalt(r),i=a.replace(/^.+\$/,""),p=r.substring(a.length);F([{label:"Bcrypt String",value:r,filename:"bcrypt-string"},{label:"Salt",value:i,filename:"bcrypt-salt"},{label:"Hash",value:p,filename:"bcrypt-hash"}])}let o=e;o.textContent=t,o.disabled=!1},an=e=>t=>{N(t*100);let n=e;n.textContent=`${String(t*100).substring(0,4)}%`},Se=document.querySelector("#hash-bcrypt button");Se?.addEventListener("click",()=>{N(0),Se.disabled=!0;try{let e=parseInt(We.value,10)||10,t=document.querySelector("#hash-bcrypt .password input");me.hash(t.value,e,rn(Se,Se.textContent),an(Se))}catch(e){C(e)}});var we=document.querySelector("#verify-bcrypt button");we?.addEventListener("click",async()=>{N(0),we.disabled=!0;let e=document.querySelector("#verify-bcrypt .password input"),t=document.querySelector("#verify-bcrypt .hash input");try{me.compare(e.value,t.value,rn(we,we.textContent),an(we))}catch(n){C(n)}});var le=document.querySelector("#hash-PBKDF2"),et=le.querySelector(".prf-select"),qe=le.querySelector("input.derivation-length"),on=le.querySelector("input.iterations"),Dn=le.querySelector("select.input-encoding"),cn=le.querySelector("input.salt"),Gn=le.querySelector(".password input"),Qe=le.querySelector("button");qe?.addEventListener("change",()=>{let e=document.querySelector("#passwords menu .output-length span");e&&(e.textContent=`${parseInt(qe.value,10)*8} bits`)});et?.addEventListener("change",()=>{let e=document.querySelector("#passwords menu .block-size span");e&&(e.textContent=String(et.selectedOptions[0].dataset.bs))});Qe?.addEventListener("click",async()=>{N(0),Qe.disabled=!0;try{let e=new TextEncoder().encode(Gn.value),t=await crypto.subtle.importKey("raw",e,{name:"PBKDF2"},!1,["deriveBits"]),n;if(cn.value){let g=Number(Dn.selectedOptions[0].value);n=U(cn.value,g)}else n=crypto.getRandomValues(new Uint8Array(16)).buffer;let r=et.selectedOptions[0].dataset.alg,o=Number(on.value.length?on.value:1e5),a={name:"PBKDF2",salt:n,hash:r,iterations:o},i=Number(qe.value.length?qe.value:32)*8;if(i<0)throw new Error("Byte length must be positive");let p=await crypto.subtle.deriveBits(a,t,i),m=g=>g.replace(/=/g,""),b=m(j(n,64)),u=m(j(p,64)),v=`$pbkdf2$prf=hmac-${r.toLowerCase()},c=${o},dklen=${i}$${b}$${u}`;F([{label:"PBKDF2 String in PHC String Format (recommended by author)",value:v,filename:"pbkdf2-string"},{label:"Hash",value:p,defaultEncoding:64,filename:"pbkdf2-hash"},{label:"Salt",value:n,defaultEncoding:64,filename:"pbkdf2-salt"}])}catch(e){C(e)}Qe.disabled=!1});var G=De(mn());var ft=class extends G.SRPRoutines{computeIdentityHash(t,n){return this.hash(new TextEncoder().encode(`${t}:${n}`).buffer)}},xt=({strict:e,primeGroup:t,hashAlgorithm:n})=>{let r=G.SRPParameters.PrimeGroup[t],o=G.SRPParameters.H[n],a=new G.SRPParameters(r,o);return e?new ft(a):new G.SRPRoutines(a)},ir=async({routines:e,username:t,password:n})=>{let{s:r,v:o}=await(0,G.createVerifierAndSalt)(e,t,n);return{salt:r,verifier:o}},lr=async e=>{let{routines:t,username:n,password:r,salt:o,serverPublicKey:a}=e,m=await(await new G.SRPClientSession(t).step1(n,r)).step2(o,a),{A:b,M1:u}=m;return{A:b,M1:u}},dr=async e=>{let{routines:t,username:n,password:r,salt:o,verifier:a}=e,i=new G.SRPClientSession(t),p=new G.SRPServerSession(t),m=await i.step1(n,r),b=await p.step1(n,o,a),u=await m.step2(o,b.B),v=await b.step2(u.A,u.M1);await u.step3(v);let{B:g}=b,{A:S,M1:h}=u;return{B:g,A:S,M1:h,M2:v}},ye=document.querySelector("#register-SRP"),ur=ye.querySelector("button"),re=document.querySelector("#authenticate-SRP"),fr=re.querySelector("button"),ae=document.querySelector("#prove-identity-SRP"),xr=ae.querySelector("button");ur.addEventListener("click",async()=>{let e=ye.querySelector("input.strict").checked,t=Number(ye.querySelector("select.prime-group").selectedOptions[0].value),n=String(ye.querySelector("select.hash-algorithm").selectedOptions[0].dataset.alg),r=ye.querySelector("input.identity")?.value||"",o=ye.querySelector('input[type="password"]')?.value||"";try{let a=xt({strict:e,primeGroup:t,hashAlgorithm:n}),{salt:i,verifier:p}=await ir({routines:a,username:r,password:o});F([{label:"Salt \u2022 Store in user record",value:i.toString(),defaultEncoding:-2,filename:"srp-salt"},{label:"Verifier \u2022 Store in user record",value:p.toString(),defaultEncoding:-2,filename:"srp-verifier"},{label:"Prime",value:G.SRPParameters.PrimeGroup[t].N.toString(),defaultEncoding:-2,filename:"srp-prime"}])}catch(a){C(a)}});fr.addEventListener("click",async()=>{let e=re.querySelector("input.strict").checked,t=Number(re.querySelector("select.prime-group").selectedOptions[0].value),n=String(re.querySelector("select.hash-algorithm").selectedOptions[0].dataset.alg),r=re.querySelector("input.identity")?.value||"",o=re.querySelector('input[type="password"]')?.value||"";try{let a=BigInt(re.querySelector(".salt textarea")?.value||""),i=BigInt(re.querySelector(".verifier textarea")?.value||""),p=xt({strict:e,primeGroup:t,hashAlgorithm:n}),m=[],{B:b,A:u,M1:v,M2:g}=await dr({routines:p,username:r,password:o,salt:a,verifier:i});m.push({label:"Client Public Key A",value:u.toString(),defaultEncoding:-2,filename:"srp-client-public-key"}),m.push({label:"Client Proof M1",value:v.toString(),defaultEncoding:-2,filename:"srp-client-proof"}),m.push({label:"Server Public Key B",value:b.toString(),defaultEncoding:-2,filename:"srp-server-public-key"}),m.push({label:"Server Proof M2",value:g.toString(),defaultEncoding:-2,filename:"srp-server-proof"}),F(m)}catch(a){C(a)}});xr.addEventListener("click",async()=>{let e=ae.querySelector("input.strict").checked,t=Number(ae.querySelector("select.prime-group").selectedOptions[0].value),n=String(ae.querySelector("select.hash-algorithm").selectedOptions[0].dataset.alg),r=ae.querySelector("input.identity")?.value||"",o=ae.querySelector('input[type="password"]')?.value||"",a=ae.querySelector(".public-key textarea")?.value||"";try{let i=BigInt(ae.querySelector(".salt textarea")?.value||""),p=xt({strict:e,primeGroup:t,hashAlgorithm:n}),m=[],b=BigInt(a),{A:u,M1:v}=await lr({routines:p,username:r,password:o,salt:i,serverPublicKey:b});m.push({label:"Client Public Key A",value:u.toString(),defaultEncoding:-2,filename:"srp-client-public-key"}),m.push({label:"Client Proof M1",value:v.toString(),defaultEncoding:-2,filename:"srp-client-proof"}),F(m)}catch(i){C(i)}});var pt=document.querySelector("#passwords menu .algorithm select"),Ve=document.querySelector("#passwords .operation"),hn=()=>{let e=Ve.selectedOptions[0].value,t=pt.selectedOptions[0].dataset.alg,n=document.querySelector(`#passwords #${e}-${t}`);document.querySelectorAll("#passwords .subsection > .settings").forEach(o=>o.classList.remove("active")),n?.classList.add("active")},pr=()=>{let e=document.querySelector("#passwords menu"),t=e.querySelector(".block-size span"),n=e.querySelector(".method span"),r=e.querySelector(".specification span"),o=pt.selectedOptions[0].dataset;t.textContent=o.bs||"",n.textContent=o.method||"",r.textContent=o.spec||"",Ve.textContent="",o.operations?.split(" ").forEach((a,i)=>{let p=document.createElement("option");i===0&&(p.selected=!0),p.value=a,p.textContent=`${a[0].toUpperCase()}${a.substring(1).replace("-"," ")}`,Ve.appendChild(p)}),Y(),hn()};pt.addEventListener("change",pr);Ve.addEventListener("change",hn);var br=e=>{e.preventDefault(),e.target.checkValidity()||e.stopImmediatePropagation()};document.addEventListener("DOMContentLoaded",()=>{document.querySelectorAll("form")?.forEach(e=>e.addEventListener("submit",br))});var yn=document.querySelectorAll(".operation-area"),mr=e=>()=>{let t=e.parentElement.querySelector(".encoding select")||e.parentElement.parentElement.querySelector(".encoding select"),n=fe(e.value);n&&t?.childNodes.forEach(r=>{let o=r;Number(o.value)===n&&(o.selected=!0)})},hr=(e,t)=>()=>{let n=fe(t.value);n&&e.childNodes.forEach(r=>{let o=r;Number(o.value)===n&&(o.selected=!0)})};yn.forEach(e=>{let t=e.querySelectorAll("textarea"),n=e.querySelectorAll("select.input-encoding");t.forEach(r=>r.addEventListener("paste",()=>setTimeout(mr(r),0))),n.forEach(r=>{let o=r.parentElement.querySelector("input");o?.addEventListener("paste",()=>setTimeout(hr(r,o),0))})});var bt=e=>{e.stopPropagation(),e.preventDefault()},Te=(e,t)=>()=>{t?e.classList.add("dragover"):e.classList.remove("dragover")};yn.forEach(e=>{e.querySelectorAll(".input > textarea").forEach(n=>{n?.addEventListener("dragenter",bt),n?.addEventListener("dragenter",Te(n,!0)),n?.addEventListener("dragover",bt),n?.addEventListener("dragover",Te(n,!0)),n?.addEventListener("dragleave",Te(n,!1)),n?.addEventListener("dragend",Te(n,!1)),n?.addEventListener("drop",bt),n?.addEventListener("drop",Te(n,!1)),n?.addEventListener("input",()=>{let r=n.parentElement.parentElement.querySelector(".character-count"),o=n.value.length;o===1?r.textContent="1 character":r.textContent=`${o} characters`})})});var mt=e=>{let t=e;t.target?.nodeName!=="TEXTAREA"&&t.dataTransfer&&(t.preventDefault(),t.dataTransfer.effectAllowed="none",t.dataTransfer.dropEffect="none")};window.addEventListener("dragover",mt);window.addEventListener("dragenter",mt);window.addEventListener("drop",mt);var yr=document.querySelectorAll(".tabs");yr.forEach(e=>{let t=Array.from(e.children),n=e.parentElement,r=n.querySelectorAll(":scope > section.settings");t.forEach(o=>{o.classList.contains("datum")||o.addEventListener("click",()=>{t.forEach(a=>a.classList.remove("active")),r?.forEach(a=>a.classList.remove("active")),o.classList.add("active"),n.querySelector(`.settings.${o.dataset.target}`)?.classList.add("active"),Y()})})});})();
+`,filename:O,extension:"key"});break}default:throw new Error(`Unsupported key algorithm ${B.algorithm.name}`)}z.push({label:`${v?"Secret":"Private"} Key "${r}" in JSON Web Key (JWK) Format`,value:JSON.stringify(await be(B,"jwk")(),null,2),filename:O}),F(z)}catch(Z){C(Z)}}),$.appendChild(A)}let P=document.createElement("a");P.href="#",P.innerHTML=Ze,P.dataset.tooltip="Delete key",P.addEventListener("click",A=>{A.preventDefault(),A.stopPropagation(),Vt(Ze,"Are you sure?","This is an unrecoverable operation.","Delete",_n(r))}),$.appendChild(P),t.appendChild(i)})},kn=()=>document.querySelector('#encryption li[data-target="list-keys"]')?.click(),Jt=(e,t,n=!1)=>{if(le[e])throw new Error(`A key by the name of "${e}" already exists.`);le[e]={saved:n,key:t},n&&Dt(e,t),Je(),kn(),N(100)},Me=e=>{if(!le[e])throw new Error(`A key by the name of "${e}" does not exist.`);return le[e]},_n=e=>async()=>{try{await Gt(e),delete le[e],Je()}catch(t){C(t)}},be=(e,t)=>async()=>window.crypto.subtle.exportKey(t,e);(async()=>{try{await Ut(),(await jt()).forEach(({name:t,key:n})=>{le[t]={saved:!0,key:n}}),Je()}catch(e){C(e)}})();var k=document.querySelector('#encryption [data-tab="generate-key"]'),Pe=k.querySelector(".control.purpose select"),Wt=()=>{let e=k.querySelector(".settings.algorithm.active select"),t=e.selectedOptions[0],n=k.querySelector(".stats .type span"),r=k.querySelector(".stats .method span"),o=k.querySelector(".stats .specification span"),a=k.querySelector(".stats .confidentiality"),i=k.querySelector(".stats .integrity");k.querySelectorAll("div.settings").forEach(m=>{m.classList.contains("algorithm")||m.classList.remove("active")}),k.querySelector(`.settings.${t.dataset.target}`)?.classList.add("active"),e.parentElement.querySelector(".label").textContent=`Algorithm \u2022 ${t.dataset.mode}`,n.textContent=t.dataset.type||"",r.textContent=t.dataset.method||"",o.textContent=t.dataset.spec||"",t.dataset.confidentiality==="true"?(a.classList.remove("no"),a.dataset.tooltip="Provides Confidentiality"):(a.classList.add("no"),a.dataset.tooltip="Provides NO Confidentiality"),t.dataset.integrity==="true"?(i.classList.remove("no"),i.dataset.tooltip="Provides Integrity Protection"):(i.classList.add("no"),i.dataset.tooltip="Provides NO Integrity Protection")};Pe.addEventListener("change",()=>{let e=Pe.selectedOptions[0].value;k.querySelectorAll(".settings.algorithm").forEach(n=>n.classList.remove("active")),k.querySelector(`.settings.${e}`)?.classList.add("active"),Wt()});k.querySelectorAll(".settings.algorithm").forEach(e=>e.addEventListener("change",()=>Wt()));var Fn=e=>{let t=e,n=[];for(;t>0;)n.unshift(t&255),t>>=8;return new Uint8Array(n)},Re=k.querySelector("button");Re?.addEventListener("click",async()=>{Re.disabled=!0,N(0);let t=k.querySelector(".name input").value,r=k.querySelector(".settings.algorithm.active select").selectedOptions[0];try{if(!t)throw new Error("A key name is required.");let a,i=Pe.value==="encryption"?["encrypt","decrypt"]:["sign","verify"];switch(r.dataset.target){case"rsa":{let u=Number(k.querySelector(".settings.rsa input.modulus-length")?.value)||2048,v=Number(k.querySelector(".settings.rsa input.public-exponent")?.value)||65537,g=k.querySelector(".settings.rsa select.hash")?.value||"SHA-256";a={name:r.dataset.alg||"",modulusLength:u,publicExponent:Fn(v),hash:g};break}case"ec":{let u=k.querySelector(".settings.ec select.curve")?.value||"P-256";a={name:r.dataset.alg||"",namedCurve:u};break}case"hmac":{let u=k.querySelector(".settings.hmac select.hash")?.value||"SHA-256";a={name:r.dataset.alg||"",hash:u};break}case"aes":{let u=Number(k.querySelector(".settings.aes select.key-length")?.value)||256;a={name:r.dataset.alg||"",length:u};break}default:throw new Error("No algorithm selected.")}let p=k.querySelector(".extractable input")?.checked||!1,m=await window.crypto.subtle.generateKey(a,p,i),b=k.querySelector(".save input")?.checked;Jt(t,m,b)}catch(a){C(a),Re.disabled=!1;return}k.querySelector("form")?.reset();let o=new Event("change");Pe.dispatchEvent(o),Re.disabled=!1});var Kn=document.querySelector("#encryption"),H=Kn.querySelector(".encrypt.operation-area"),en=H.querySelector(".input textarea"),tn=async(e,t,n,r="")=>{let o=[],a=r?` \u2022 ${r}`:"",i=Me(H.dataset.key||"").key,p=i instanceof CryptoKey,m=n?`${n}-${e}`:e;switch(e){case"verify":{let b=H.querySelector(".signature textarea");if(!b.value)throw new Error("A signature is required to verify the plaintext.");let u=Number(H.querySelector(".signature .encoding select")?.selectedOptions[0].value),v=U(b.value,u),g=p?i:i.publicKey,S={name:g.algorithm.name};if(g.algorithm.name==="RSA-PSS"){let w=H.querySelector(".rsa-pss.verify input.salt-length")?.value;S.saltLength=w?Number(w):32}if(g.algorithm.name==="ECDSA"){let w=H.querySelector(".ecdsa.verify select")?.selectedOptions[0].value;S.hash=w}let h=await window.crypto.subtle.verify(S,g,v,t);o.unshift({label:"Signature verifies data?",value:String(h)});break}case"encrypt":{let b=p?i:i.publicKey,u={name:b.algorithm.name};if(b.algorithm.name==="AES-CTR"){let g=H.querySelector(".aes-ctr.encrypt input.counter")?.value,S=H.querySelector(".aes-ctr.encrypt select")?.selectedOptions[0].value,h;if(g){let w=Number(S);if(h=U(g,w),h.byteLength!==16)throw new Error("Counter length must be exactly 16 bytes.")}else h=crypto.getRandomValues(new Uint8Array(16)).buffer;u.counter=h,u.length=64,o.push({label:"Counter \u2022 Needed to decrypt",value:h,defaultEncoding:64,filename:`${m}-counter`},{label:"Counter Length \u2022 Needed to decrypt",value:u.length,defaultEncoding:-3,filename:`${m}-counter-length`})}if(b.algorithm.name==="AES-CBC"){let g=H.querySelector(".aes-cbc.encrypt .iv input")?.value,S=H.querySelector(".aes-cbc.encrypt .iv select")?.selectedOptions[0].value,h;if(g){let w=Number(S);if(h=U(g,w),h.byteLength!==16)throw new Error("IV length must be exactly 16 bytes.")}else h=crypto.getRandomValues(new Uint8Array(16)).buffer;u.iv=h,o.push({label:"Initialization Vector (IV) \u2022 Needed to decrypt \u2022 Not Secret",value:h,defaultEncoding:64,filename:`${m}-iv`})}if(b.algorithm.name==="AES-GCM"){let g=H.querySelector(".aes-gcm.encrypt .iv input")?.value,S=H.querySelector(".aes-gcm.encrypt .iv select")?.selectedOptions[0].value,h=H.querySelector(".aes-gcm.encrypt textarea"),w=H.querySelector(".aes-gcm.encrypt .tag-length select")?.selectedOptions[0].value;u.tagLength=Number(w)||128;let $;if(g){let P=Number(S);$=U(g,P)}else $=crypto.getRandomValues(new Uint8Array(12)).buffer;if(u.iv=$,o.unshift({label:"Initialization Vector (IV) \u2022 Needed to decrypt \u2022 Not Secret",value:$,defaultEncoding:64,filename:`${m}-iv`}),h?.value.length){let P=new TextEncoder().encode(h?.value);u.additionalData=P,o.push({label:"Authenticated but Unencrypted Data \u2022 Needed to decrypt",value:h.value,defaultEncoding:100,filename:`${m}-authenticated-data`})}}let v=await window.crypto.subtle.encrypt(u,b,t);o.unshift({label:`Ciphertext${a}`,value:v,defaultEncoding:64,filename:n?`${n}.enc`:"ciphertext"});break}case"sign":{let b=p?i:i.privateKey,u={name:b.algorithm.name};if(b.algorithm.name==="RSA-PSS"){let g=H.querySelector(".rsa-pss.sign input.salt-length")?.value,S;switch(b.algorithm.hash.name){case"SHA-1":S=160/8;break;case"SHA-256":S=256/8;break;case"SHA-384":S=384/8;break;case"SHA-512":S=512/8;break;default:S=0}u.saltLength=g===""?S:Number(g),o.push({label:"RSA-PSS Salt Length \u2022 Needed for signature verification",value:String(u.saltLength),defaultEncoding:-3,filename:`${m}-salt-length`});let h=b.algorithm.modulusLength;console.debug("Limit",Math.ceil((h-1)/8)-S-2)}if(b.algorithm.name==="ECDSA"){let g=H.querySelector(".ecdsa.sign select")?.selectedOptions[0].value;u.hash=g,o.push({label:"Hash Function \u2022 Needed for signature verification",value:String(g),defaultEncoding:100,filename:`${m}-hash`})}let v=await window.crypto.subtle.sign(u,b,t);o.unshift({label:`Signature${a}`,value:v,defaultEncoding:64,filename:n?`${n}.sig`:"signature"});break}case"decrypt":{let b=p?i:i.privateKey,u={name:b.algorithm.name};if(b.algorithm.name==="AES-CTR"){let g=H.querySelector(".aes-ctr.decrypt input.counter")?.value,S=H.querySelector(".aes-ctr.decrypt select")?.selectedOptions[0].value,h;if(g){let w=Number(S);if(h=U(g,w),h.byteLength!==16)throw new Error("Counter length must be exactly 16 bytes.")}else throw new Error("AES-CTR requires the initial counter value in order to decrypt the ciphertext.");u.counter=h,u.length=64}if(b.algorithm.name==="AES-CBC"){let g=H.querySelector(".aes-cbc.decrypt .iv input")?.value,S=H.querySelector(".aes-cbc.decrypt .iv select")?.selectedOptions[0].value,h;if(g){let w=Number(S);if(h=U(g,w),h.byteLength!==16)throw new Error("IV length must be exactly 16 bytes.")}else throw new Error("AES-CBC requires the Initialization Vector (IV) in order to decrypt the ciphertext.");u.iv=h}if(b.algorithm.name==="AES-GCM"){let g=H.querySelector(".aes-gcm.decrypt .iv input")?.value,S=H.querySelector(".aes-gcm.decrypt .iv select")?.selectedOptions[0].value,h=H.querySelector(".aes-gcm.decrypt textarea"),w=H.querySelector(".aes-gcm.decrypt .tag-length select")?.selectedOptions[0].value;u.tagLength=Number(w)||128;let $;if(g){let P=Number(S);$=U(g,P)}else throw new Error("AES-GCM requires the Initialization Vector (IV) in order to decrypt the ciphertext.");if(u.iv=$,h?.value.length){let P=new TextEncoder().encode(h?.value);u.additionalData=P}}let v=await window.crypto.subtle.decrypt(u,b,t);o.unshift({label:`Plaintext${a}`,value:v,defaultEncoding:100,filename:n?n.replace(/(\.?enc)|(\.?bin)/ig,""):"plaintext",extension:null});break}default:throw new Error(`Operation ${e} is not implemented.`)}return o},Vn=H.querySelector("button");Vn?.addEventListener("click",async()=>{N(0);try{let e=H.querySelector("menu li.active")?.dataset.op;if(!e)throw new Error("An operation must be defined.");let t=Number(H.querySelector(".input .encoding select")?.selectedOptions[0].value),n=U(en.value,t),r=await tn(e,n);F(r)}catch(e){C(e)}});var nn=e=>{if(!(!e||!e.length)){N(0);try{let t=H.querySelector("menu li.active")?.dataset.op;if(!t)throw new Error("An operation must be defined.");let n=Array.from(e),r=n.reduce((i,p)=>i+p.size,0),o=[],a=[];n.forEach((i,p)=>{let m=new FileReader;m.onprogress=b=>{o[p]=b.loaded,N(o.reduce((u,v)=>u+v,0)/r*100)},m.onload=async b=>{if(!b.target?.result||!(b.target.result instanceof ArrayBuffer)){C(new Error(`File [${i.name}] of type [${i.type}] failed to load.`));return}let u=`${i.name} \u2022 ${i.size.toLocaleString()} bytes \u2022 ${i.type||"Unknown type"}`;try{a.push(await tn(t,b.target.result,i.name,u))}catch(v){C(v)}a.length===e.length&&F(a.flat())},m.readAsArrayBuffer(i)})}catch(t){C(t)}}},Qt=H.querySelector(".upload input");Qt?.addEventListener("change",()=>nn(Qt.files));en.addEventListener("drop",e=>{nn(e.dataTransfer?.files)});var me=De(ze());var Un=document.querySelector("#hash-bcrypt .control.cost span:first-child"),We=document.querySelector("#hash-bcrypt input.cost");We.addEventListener("change",()=>{let e=parseInt(We.value,10);Un.textContent=`Cost \u2022 ${(2**e).toLocaleString()} iterations`});var rn=(e,t)=>(n,r)=>{if(n&&C(n),typeof r=="boolean")F([{label:"Bcrypt Verification Result",value:String(r)}]);else{let a=me.getSalt(r),i=a.replace(/^.+\$/,""),p=r.substring(a.length);F([{label:"Bcrypt String",value:r,filename:"bcrypt-string"},{label:"Salt",value:i,filename:"bcrypt-salt"},{label:"Hash",value:p,filename:"bcrypt-hash"}])}let o=e;o.textContent=t,o.disabled=!1},an=e=>t=>{N(t*100);let n=e;n.textContent=`${String(t*100).substring(0,4)}%`},Se=document.querySelector("#hash-bcrypt button");Se?.addEventListener("click",()=>{N(0),Se.disabled=!0;try{let e=parseInt(We.value,10)||10,t=document.querySelector("#hash-bcrypt .password input");me.hash(t.value,e,rn(Se,Se.textContent),an(Se))}catch(e){C(e)}});var we=document.querySelector("#verify-bcrypt button");we?.addEventListener("click",async()=>{N(0),we.disabled=!0;let e=document.querySelector("#verify-bcrypt .password input"),t=document.querySelector("#verify-bcrypt .hash input");try{me.compare(e.value,t.value,rn(we,we.textContent),an(we))}catch(n){C(n)}});var de=document.querySelector("#hash-PBKDF2"),et=de.querySelector(".prf-select"),qe=de.querySelector("input.derivation-length"),on=de.querySelector("input.iterations"),Dn=de.querySelector("select.input-encoding"),cn=de.querySelector("input.salt"),Gn=de.querySelector(".password input"),Qe=de.querySelector("button");qe?.addEventListener("change",()=>{let e=document.querySelector("#passwords menu .output-length span");e&&(e.textContent=`${parseInt(qe.value,10)*8} bits`)});et?.addEventListener("change",()=>{let e=document.querySelector("#passwords menu .block-size span");e&&(e.textContent=String(et.selectedOptions[0].dataset.bs))});Qe?.addEventListener("click",async()=>{N(0),Qe.disabled=!0;try{let e=new TextEncoder().encode(Gn.value),t=await crypto.subtle.importKey("raw",e,{name:"PBKDF2"},!1,["deriveBits"]),n;if(cn.value){let g=Number(Dn.selectedOptions[0].value);n=U(cn.value,g)}else n=crypto.getRandomValues(new Uint8Array(16)).buffer;let r=et.selectedOptions[0].dataset.alg,o=Number(on.value.length?on.value:1e5),a={name:"PBKDF2",salt:n,hash:r,iterations:o},i=Number(qe.value.length?qe.value:32)*8;if(i<0)throw new Error("Byte length must be positive");let p=await crypto.subtle.deriveBits(a,t,i),m=g=>g.replace(/=/g,""),b=m(j(n,64)),u=m(j(p,64)),v=`$pbkdf2$prf=hmac-${r.toLowerCase()},c=${o},dklen=${i}$${b}$${u}`;F([{label:"PBKDF2 String in PHC String Format (recommended by author)",value:v,filename:"pbkdf2-string"},{label:"Hash",value:p,defaultEncoding:64,filename:"pbkdf2-hash"},{label:"Salt",value:n,defaultEncoding:64,filename:"pbkdf2-salt"}])}catch(e){C(e)}Qe.disabled=!1});var G=De(mn());var ft=class extends G.SRPRoutines{computeIdentityHash(t,n){return this.hash(new TextEncoder().encode(`${t}:${n}`).buffer)}},xt=({strict:e,primeGroup:t,hashAlgorithm:n})=>{let r=G.SRPParameters.PrimeGroup[t],o=G.SRPParameters.H[n],a=new G.SRPParameters(r,o);return e?new ft(a):new G.SRPRoutines(a)},ir=async({routines:e,username:t,password:n})=>{let{s:r,v:o}=await(0,G.createVerifierAndSalt)(e,t,n);return{salt:r,verifier:o}},lr=async e=>{let{routines:t,username:n,password:r,salt:o,serverPublicKey:a}=e,m=await(await new G.SRPClientSession(t).step1(n,r)).step2(o,a),{A:b,M1:u}=m;return{A:b,M1:u}},dr=async e=>{let{routines:t,username:n,password:r,salt:o,verifier:a}=e,i=new G.SRPClientSession(t),p=new G.SRPServerSession(t),m=await i.step1(n,r),b=await p.step1(n,o,a),u=await m.step2(o,b.B),v=await b.step2(u.A,u.M1);await u.step3(v);let{B:g}=b,{A:S,M1:h}=u;return{B:g,A:S,M1:h,M2:v}},ye=document.querySelector("#register-SRP"),ur=ye.querySelector("button"),re=document.querySelector("#authenticate-SRP"),fr=re.querySelector("button"),ae=document.querySelector("#prove-identity-SRP"),xr=ae.querySelector("button");ur.addEventListener("click",async()=>{let e=ye.querySelector("input.strict").checked,t=Number(ye.querySelector("select.prime-group").selectedOptions[0].value),n=String(ye.querySelector("select.hash-algorithm").selectedOptions[0].dataset.alg),r=ye.querySelector("input.identity")?.value||"",o=ye.querySelector('input[type="password"]')?.value||"";try{let a=xt({strict:e,primeGroup:t,hashAlgorithm:n}),{salt:i,verifier:p}=await ir({routines:a,username:r,password:o});F([{label:"Salt \u2022 Store in user record",value:i.toString(),defaultEncoding:-2,filename:"srp-salt"},{label:"Verifier \u2022 Store in user record",value:p.toString(),defaultEncoding:-2,filename:"srp-verifier"},{label:"Prime",value:G.SRPParameters.PrimeGroup[t].N.toString(),defaultEncoding:-2,filename:"srp-prime"}])}catch(a){C(a)}});fr.addEventListener("click",async()=>{let e=re.querySelector("input.strict").checked,t=Number(re.querySelector("select.prime-group").selectedOptions[0].value),n=String(re.querySelector("select.hash-algorithm").selectedOptions[0].dataset.alg),r=re.querySelector("input.identity")?.value||"",o=re.querySelector('input[type="password"]')?.value||"";try{let a=BigInt(re.querySelector(".salt textarea")?.value||""),i=BigInt(re.querySelector(".verifier textarea")?.value||""),p=xt({strict:e,primeGroup:t,hashAlgorithm:n}),m=[],{B:b,A:u,M1:v,M2:g}=await dr({routines:p,username:r,password:o,salt:a,verifier:i});m.push({label:"Client Public Key A",value:u.toString(),defaultEncoding:-2,filename:"srp-client-public-key"}),m.push({label:"Client Proof M1",value:v.toString(),defaultEncoding:-2,filename:"srp-client-proof"}),m.push({label:"Server Public Key B",value:b.toString(),defaultEncoding:-2,filename:"srp-server-public-key"}),m.push({label:"Server Proof M2",value:g.toString(),defaultEncoding:-2,filename:"srp-server-proof"}),F(m)}catch(a){C(a)}});xr.addEventListener("click",async()=>{let e=ae.querySelector("input.strict").checked,t=Number(ae.querySelector("select.prime-group").selectedOptions[0].value),n=String(ae.querySelector("select.hash-algorithm").selectedOptions[0].dataset.alg),r=ae.querySelector("input.identity")?.value||"",o=ae.querySelector('input[type="password"]')?.value||"",a=ae.querySelector(".public-key textarea")?.value||"";try{let i=BigInt(ae.querySelector(".salt textarea")?.value||""),p=xt({strict:e,primeGroup:t,hashAlgorithm:n}),m=[],b=BigInt(a),{A:u,M1:v}=await lr({routines:p,username:r,password:o,salt:i,serverPublicKey:b});m.push({label:"Client Public Key A",value:u.toString(),defaultEncoding:-2,filename:"srp-client-public-key"}),m.push({label:"Client Proof M1",value:v.toString(),defaultEncoding:-2,filename:"srp-client-proof"}),F(m)}catch(i){C(i)}});var pt=document.querySelector("#passwords menu .algorithm select"),Ve=document.querySelector("#passwords .operation"),hn=()=>{let e=Ve.selectedOptions[0].value,t=pt.selectedOptions[0].dataset.alg,n=document.querySelector(`#passwords #${e}-${t}`);document.querySelectorAll("#passwords .subsection > .settings").forEach(o=>o.classList.remove("active")),n?.classList.add("active")},pr=()=>{let e=document.querySelector("#passwords menu"),t=e.querySelector(".block-size span"),n=e.querySelector(".method span"),r=e.querySelector(".specification span"),o=pt.selectedOptions[0].dataset;t.textContent=o.bs||"",n.textContent=o.method||"",r.textContent=o.spec||"",Ve.textContent="",o.operations?.split(" ").forEach((a,i)=>{let p=document.createElement("option");i===0&&(p.selected=!0),p.value=a,p.textContent=`${a[0].toUpperCase()}${a.substring(1).replace("-"," ")}`,Ve.appendChild(p)}),Y(),hn()};pt.addEventListener("change",pr);Ve.addEventListener("change",hn);var br=e=>{e.preventDefault(),e.target.checkValidity()||e.stopImmediatePropagation()};document.addEventListener("DOMContentLoaded",()=>{document.querySelectorAll("form")?.forEach(e=>e.addEventListener("submit",br))});var mr=(e,t)=>()=>{let n=oe(t.value);n&&e.childNodes.forEach(r=>{let o=r;Number(o.value)===n&&(o.selected=!0)})},hr=document.querySelectorAll("label.multi");hr.forEach(e=>{let t=e.querySelector("input"),n=e.querySelector("select");t.addEventListener("paste",()=>setTimeout(mr(n,t),0))});var yn=document.querySelectorAll(".operation-area"),yr=e=>()=>{let t=e.parentElement.querySelector(".encoding select")||e.parentElement.parentElement.querySelector(".encoding select"),n=oe(e.value);n&&t?.childNodes.forEach(r=>{let o=r;Number(o.value)===n&&(o.selected=!0)})};yn.forEach(e=>{e.querySelectorAll("textarea").forEach(n=>{n.addEventListener("paste",()=>setTimeout(yr(n),0)),n.addEventListener("input",()=>{let r=n.parentElement.querySelector(".character-count")||n.parentElement.parentElement.querySelector(".character-count");if(r){let o=n.value.length;o===1?r.textContent="1 character":r.textContent=`${o} characters`}})})});var bt=e=>{e.stopPropagation(),e.preventDefault()},Te=(e,t)=>()=>{t?e.classList.add("dragover"):e.classList.remove("dragover")};yn.forEach(e=>{e.querySelectorAll(".input > textarea").forEach(n=>{n?.addEventListener("dragenter",bt),n?.addEventListener("dragenter",Te(n,!0)),n?.addEventListener("dragover",bt),n?.addEventListener("dragover",Te(n,!0)),n?.addEventListener("dragleave",Te(n,!1)),n?.addEventListener("dragend",Te(n,!1)),n?.addEventListener("drop",bt),n?.addEventListener("drop",Te(n,!1))})});var mt=e=>{let t=e;t.target?.nodeName!=="TEXTAREA"&&t.dataTransfer&&(t.preventDefault(),t.dataTransfer.effectAllowed="none",t.dataTransfer.dropEffect="none")};window.addEventListener("dragover",mt);window.addEventListener("dragenter",mt);window.addEventListener("drop",mt);var gr=document.querySelectorAll(".tabs");gr.forEach(e=>{let t=Array.from(e.children),n=e.parentElement,r=n.querySelectorAll(":scope [data-tab]");t.forEach(o=>{o.classList.contains("datum")||o.addEventListener("click",()=>{t.forEach(a=>a.classList.remove("active")),r?.forEach(a=>a.classList.remove("active")),o.classList.add("active"),n.querySelector(`[data-tab="${o.dataset.target}"]`)?.classList.add("active"),Y()})})});})();
 /**
  * CryptoTools
  * @author Micah Henning <hello@micah.soy>
