+++ src/lib/tabs.ts
@@ -12,7 +12,7 @@ const tabsElements = document.querySelectorAll('.tabs');
 tabsElements.forEach((tabsElement) => {
   const tabs = Array.from(tabsElement.children) as HTMLElement[];
   const parent = tabsElement.parentElement!;
-  const sections = parent.querySelectorAll(':scope > section.settings');
+  const sections = parent.querySelectorAll(':scope [data-tab]');
 
   tabs.forEach((tab) => {
     if (tab.classList.contains('datum')) return; // don't interact with stats
@@ -22,7 +22,7 @@ tabsElements.forEach((tabsElement) => {
       sections?.forEach((s) => s.classList.remove('active'));
 
       tab.classList.add('active');
-      parent.querySelector(`.settings.${tab.dataset.target}`)?.classList.add('active');
+      parent.querySelector(`[data-tab="${tab.dataset.target}"]`)?.classList.add('active');
 
       hideResults();
     });
