+++ src/styles/responsive.sass
@@ -10,7 +10,7 @@
     width: auto
     margin: 2rem 1rem 0
 
-  #encryption .list-keys li button:last-of-type
+  #encryption [data-tab="list-keys"] li button:last-of-type
     margin-right: auto
 
   // Don't show tooltips on mobile
