+++ src/styles/sections.sass
@@ -92,7 +92,7 @@ form.settings
   .operation-area .settings > label.additional-data
     width: 100%
 
-  .operation-area button
+  .crypto-operations button
     width: 100%
     border-radius: 0 0 5px 5px
 
@@ -109,63 +109,63 @@ form.settings
     top: 4px
     left: 11px
 
-  .list-keys ul
+  [data-tab="list-keys"] ul
     text-align: left
     padding: 0
     margin: 1rem
     width: 100%
 
-  .list-keys li
-    display: flex
-    flex-wrap: wrap
-    line-height: 1.5rem
-    padding: 1rem
-    border-top: 2px solid $border
-    cursor: pointer
-
-    &:first-child
-      border-top: none
-
-    // Hover effect prevents tooltip visibility 
-    // &:hover
-      // background: $border
-      // border-radius: 5px
-    // &:hover + li
-    //   border-top-color: transparent
-
-    .container
+    li
       display: flex
-      align-items: center
+      flex-wrap: wrap
+      line-height: 1.5rem
+      padding: 1rem
+      border-top: 2px solid $border
+      cursor: pointer
 
-      svg
+      &:first-child
+        border-top: none
+
+      // Hover effect prevents tooltip visibility 
+      // &:hover
+        // background: $border
+        // border-radius: 5px
+      // &:hover + li
+      //   border-top-color: transparent
+
+      .container
+        display: flex
+        align-items: center
+
+        svg
+          color: $primary
+
+      .meta
+        display: flex
+        flex-direction: column
+        margin-left: 1rem
+
+      span:first-child
+        font: 100 italic 1rem 'Lora', serif
         color: $primary
 
-    .meta
-      display: flex
-      flex-direction: column
-      margin-left: 1rem
+      .saved svg
+        color: $text
+        vertical-align: text-top
+        margin-left: 0.5rem
 
-    span:first-child
-      font: 100 italic 1rem 'Lora', serif
-      color: $primary
+      .actions
+        display: flex
+        align-items: center
+        margin: 0 0 0 auto
 
-    .saved svg
-      color: $text
-      vertical-align: text-top
-      margin-left: 0.5rem
+        a
+          color: $secondary
+          padding: 0 0.5rem
 
-    .actions
-      display: flex
-      align-items: center
-      margin: 0 0 0 auto
-
-      a
-        color: $secondary
-        padding: 0 0.5rem
-
-        &:last-child
-          color: $error
-          padding-right: 0
+          &:last-child
+            color: $error
+            padding-right: 0
 
 // Advanced Settings 
 details
